### 2021-03-14 #3103 @tsalo
The manuscript is located at https://github.com/ME-ICA/tedana/blob/joss/paper/paper.md. Is there a way to make it easier for the bot to locate it?
