### 2021-03-15 2403af8
Nest dict, move to new file constants.py
