+++ tedana/io.py
@@ -11,6 +11,8 @@ from nilearn.image import new_img_like
 
 from tedana import utils
 from tedana.stats import computefeats2, get_coeffs
+from .constants import allowed_conventions, img_table
+
 
 LGR = logging.getLogger(__name__)
 RepLGR = logging.getLogger('REPORT')
@@ -18,81 +20,7 @@ RefLGR = logging.getLogger('REFERENCES')
 
 outdir = '.'
 prefix = ''
-convention = 'bids' # overridden in API or CLI calls
-
-
-img_table = {
-    'adaptive mask': ('adaptive_mask', 'desc-adaptiveGoodSignal_mask'),
-    't2star map': ('t2sv', 'T2starmap'),
-    's0 map': ('s0v', 'S0map'),
-    'combined': ('ts_OC', 'desc-optcom_bold'),
-    'ICA components': ('ica_components', 'desc-ICA_components'),
-    'z-scored PCA components': (
-        'pca_components', 'desc-PCA_stat-z_components'
-    ),
-    'z-scored ICA components': (
-        'betas_OC', 'desc-ICA_stat-z_components'
-    ),
-    'ICA accepted components': (
-        'betas_hik_OC', 'desc-ICAAccepted_components'
-    ),
-    'z-scored ICA accepted components': (
-        'feats_OC2', 'desc-ICAAccepted_stat-z_components'
-    ),
-    'denoised ts': ('dn_ts_OC', 'desc-optcomDenoised_bold'),
-    'high kappa ts': ('hik_ts_OC', 'desc-optcomAccepted_bold'),
-    'low kappa ts': ('lowk_ts_OC', 'desc-optcomRejected_bold'),
-    # verbose outputs
-    'full t2star map': ('t2svG', 'desc-full_T2starmap'),
-    'full s0 map': ('s0vG', 'desc-full_S0map'),
-    'whitened': ('ts_OC_whitened', 'desc-optcomPCAReduced_bold'),
-    'echo weight PCA map split': (
-        'e{0}_PCA_comp', 'echo-{0}_desc-PCA_components'
-    ),
-    'echo R2 PCA split': (
-        'e{0}_PCA_R2', 'echo-{0}_desc-PCAR2ModelPredictions_components'
-    ),
-    'echo S0 PCA split': (
-        'e{0}_PCA_S0', 'echo-{0}_desc-PCAS0ModelPredictions_components'
-    ),
-    'PCA component weights': (
-        'pca_weights', 'desc-PCAAveragingWeights_components'
-    ),
-    'PCA reduced': ('oc_reduced', 'desc-optcomPCAReduced_bold'),
-    'echo weight ICA map split': (
-        'e{0}_ICA_comp', 'echo-{0}_desc-ICA_components'
-    ),
-    'echo R2 ICA split': (
-        'e{0}_ICA_R2', 'echo-{0}_desc-ICAR2ModelPredictions_components'
-    ),
-    'echo S0 ICA split': (
-        'e{0}_ICA_S0', 'echo-{0}_desc-ICAS0ModelPredictions_components'
-    ),
-    'ICA component weights': (
-        'ica_weights', 'desc-ICAAveragingWeights_components'
-    ),
-    'high kappa ts split': (
-        'hik_ts_e{0}', 'echo-{0}_desc-Accepted_bold'
-    ),
-    'low kappa ts split': (
-        'lowk_ts_e{0}', 'echo-{0}_desc-Rejected_bold'
-    ),
-    'denoised ts split': ('dn_ts_e{0}', 'echo-{0}_desc-Denoised_bold'),
-    # global signal outputs
-    'gs map': ('T1gs', 'desc-globalSignal_map'),
[…]
