### 2021-03-15 fad0845
Switches bids pointer method
