+++ tedana/workflows/tedana.py
@@ -18,7 +18,7 @@ from threadpoolctl import threadpool_limits
 from nilearn.masking import compute_epi_mask
 
 from tedana import (decay, combine, decomposition, io, metrics,
-                    reporting, selection, utils, __version__)
+                    reporting, selection, utils, constants, __version__)
 import tedana.gscontrol as gsc
 from tedana.stats import computefeats2
 from tedana.workflows.parser_utils import is_valid_file, check_tedpca_value, ContextFilter
@@ -348,6 +348,8 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
     if prefix and prefix[-1] != '_':
         prefix += '_'
     io.prefix = prefix
+    if convention == 'bids':
+        convention = constants.bids
     io.convention = convention
 
     # boilerplate
