### 2021-03-15 e479836
Adds tables for json/tsv
