+++ tedana/workflows/tedana.py
@@ -437,8 +437,9 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
     if mixm is not None and op.isfile(mixm):
         mixm = op.abspath(mixm)
         # Allow users to re-run on same folder
-        if mixm != op.join(out_dir, 'desc-ICA_mixing.tsv'):
-            shutil.copyfile(mixm, op.join(out_dir, 'desc-ICA_mixing.tsv'))
+        mixing_name = io.gen_tsv_name("ICA mixing")
+        if mixm != mixing_name:
+            shutil.copyfile(mixm, mixing_name)
             shutil.copyfile(mixm, op.join(out_dir, op.basename(mixm)))
     elif mixm is not None:
         raise IOError('Argument "mixm" must be an existing file.')
@@ -446,8 +447,9 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
     if ctab is not None and op.isfile(ctab):
         ctab = op.abspath(ctab)
         # Allow users to re-run on same folder
-        if ctab != op.join(out_dir, 'desc-tedana_metrics.tsv'):
-            shutil.copyfile(ctab, op.join(out_dir, 'desc-tedana_metrics.tsv'))
+        metrics_name = io.gen_tsv_name("ICA metrics")
+        if ctab != metrics_name:
+            shutil.copyfile(ctab, metrics_name)
             shutil.copyfile(ctab, op.join(out_dir, op.basename(ctab)))
     elif ctab is not None:
         raise IOError('Argument "ctab" must be an existing file.')
@@ -589,13 +591,14 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
                 keep_restarting = False
     else:
         LGR.info('Using supplied mixing matrix from ICA')
-        mmix_orig = pd.read_table(op.join(out_dir, 'desc-ICA_mixing.tsv')).values
+        mixing_file = io.gen_tsv_name("ICA mixing")
+        mmix_orig = pd.read_table(mixing_file).values
 
         if ctab is None:
             comptable, metric_maps, metric_metadata, betas, mmix = metrics.dependence_metrics(
                         catd, data_oc, mmix_orig, masksum, tes,
-                        ref_img, label='ICA', out_dir=out_dir,
-                        algorithm='kundu_v2', verbose=verbose)
+                        ref_img, label='ICA', algorithm='kundu_v2',
+                        verbose=verbose)
             comptable, metric_metadata = metrics.kundu_metrics(
                 comptable,
                 metric_maps,
@@ -611,8 +614,7 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
             mmix = mmix_orig.copy()
             comptable = pd.read_table(ctab)
             # Try to find and load the metric metadata file
-            ctab_parts = ctab.split(".")
-            metadata_file = ctab_parts[0] + ".json"
+            metadata_file = io.gen_json_name('ICA metrics')
             if op.isfile(metadata_file):
                 with open(metadata_file, "r") as fo:
                     metric_metadata = json.load(fo)
@@ -629,14 +631,14 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
     # Write out ICA files.
     comp_names = comptable["Component"].values
     mixing_df = pd.DataFrame(data=mmix, columns=comp_names)
-    mixing_df.to_csv(op.join(out_dir, "desc-ICA_mixing.tsv"), sep="\t", index=False)
+    mixing_df.to_csv(io.gen_tsv_name("ICA mixing"), sep="\t", index=False)
     betas_oc = utils.unmask(computefeats2(data_oc, mmix, mask), mask)
     io.filewrite(betas_oc, 'z-scored ICA components', ref_img)
 
[…]
