+++ tedana/info.py
@@ -26,7 +26,7 @@ DOWNLOAD_URL = (
 
 REQUIRES = [
     'bokeh',
-    'mapca>=0.0.1',
+    'mapca~=0.0.1',
     'matplotlib',
     'nibabel>=2.5.1',
     'nilearn>=0.5.2',
