### 2021-03-16 46a1f21
Apply normalization in ma_pca.
