+++ tedana/io.py
@@ -1,5 +1,63 @@
 """
-Functions to handle file input/output
+=============================
+io module (:mod: `tedana.io`)
+=============================
+
+.. currentmodule:: tedana.io
+
+The io module handles most file input and output in the `tedana` workflow,
+and simplifies some naming function calls with module globals (see "Globals"
+and "Notes" below). Other functions in the module help simplify writing out
+data from multiple echoes or write very complex outputs.
+
+
+Globals
+-------
+outdir
+prefix
+convention
+
+
+Naming Functions
+----------------
+get_img_name
+get_json_name
+get_tsv_name
+add_decomp_prefix
+
+
+File Writing Functions
+----------------------
+write_split_ts
+writefeats
+writeresults
+writeresults_echoes
+filewrite
+
+
+File Loading Functions
+----------------------
+load_data
+
+
+Helper Functions
+----------------
+new_nii_like
+split_ts
+check_convention
+
+See Also
+--------
+`tedana.constants`
+
+
+Notes
+-----
+The global variables are set by default in the module to guarantee that the
+functions that use them won't fail if a workflow API is not used.
+However, API calls can override the default settings. Additionally, the
+naming functions beginning with "get" all leverage dictionaries defined in
+the `constants.py` module, as the definitions are large.
 """
 import logging
 import os.path as op
@@ -24,19 +82,7 @@ outdir = '.'
 prefix = ''
 convention = bids   # overridden in API or CLI calls
 
-
-def check_convention() -> None:
-    """Checks set convention for io module
-
-    Raises
-    ------
-    RuntimeError, if invalid convention is set
-    """
-    if convention not in allowed_conventions:
-        raise RuntimeError(
-            ('Convention %s is not valid; allowed: ' % convention) +
-            str(allowed_conventions)
-        )
+# Naming Functions
 
 
 def gen_img_name(img_type: str, echo: str = 0) -> str:
@@ -128,46 +174,34 @@ def gen_tsv_name(tsv_type: str) -> str:
     return op.join(outdir, prefix + basename + '.tsv')
 
 
-def split_ts(data, mmix, mask, comptable):
+def add_decomp_prefix(comp_num, prefix, max_value):
     """
-    Splits `data` time series into accepted component time series and remainder
+    Create component name with leading zeros matching number of components
 
     Parameters
     ----------
-    data : (S x T) array_like
-        Input data, where `S` is samples and `T` is time
-    mmix : (T x C) array_like
-        Mixing matrix for converting input data to component space, where `C`
-        is components and `T` is the same as in `data`
-    mask : (S,) array_like
-        Boolean mask array
-    comptable : (C x X) :obj:`pandas.DataFrame`
-        Component metric table. One row for each component, with a column for
-        each metric. Requires at least two columns: "component" and
-        "classification".
+    comp_num : :obj:`int`
+        Component number
+    prefix : :obj:`str`
+        A prefix to prepend to the component name. An underscore is
+        automatically added between the prefix and the component number.
+    max_value : :obj:`int`
+        The maximum component number in the whole decomposition. Used to
+        determine the appropriate number of leading zeros in the component
+        name.
 
     Returns
     -------
-    hikts : (S x T) :obj:`numpy.ndarray`
-        Time series reconstructed using only components in `acc`
-    resid : (S x T) :obj:`numpy.ndarray`
-        Original data with `hikts` removed
+    comp_name : :obj:`str`
+        Component name in the form <prefix>_<zeros><comp_num>
     """
[…]
