### 2021-03-17 d67d9bd
[REF] Replace MAPCA code with mapca library (#641)
* Move mapca out of tedana.

* Pin to roughly 0.0.1.
