+++ tedana/decomposition/pca.py
@@ -7,11 +7,11 @@ from numbers import Number
 
 import numpy as np
 import pandas as pd
+from mapca import ma_pca
 from scipy import stats
 from sklearn.decomposition import PCA
 
 from tedana import metrics, utils, io
-from tedana.decomposition import ma_pca
 from tedana.stats import computefeats2
 from tedana.selection import kundu_tedpca
 
@@ -206,7 +206,7 @@ def tedpca(data_cat, data_oc, combmode, mask, adaptive_mask, t2sG,
     if algorithm in ['mdl', 'aic', 'kic']:
         data_img = io.new_nii_like(ref_img, utils.unmask(data, mask))
         mask_img = io.new_nii_like(ref_img, mask.astype(int))
-        voxel_comp_weights, varex, varex_norm, comp_ts = ma_pca.ma_pca(
+        voxel_comp_weights, varex, varex_norm, comp_ts = ma_pca(
             data_img, mask_img, algorithm)
     elif isinstance(algorithm, Number):
         ppca = PCA(copy=False, n_components=algorithm, svd_solver="full")
