+++ tedana/decomposition/ma_pca.py
@@ -1,619 +1,0 @@
-"""
-PCA based on Moving Average (stationary Gaussian) process
-"""
-import logging
-
-import numpy as np
-
-from sklearn.decomposition import PCA
-from sklearn.preprocessing import StandardScaler
-
-from scipy.linalg import svd
-from scipy.signal import detrend, fftconvolve
-from scipy.fftpack import fftshift, fftn
-
-LGR = logging.getLogger(__name__)
-
-
-def _autocorr(data):
-    """
-    Calculates the auto correlation of a given array.
-
-    Parameters
-    ----------
-    data : array-like
-        The array to calculate the autocorrelation of
-
-    Returns
-    -------
-    u : ndarray
-        The array of autocorrelations
-    """
-    u = np.correlate(data, data, mode='full')
-    # Take upper half of correlation matrix
-    return u[u.size // 2:]
-
-
-def _check_order(order_in):
-    """
-    Checks the order passed to the window functions.
-
-    Parameters
-    ----------
-    order_in : int
-        The order to be passed to the window function
-
-    Returns
-    -------
-    n_out : ndarray
-        An integer order array
-    w : list
-        The window to be used
-    trivialwin : boolean
-        Whether the window is trivial (w in [0,1])
-    """
-
-    w = []
-    trivialwin = False
-
-    # Special case of negative orders:
-    if order_in < 0:
-        raise ValueError('Order cannot be less than zero.')
-
-    order_out = np.round(order_in)
-    if not np.array_equal(order_in, order_out):
-        LGR.warning('Rounded order to nearest integer')
-
-    # Special cases:
-    if not order_out or order_out == 0:
-        w = np.zeros((0, 1))  # Empty matrix: 0-by-1
-        trivialwin = True
-    elif order_out == 1:
-        w = 1
-        trivialwin = True
-
-    return order_out, w, trivialwin
-
-
-def _parzen_win(n_points):
-    """
-    Returns the N-point Parzen (de la Valle-Poussin) window in a column vector.
-
-    Parameters
-    ----------
-    n_points : int
-        Number of non-zero points the window must contain
-
-    Returns
-    -------
-    parzen_w : 1D array
-        The Parzen window
-
-    Notes
-    -----
-    Maths are described in the following MATLAB documentation page:
-    https://www.mathworks.com/help/signal/ref/parzenwin.html
-
-    References
-    ----------
-    * Harris, Fredric J. “On the Use of Windows for Harmonic Analysis
-      with the Discrete Fourier Transform.” Proceedings of the IEEE.
-      Vol. 66, January 1978, pp. 51–83.
-    """
-
-    # Check for valid window length (i.e., n < 0)
-    n_points, parzen_w, trivialwin = _check_order(n_points)
-    if trivialwin:
-        return parzen_w
-
-    # Index vectors
-    k = np.arange(-(n_points - 1) / 2, ((n_points - 1) / 2) + 1)
-    k1 = k[k < -(n_points - 1) / 4]
-    k2 = k[abs(k) <= (n_points - 1) / 4]
-
-    # Equation 37 of [1]: window defined in three sections
-    parzen_w1 = 2 * (1 - abs(k1) / (n_points / 2))**3
-    parzen_w2 = 1 - 6 * (abs(k2) / (n_points / 2))**2 + 6 * (abs(k2) / (n_points / 2))**3
-    parzen_w = np.hstack((parzen_w1, parzen_w2, parzen_w1[::-1])).T
-
-    return parzen_w
-
[…]
