### 2021-03-18 8df1331
Adds module text for constants.py
