### 2021-03-18 d5ce0f3
Adds set_convention
