### 2021-03-18 61ba799
Adds set_prefix function
