### 2021-03-19 7b955b2
Clarifies PCA docstring
