### 2021-03-19 5fdcd88
Update tedana/io.py
Co-authored-by: Taylor Salo <tsalo006@fiu.edu>
