### 2021-03-19 cdbe1e4
Update tedana/io.py
Co-authored-by: Taylor Salo <tsalo006@fiu.edu>
