### 2021-03-19 6d93468
Update tedana/io.py
Co-authored-by: Taylor Salo <tsalo006@fiu.edu>
