### 2021-03-19 02726df
Update tedana/io.py
Co-authored-by: Taylor Salo <tsalo006@fiu.edu>
