### 2021-03-19 c349c0d
Update tedana/io.py
Co-authored-by: Taylor Salo <tsalo006@fiu.edu>
