+++ tedana/io.py
@@ -132,7 +132,7 @@ def gen_img_name(img_type: str, echo: int = 0) -> str:
     ----------
     img_type : str
         The description of the image. Must be a key in constants.img_table
-    echo : :obj: `int`
+    echo : :obj:`int`, optional
         The echo number of the image. Default is 0.
 
     Returns
