+++ tedana/workflows/t2smap.py
@@ -64,11 +64,6 @@ def _get_parser():
                                 'Dependent ANAlysis. Must be in the same '
                                 'space as `data`.'),
                           default=None)
-    optional.add_argument('--prefix',
-                          dest='prefix',
-                          type=str,
-                          help='Prefix for filenames generated.',
-                          default='')
     optional.add_argument('--convention',
                           dest='convention',
                           action='store',
@@ -129,7 +124,7 @@ def _get_parser():
 
 
 def t2smap_workflow(data, tes, out_dir='.', mask=None,
-                    prefix='', convention='bids',
+                    convention='bids',
                     fittype='loglin', fitmode='all', combmode='t2s',
                     debug=False, quiet=False):
     """
@@ -192,7 +187,6 @@ def t2smap_workflow(data, tes, out_dir='.', mask=None,
     if not op.isdir(out_dir):
         os.mkdir(out_dir)
     io.outdir = out_dir
-    io.set_prefix(prefix)
     io.set_convention(convention)
 
     if debug and not quiet:
