### 2021-03-19 482851d
Switches tables to json files
