### 2021-03-19 c3aba1c
Revert "Addresses some review concerns"
This reverts commit 63903a4868e3a2eb895ebde47eb0d9af47981814.
