### 2021-03-21 eee5afe
Enable normalization in mapca call
