### 2021-03-22 017c861
Actually fix the string conversion
