### 2021-03-22 170422e
[DOC] Add references to HTML report (#695)
