### 2021-03-22 9ba0ec9
Replace hard-coded T2starmap with io call
