+++ docs/multi-echo_walkthrough.rst
@@ -2,6 +2,7 @@ What is multi-echo fMRI?
 ========================
 
 .. admonition:: TL;DR
+
     Most echo-planar image (EPI) sequences collect a single brain image following
     a radio frequency (RF) pulse, at a rate known as the repetition time (TR).
     This typical approach is known as single-echo fMRI.
@@ -14,7 +15,7 @@ What is multi-echo fMRI?
 To understand what multi-echo fMRI is and why it's useful,
 we will walk through a number of things.
 
-To start, let's look at how fMRI signal decays as a function of echo time.
+To start, let's look at the most common type of fMRI, single-echo EPI.
 
 What is single-echo fMRI?
 -------------------------
@@ -27,6 +28,10 @@ and that energy is actively measured at an "echo time" (also known as `TE`_).
 When you have a single echo time, you acquire one value for each voxel, at each time point.
 
 .. image:: /_static/mr_schematic_single_echo.png
+
+This works rather well, in that there is relatively high BOLD contrast as every voxel,
+although the level of contrast (and thus the signal-to-noise ratio) will vary across the brain.
+However, this
 
 .. _multi-echo physics:
 
@@ -50,16 +55,12 @@ By using the information across multiple echos these images can be combined in
 an optimal manner to take advantage of the signal
 in the earlier echos (see :ref:`optimal combination`).
 
-.. image:: /_static/physics_kundu_2017_multiple_echoes.jpg
+.. image:: /_static/physics_signal_decay.png
 
-Adapted from  `Kundu et al. (2017)`_.
+.. image:: /_static/physics_multiple_echos.png
 
 In order to classify the relationship between the signal and the echo time we can consider a
 single voxel at two timepoints (x and y) and the measured signal measured at three different echo times - :math:`S(TE_n)`.
-
-.. image:: /_static/physics_kundu_2017_TE_dependence.jpg
-
-Adapted from  `Kundu et al. (2017)`_.
 
 For the left column, we are observing a change that we term :math:`{\Delta}{S_0}` - that is a change
 in the intercept or raw signal intensity.
@@ -84,6 +85,7 @@ that the magnitude of the changes is dependent on the echo time.
 
 For a more comprehensive review of these topics and others, see `Kundu et al. (2017)`_.
 
+.. _TE: http://mriquestions.com/tr-and-te.html
 .. _TEs: http://mriquestions.com/tr-and-te.html
 .. _BOLD signal: http://www.fil.ion.ucl.ac.uk/spm/course/slides10-zurich/Kerstin_BOLD.pdf
 .. _Kundu et al. (2017): https://www.sciencedirect.com/science/article/pii/S1053811917302410?via%3Dihub
@@ -117,84 +119,3 @@ as well as additional information only available when looking at signal decay ac
 We can use this information to denoise the optimally combined time series.
 
 .. _Pruim et al. (2015): https://www.sciencedirect.com/science/article/pii/S1053811915001822
-
-
-Considerations for ME-fMRI
---------------------------
-Multi-echo fMRI acquisition sequences and analysis methods are rapidly maturing.
-Someone who has access to a multi-echo fMRI sequence should seriously consider using it.
-
-Costs and benefits of multi-echo fMRI
--------------------------------------
-The following are a few points to consider when deciding whether or not to collect multi-echo data.
-
-Possible increase in TR
-```````````````````````
-The one difference with multi-echo is a slight time cost.
-For multi-echo fMRI, the shortest echo time (TE) is essentially free since it is collected in the
-gap between the RF pulse and the single-echo acquisition.
-The second echo tends to roughly match the single-echo TE.
-Additional echos require more time.
[…]
