### 2021-03-23 0c87c74
Fixes docstring for io.py
