+++ tedana/io.py
@@ -502,7 +502,7 @@ def filewrite(data, img_type, ref_img, gzip=True, copy_header=True,
     ----------
     data : (S [x T]) array_like
         Data to be saved
-    img_type : :obj: `str`
+    img_type : :obj:`str`
         The type of file to write
     ref_img : :obj:`str` or img_like
         Reference image
