### 2021-03-24 0f4d54c
[DOC] Added MAPCA to list of dependencies (#709)
* added MAPCA dependency

* Added mapca dependency to installation.rst
