### 2021-04-01 6466882
Adds f_max to right place
