+++ tedana/metrics/kundu_fit.py
@@ -174,6 +174,7 @@ def dependence_metrics(catd, tsoc, mmix, adaptive_mask, tes, ref_img,
             SSE_S0 = (comp_betas[:j_echo] - pred_S0)**2
             SSE_S0 = SSE_S0.sum(axis=0)  # (S,) prediction error map
             F_S0 = (alpha - SSE_S0) * (j_echo - 1) / (SSE_S0)
+            F_S0[F_S0 > F_MAX] = F_MAX
             F_S0_maps[mask_idx[mask], i_comp] = F_S0[mask_idx[mask]]
 
             # R2 Model
@@ -183,6 +184,7 @@ def dependence_metrics(catd, tsoc, mmix, adaptive_mask, tes, ref_img,
             SSE_R2 = (comp_betas[:j_echo] - pred_R2)**2
             SSE_R2 = SSE_R2.sum(axis=0)
             F_R2 = (alpha - SSE_R2) * (j_echo - 1) / (SSE_R2)
+            F_R2[F_R2 > F_MAX] = F_MAX
             F_R2_maps[mask_idx[mask], i_comp] = F_R2[mask_idx[mask]]
 
             if verbose:
