### 2021-04-01 2b9a7ab
[REF] Modularize io (#692)
* Covers non-verbose non-global mod

* Adds .swp to gitignore

* Adds full image functionality

* Adds --prefix, --convention to t2map.py

* Fix io tests

* Switches bids pointer method

* Adds tsv and json tables

* Add documentation and rearrange module

* Adds module text for constants.py

* Adds set_convention

* Adds set_prefix function

* Clarifies PCA docstring

Co-authored-by: Taylor Salo <tsalo006@fiu.edu>
