+++ tedana/tests/test_io.py
@@ -8,6 +8,7 @@ import pytest
 import pandas as pd
 
 from tedana import io as me
+from tedana import constants
 from tedana.tests.test_utils import fnames, tes
 
 from tedana.tests.utils import get_test_data_path
@@ -106,13 +107,12 @@ def test_smoke_write_split_ts():
     assert me.write_split_ts(data, mmix, mask, comptable, ref_img) is not None
 
     # TODO: midk_ts.nii is never generated?
-    for filename in ["hik_ts_.nii.gz", "lowk_ts_.nii.gz", "dn_ts_.nii.gz"]:
+    fn = me.gen_img_name
+    split = ('high kappa ts', 'low kappa ts', 'denoised ts')
+    fnames = [fn(f) + '.nii.gz' for f in split]
+    for filename in fnames:
         # remove all files generated
-        try:
-            os.remove(filename)
-        except OSError:
-            print(filename + " not generated")
-            pass
+        os.remove(filename)
 
 
 def test_smoke_writefeats():
@@ -129,31 +129,32 @@ def test_smoke_writefeats():
     assert me.writefeats(data, mmix, mask, ref_img) is not None
 
     # this only generates feats_.nii, so delete that
-    try:
-        os.remove("feats_.nii.gz")
-    except OSError:
-        print("feats_.nii not generated")
-        pass
+    os.remove(
+            me.gen_img_name('z-scored ICA accepted components') +
+            '.nii.gz'
+    )
 
 
 def test_smoke_filewrite():
     """
-    Ensures that filewrite writes out a neuroimage with random input,
-    since there is no name, remove the image named .nii
+    Ensures that filewrite fails for no known image type, write a known key
+    in both bids and orig formats
     """
-    n_samples, n_times, _ = 64350, 10, 6
+    n_samples, _, _ = 64350, 10, 6
     data_1d = np.random.random((n_samples))
-    data_2d = np.random.random((n_samples, n_times))
-    filename = ""
     ref_img = os.path.join(data_dir, 'mask.nii.gz')
 
-    assert me.filewrite(data_1d, filename, ref_img) is not None
-    assert me.filewrite(data_2d, filename, ref_img) is not None
+    with pytest.raises(KeyError):
+        me.filewrite(data_1d, '', ref_img)
 
-    try:
-        os.remove(".nii.gz")
-    except OSError:
-        print(".nii not generated")
+    for convention in (constants.bids, 'orig'):
+        me.set_convention(convention)
+        fname = me.filewrite(data_1d, 't2star map', ref_img)
+        assert fname is not None
+        try:
+            os.remove(fname)
+        except OSError:
+            print('File not generated!')
 
 
 def test_smoke_load_data():
