+++ tedana/reporting/html_report.py
@@ -5,6 +5,7 @@ from os.path import join as opj
 from string import Template
 from tedana.info import __version__
 from tedana.reporting import dynamic_figures as df
+from tedana.io import gen_tsv_name
 
 
 def _update_template_bokeh(bokeh_id, about, bokeh_js):
@@ -70,12 +71,12 @@ def generate_report(out_dir, tr):
         A generated HTML report
     """
     # Load the component time series
-    comp_ts_path = opj(out_dir, 'desc-ICA_mixing.tsv')
+    comp_ts_path = gen_tsv_name("ICA mixing")
     comp_ts_df = pd.read_csv(comp_ts_path, sep='\t', encoding='utf=8')
     n_vols, n_comps = comp_ts_df.shape
 
     # Load the component table
-    comptable_path = opj(out_dir, 'desc-tedana_metrics.tsv')
+    comptable_path = gen_tsv_name("ICA metrics")
     comptable_cds = df._create_data_struct(comptable_path)
 
     # Create kappa rho plot
