### 2021-04-01 9d68c23
Initial work on a class.
