### 2021-04-08 94f2a0e
Removes redundant thresholding
