### 2021-04-08 7b1bc76
Calculate rho and kappa on full maps.
