### 2021-04-12 167231b
[FIX] Adds f_max to right place (#712)
Moves f_max thresholding to the step where the R2/S0 maps are written out, rather than just before kappa/rho calculation.
