### 2021-04-15 2363b4f
[FIX] Calculate Kappa and Rho on full F-statistic maps (#714)
* Calculate rho and kappa on full maps.

* Add missing test outputs.

* Revert changes to outputs.
