### 2021-04-16 4d86597
Improved walkthrough (#716)
* Update gitignore

* New gifs and notebook.

* Improve gifs.

* More lovely figures.

* Reduce gifs for speed-testing.

* Start adding.

* Some more work.

* Update multi-echo_walkthrough.rst

* Update multi-echo_walkthrough.rst
