+++ docs/multi-echo_walkthrough.rst
@@ -1,0 +1,151 @@
+What is multi-echo fMRI?
+========================
+
+.. admonition:: TL;DR
+
+    Most echo-planar image (EPI) sequences collect a single brain image
+    following a radio frequency (RF) pulse, at a rate known as the
+    repetition time (TR).
+    This typical approach is known as single-echo fMRI.
+
+    In contrast, multi-echo (ME) fMRI refers to collecting data at multiple
+    echo times,
+    resulting in multiple volumes with varying levels of contrast acquired per
+    RF pulse.
+    Multi-echo fMRI can be used to identify and remove certain types of noise
+    that can't be removed from single-echo data.
+
+To understand what multi-echo fMRI is and why it's useful,
+we will walk through a number of things.
+
+The physics of fMRI
+-------------------
+
+In a typical fMRI sequence,
+the protons in a brain slice are aligned perpendicularly to the main magnetic
+field of the scanner (:math:`B_0`) with an "excitation pulse".
+Those protons start releasing energy from the excitation pulse as they fall
+back in line with :math:`B_0`,
+and that energy is actively measured at an "echo time" (also known as `TE`_).
+
+The time it takes for 37% of the protons fall back in line with :math:`B_0`
+is known as the transverse relaxation time, or "T2".
+The exact value of T2 varies across voxels in the brain, based on tissue type,
+which is why protocols which are T2-weighted or which quantitatively measure
+T2 are useful for structural analyses.
+
+However,
+
+To start, let's look at the most common type of fMRI, single-echo EPI.
+
+What is single-echo fMRI?
+-------------------------
+
+
+
+When you have a single echo time, you acquire one value for each voxel, at each time point.
+
+.. image:: /_static/mr_schematic_single_echo.png
+
+This works rather well, in that there is relatively high BOLD contrast as every voxel,
+although the level of contrast (and thus the signal-to-noise ratio) will vary across the brain.
+
+There are many factors that impact BOLD signal, but we will focus on a small number.
+First, we have neurally-driven BOLD signal.
+This is the signal we generally care about in fMRI.
+Next, we have non-BOLD noise.
+This noise is often driven by things like instrument noise, subject motion, and thermal noise.
+Finally, we have non-neural BOLD signal.
+There are physiological sources of changes in blood oxygenation that are unrelated to neural activity,
+including heart rate and breathing changes.
+
+However, this ignores a number of useful features of the BOLD signal, including:
+(1) BOLD signal decays as echo time increases, (2) BOLD contrast _increases_ as echo time increases,
+(3) BOLD fluctuations may be driven by
+
+.. _multi-echo physics:
+
+The physics of multi-echo fMRI
+------------------------------
+Multi-echo fMRI data is obtained by acquiring multiple echo times (commonly called
+`TEs`_) for each MRI volume during data collection.
+While fMRI signal contains important neural information (termed the blood
+oxygen-level dependent, or `BOLD signal`_,
+it also contains "noise" (termed non-BOLD signal) caused by things like
+participant motion and changes in breathing.
+Because the BOLD signal is known to decay at a set rate, collecting multiple
+echos allows us to assess non-BOLD.
+
+The image below shows the basic relationship between echo times and the image acquired at
+3T (top, A) and 7T (bottom, B). Note that the earliest echo time is the brightest, as the
+signal has only had a limited amount of time to decay.
+In addition, the latter echo times show areas in which is the signal has decayed completely ('drop out')
+due to inhomogeneity in the magnetic field.
+By using the information across multiple echos these images can be combined in
+an optimal manner to take advantage of the signal
+in the earlier echos (see :ref:`optimal combination`).
+
+.. image:: /_static/physics_signal_decay.png
+
+.. image:: /_static/physics_multiple_echos.png
+
[…]
