### 2021-04-16 4813a58
Fill out docstrings for file-writing class.
