+++ tedana/io.py
@@ -1,7 +1,7 @@
 """The io module handles most file input and output in the `tedana` workflow.
 
-Other functions in the module help simplify writing out
-data from multiple echoes or write very complex outputs.
+Other functions in the module help simplify writing out data from multiple echoes or write very
+complex outputs.
 """
 import logging
 import os
@@ -26,6 +26,36 @@ RefLGR = logging.getLogger('REFERENCES')
 
 class OutputGenerator():
     """A class for managing tedana outputs.
+
+    Parameters
+    ----------
+    reference_img : img_like
+        The reference image which defines affine, shape, etc. of output images.
+    convention : {"bidsv1.5.0", "orig", or other str}, optional
+        Default is "bidsv1.5.0". Must correspond to a key in ``config``.
+    out_dir : str, optional
+        Output directory. Default is current working directory (".").
+    prefix : None or str, optional
+        Prefix to prepend to output filenames. Default is None, which means no prefix will be used.
+    config : str, optional
+        Path to configuration json file, which determines appropriate filenames based on file
+        descriptions. Default is "auto", which uses tedana's default configuration file.
+
+    Attributes
+    ----------
+    config : dict
+        File naming configuration information.
+    reference_img : img_like
+        The reference image which defines affine, shape, etc. of output images.
+    convention : str
+        The naming convention for output files.
+    out_dir : str
+        Directory in which outputs will be saved.
+    figures_dir : str
+        Directory in which figures will be saved.
+        This will correspond to a "figures" subfolder of ``out_dir``.
+    prefix : str
+        Prefix to prepend to output filenames.
     """
     def __init__(
         self,
@@ -63,48 +93,79 @@ class OutputGenerator():
             LGR.info(f"Generating figures directory: {self.figures_dir}")
             os.mkdir(self.figures_dir)
 
-    def _determine_suffix(self, description, name):
-        if description.endswith("img"):
-            allowed_suffixes = [".nii", ".nii.gz"]
-            preferred_suffix = ".nii.gz"
-        elif description.endswith("json"):
-            allowed_suffixes = [".json"]
-            preferred_suffix = ".json"
-        elif description.endswith("tsv"):
-            allowed_suffixes = [".tsv"]
-            preferred_suffix = ".tsv"
-
-        if not any(name.endswith(suff) for suff in allowed_suffixes):
-            suffix = preferred_suffix
-
-        return suffix
-
-    def get_name(self, description, **kwargs):
-        """Generates an image file full path to simplify file output
+    def _determine_extension(self, description, name):
+        """Infer the extension for a file based on its description.
 
         Parameters
         ----------
         description : str
-            The description of the file. Must be a key in self.config.img_table
+            The description of the file. Corresponds to a key in ``self.config``.
+        name : str
+            Filename corresponding to the description within ``self.config``.
+
+        Returns
+        -------
+        extension : str
+            File extension for the filename.
+        """
+        if description.endswith("img"):
+            allowed_extensions = [".nii", ".nii.gz"]
+            preferred_extension = ".nii.gz"
+        elif description.endswith("json"):
+            allowed_extensions = [".json"]
+            preferred_extension = ".json"
+        elif description.endswith("tsv"):
+            allowed_extensions = [".tsv"]
+            preferred_extension = ".tsv"
+
+        if not any(name.endswith(ext) for ext in allowed_extensions):
+            extension = preferred_extension
+
+        return extension
+
+    def get_name(self, description, **kwargs):
+        """Generate a file full path to simplify file output.
+
+        Parameters
+        ----------
+        description : str
+            The description of the file. Must be a key in ``self.config``.
         kwargs : keyword arguments
+            Additional arguments used to format the base filename string.
[…]
