+++ tedana/workflows/tedana.py
@@ -344,10 +344,6 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
     if not op.isdir(out_dir):
         os.mkdir(out_dir)
 
-    io.outdir = out_dir
-    io.set_prefix(prefix)
-    io.set_convention(convention)
-
     # boilerplate
     basename = 'report'
     extension = 'txt'
@@ -441,20 +437,20 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
     if mixm is not None and op.isfile(mixm):
         mixm = op.abspath(mixm)
         # Allow users to re-run on same folder
-        mixing_name = io.gen_tsv_name("ICA mixing")
+        mixing_name = generator.get_name("ICA mixing tsv")
         if mixm != mixing_name:
             shutil.copyfile(mixm, mixing_name)
-            shutil.copyfile(mixm, op.join(out_dir, op.basename(mixm)))
+            shutil.copyfile(mixm, op.join(generator.out_dir, op.basename(mixm)))
     elif mixm is not None:
         raise IOError('Argument "mixm" must be an existing file.')
 
     if ctab is not None and op.isfile(ctab):
         ctab = op.abspath(ctab)
         # Allow users to re-run on same folder
-        metrics_name = io.gen_tsv_name("ICA metrics")
+        metrics_name = generator.get_name("ICA metrics tsv")
         if ctab != metrics_name:
             shutil.copyfile(ctab, metrics_name)
-            shutil.copyfile(ctab, op.join(out_dir, op.basename(ctab)))
+            shutil.copyfile(ctab, op.join(generator.out_dir, op.basename(ctab)))
     elif ctab is not None:
         raise IOError('Argument "ctab" must be an existing file.')
 
@@ -469,10 +465,10 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
         manacc = [int(m) for m in manacc]
 
     if t2smap is not None and op.isfile(t2smap):
-        t2smap_file = io.gen_img_name('t2star map')
+        t2smap_file = generator.get_name('t2star img')
         t2smap = op.abspath(t2smap)
         # Allow users to re-run on same folder
-        if t2smap != io.gen_img_name('t2star map'):
+        if t2smap != t2smap_file:
             shutil.copyfile(t2smap, t2smap_file)
     elif t2smap is not None:
         raise IOError('Argument "t2smap" must be an existing file.')
@@ -519,12 +515,12 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
         LGR.debug('Setting cap on T2* map at {:.5f}s'.format(
             utils.millisec2sec(cap_t2s)))
         t2s_limited[t2s_limited > cap_t2s * 10] = cap_t2s
-        generator.save_file(utils.millisec2sec(t2s_limited), 't2star map')
-        generator.save_file(s0_limited, 's0 map')
+        generator.save_file(utils.millisec2sec(t2s_limited), 't2star img')
+        generator.save_file(s0_limited, 's0 img')
 
         if verbose:
-            generator.save_file(utils.millisec2sec(t2s_full), 'full t2star map')
-            generator.save_file(s0_full, 'full s0 map')
+            generator.save_file(utils.millisec2sec(t2s_full), 'full t2star img')
+            generator.save_file(s0_full, 'full s0 img')
 
     # optimally combine data
     data_oc = combine.make_optcom(catd, tes, masksum, t2s=t2s_full, combmode=combmode)
@@ -533,7 +529,7 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
     if 'gsr' in gscontrol:
[…]
