+++ tedana/gscontrol.py
@@ -8,7 +8,7 @@ import pandas as pd
 from scipy import stats
 from scipy.special import lpmv
 
-from tedana import io, utils
+from tedana import utils
 from tedana.due import due, Doi
 
 LGR = logging.getLogger(__name__)
@@ -16,7 +16,7 @@ RepLGR = logging.getLogger("REPORT")
 RefLGR = logging.getLogger("REFERENCES")
 
 
-def gscontrol_raw(catd, optcom, n_echos, ref_img, out_dir='.', dtrank=4):
+def gscontrol_raw(catd, optcom, n_echos, generator, dtrank=4):
     """
     Removes global signal from individual echo `catd` and `optcom` time series
 
@@ -78,11 +78,7 @@ def gscontrol_raw(catd, optcom, n_echos, ref_img, out_dir='.', dtrank=4):
     detr = dat - np.dot(sol.T, Lmix.T)[0]
     sphis = (detr).min(axis=1)
     sphis -= sphis.mean()
-    io.filewrite(
-        utils.unmask(sphis, Gmask),
-        'gs map',
-        ref_img
-    )
+    generator.save_file(utils.unmask(sphis, Gmask), "gs img")
 
     # find time course ofc the spatial global signal
     # make basis with the Legendre basis
@@ -90,8 +86,7 @@ def gscontrol_raw(catd, optcom, n_echos, ref_img, out_dir='.', dtrank=4):
     glsig = stats.zscore(glsig, axis=None)
 
     glsig_df = pd.DataFrame(data=glsig.T, columns=['global_signal'])
-    glsig_df.to_csv(io.gen_tsv_name("global signal time series"),
-                    sep='\t', index=False)
+    generator.save_file(glsig_df, "global signal time series tsv")
     glbase = np.hstack([Lmix, glsig.T])
 
     # Project global signal out of optimally combined data
@@ -99,17 +94,9 @@ def gscontrol_raw(catd, optcom, n_echos, ref_img, out_dir='.', dtrank=4):
     tsoc_nogs = dat - np.dot(np.atleast_2d(sol[dtrank]).T,
                              np.atleast_2d(glbase.T[dtrank])) + Gmu[Gmask][:, np.newaxis]
 
-    io.filewrite(
-        optcom,
-        'has gs combined',
-        ref_img
-    )
+    generator.save_file(optcom, "has gs combined img")
     dm_optcom = utils.unmask(tsoc_nogs, Gmask)
-    io.filewrite(
-        dm_optcom,
-        'removed gs combined',
-        ref_img
-    )
+    generator.save_file(dm_optcom, "removed gs combined img")
 
     # Project glbase out of each echo
     dm_catd = catd.copy()  # don't overwrite catd
@@ -126,7 +113,7 @@ def gscontrol_raw(catd, optcom, n_echos, ref_img, out_dir='.', dtrank=4):
 @due.dcite(Doi("10.1073/pnas.1301725110"),
            description="Minimum image regression to remove T1-like effects "
                        "from the denoised data.")
-def minimum_image_regression(optcom_ts, mmix, mask, comptable, ref_img, out_dir="."):
+def minimum_image_regression(optcom_ts, mmix, mask, comptable, generator):
     """
     Perform minimum image regression (MIR) to remove T1-like effects from
     BOLD-like components.
@@ -217,11 +204,7 @@ def minimum_image_regression(optcom_ts, mmix, mask, comptable, ref_img, out_dir=
     mehk_ts = np.dot(comp_pes[:, acc], mmix[:, acc].T)
     t1_map = mehk_ts.min(axis=-1)  # map of T1-like effect
     t1_map -= t1_map.mean()
[…]

[…]
