### 2021-04-16 18e7df7
Update docstrings with correct references, details
