### 2021-04-25 f0736e6
return list of tuples for iterating
