### 2021-04-28 8477aab
[MAINT] Modifies actions to run on release publish (#725)
