### 2021-04-30 b83ffa5
Progress on refactor
