+++ tedana/decomposition/pca.py
@@ -202,8 +202,8 @@ def tedpca(data_cat, data_oc, combmode, mask, adaptive_mask, t2sG,
     data_z = (data_z - data_z.mean()) / data_z.std()  # var normalize everything
 
     if algorithm in ['mdl', 'aic', 'kic']:
-        data_img = io.new_nii_like(generator.reference_image, utils.unmask(data, mask))
-        mask_img = io.new_nii_like(generator.reference_image, mask.astype(int))
+        data_img = io.new_nii_like(generator.reference_img, utils.unmask(data, mask))
+        mask_img = io.new_nii_like(generator.reference_img, mask.astype(int))
         voxel_comp_weights, varex, varex_norm, comp_ts = ma_pca(
             data_img, mask_img, algorithm, normalize=True)
     elif isinstance(algorithm, Number):
@@ -243,7 +243,7 @@ def tedpca(data_cat, data_oc, combmode, mask, adaptive_mask, t2sG,
 
     # write component maps to 4D image
     comp_maps = utils.unmask(computefeats2(data_oc, comp_ts, mask), mask)
-    generator.save_file(comp_maps, 'z-scored PCA components')
+    generator.save_file(comp_maps, 'z-scored PCA components img')
 
     # Select components using decision tree
     if algorithm == 'kundu':
