+++ tedana/tests/test_model_fit_dependence_metrics.py
@@ -1,11 +1,17 @@
 """
 Tests for tedana.metrics.fit
 """
+import os
 
 import numpy as np
 import pytest
 
 from tedana.metrics import kundu_fit
+from tedana.io import OutputGenerator
+from tedana.tests.utils import get_test_data_path
+
+data_dir = get_test_data_path()
+ref_img = os.path.join(data_dir, 'mask.nii.gz')
 
 
 def test_break_dependence_metrics():
@@ -19,14 +25,14 @@ def test_break_dependence_metrics():
     mmix = np.empty((n_vols, n_comps))
     adaptive_mask = np.empty((n_samples))
     tes = np.empty((n_echos))
-    ref_img = ''
+    generator = OutputGenerator(ref_img)
 
     # Shape of catd is wrong
     catd = np.empty((n_samples + 1, n_echos, n_vols))
     with pytest.raises(ValueError):
         kundu_fit.dependence_metrics(
             catd=catd, tsoc=tsoc, mmix=mmix,
-            adaptive_mask=adaptive_mask, tes=tes, ref_img=ref_img,
+            adaptive_mask=adaptive_mask, tes=tes, generator=generator,
             reindex=False, mmixN=None, algorithm='kundu_v3')
 
     # Shape of adaptive_mask is wrong
@@ -35,7 +41,7 @@ def test_break_dependence_metrics():
     with pytest.raises(ValueError):
         kundu_fit.dependence_metrics(
             catd=catd, tsoc=tsoc, mmix=mmix,
-            adaptive_mask=adaptive_mask, tes=tes, ref_img=ref_img,
+            adaptive_mask=adaptive_mask, tes=tes, generator=generator,
             reindex=False, mmixN=None, algorithm='kundu_v3')
 
     # Shape of tsoc is wrong
@@ -44,7 +50,7 @@ def test_break_dependence_metrics():
     with pytest.raises(ValueError):
         kundu_fit.dependence_metrics(
             catd=catd, tsoc=tsoc, mmix=mmix,
-            adaptive_mask=adaptive_mask, tes=tes, ref_img=ref_img,
+            adaptive_mask=adaptive_mask, tes=tes, generator=generator,
             reindex=False, mmixN=None, algorithm='kundu_v3')
 
     # Shape of catd is wrong
@@ -53,7 +59,7 @@ def test_break_dependence_metrics():
     with pytest.raises(ValueError):
         kundu_fit.dependence_metrics(
             catd=catd, tsoc=tsoc, mmix=mmix,
-            adaptive_mask=adaptive_mask, tes=tes, ref_img=ref_img,
+            adaptive_mask=adaptive_mask, tes=tes, generator=generator,
             reindex=False, mmixN=None, algorithm='kundu_v3')
 
     # Shape of catd is wrong
@@ -61,5 +67,5 @@ def test_break_dependence_metrics():
     with pytest.raises(ValueError):
         kundu_fit.dependence_metrics(
             catd=catd, tsoc=tsoc, mmix=mmix,
-            adaptive_mask=adaptive_mask, tes=tes, ref_img=ref_img,
+            adaptive_mask=adaptive_mask, tes=tes, generator=generator,
             reindex=False, mmixN=None, algorithm='kundu_v3')
