+++ tedana/workflows/t2smap.py
@@ -1,7 +1,6 @@
 """
 Estimate T2 and S0, and optimally combine data across TEs.
 """
-import json
 import os
 import os.path as op
 import logging
