### 2021-04-30 61f5fbf
Switch to fstrings where possible
