### 2021-04-30 0ed2ad9
Update tedana/io.py
Co-authored-by: Taylor Salo <tsalo006@fiu.edu>
