### 2021-04-30 21b8055
Update tedana/io.py
Co-authored-by: Taylor Salo <tsalo006@fiu.edu>
