### 2021-05-04 5045fc3
Address @handwerkerd docstring review
