+++ tedana/gscontrol.py
@@ -16,7 +16,7 @@ RepLGR = logging.getLogger("REPORT")
 RefLGR = logging.getLogger("REFERENCES")
 
 
-def gscontrol_raw(catd, optcom, n_echos, generator, dtrank=4):
+def gscontrol_raw(catd, optcom, n_echos, io_generator, dtrank=4):
     """
     Removes global signal from individual echo `catd` and `optcom` time series
 
@@ -34,10 +34,8 @@ def gscontrol_raw(catd, optcom, n_echos, generator, dtrank=4):
         Optimally combined functional data (i.e., the output of `make_optcom`)
     n_echos : :obj:`int`
         Number of echos in data. Should be the same as `E` dimension of `catd`
-    ref_img : :obj:`str` or img_like
-        Reference image to dictate how outputs are saved to disk
-    out_dir : :obj:`str`, optional
-        Output directory.
+    io_generator : :obj:`tedana.io.OutputGenerator`
+        The output generator for this workflow
     dtrank : :obj:`int`, optional
         Specifies degree of Legendre polynomial basis function for estimating
         spatial global signal. Default: 4
@@ -78,7 +76,7 @@ def gscontrol_raw(catd, optcom, n_echos, generator, dtrank=4):
     detr = dat - np.dot(sol.T, Lmix.T)[0]
     sphis = (detr).min(axis=1)
     sphis -= sphis.mean()
-    generator.save_file(utils.unmask(sphis, Gmask), "gs img")
+    io_generator.save_file(utils.unmask(sphis, Gmask), "gs img")
 
     # find time course ofc the spatial global signal
     # make basis with the Legendre basis
@@ -86,7 +84,7 @@ def gscontrol_raw(catd, optcom, n_echos, generator, dtrank=4):
     glsig = stats.zscore(glsig, axis=None)
 
     glsig_df = pd.DataFrame(data=glsig.T, columns=['global_signal'])
-    generator.save_file(glsig_df, "global signal time series tsv")
+    io_generator.save_file(glsig_df, "global signal time series tsv")
     glbase = np.hstack([Lmix, glsig.T])
 
     # Project global signal out of optimally combined data
@@ -94,9 +92,9 @@ def gscontrol_raw(catd, optcom, n_echos, generator, dtrank=4):
     tsoc_nogs = dat - np.dot(np.atleast_2d(sol[dtrank]).T,
                              np.atleast_2d(glbase.T[dtrank])) + Gmu[Gmask][:, np.newaxis]
 
-    generator.save_file(optcom, "has gs combined img")
+    io_generator.save_file(optcom, "has gs combined img")
     dm_optcom = utils.unmask(tsoc_nogs, Gmask)
-    generator.save_file(dm_optcom, "removed gs combined img")
+    io_generator.save_file(dm_optcom, "removed gs combined img")
 
     # Project glbase out of each echo
     dm_catd = catd.copy()  # don't overwrite catd
@@ -113,7 +111,7 @@ def gscontrol_raw(catd, optcom, n_echos, generator, dtrank=4):
 @due.dcite(Doi("10.1073/pnas.1301725110"),
            description="Minimum image regression to remove T1-like effects "
                        "from the denoised data.")
-def minimum_image_regression(optcom_ts, mmix, mask, comptable, generator):
+def minimum_image_regression(optcom_ts, mmix, mask, comptable, io_generator):
     """
     Perform minimum image regression (MIR) to remove T1-like effects from
     BOLD-like components.
@@ -133,10 +131,8 @@ def minimum_image_regression(optcom_ts, mmix, mask, comptable, generator):
     comptable : (C x X) :obj:`pandas.DataFrame`
[…]
