### 2021-05-04 aa499b5
Use class for managing outputs (#711)
* Initial work on a class.

* Clean up.

* More work.

* Revert mistake.

* [FIX] Adds f_max to right place (#712)

Moves f_max thresholding to the step where the R2/S0 maps are written out, rather than just before kappa/rho calculation.

* [FIX] Calculate Kappa and Rho on full F-statistic maps (#714)

* Calculate rho and kappa on full maps.

* Add missing test outputs.

* Revert changes to outputs.

* Fill out docstrings for file-writing class.

* More work.

* [MAINT] Drop 3.5 support and begin 3.8 and 3.9 support (#721)

* Drop 3.5 support and add 3.8/3
[…]
