+++ tedana/workflows/tedana.py
@@ -344,10 +344,6 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
     if not op.isdir(out_dir):
         os.mkdir(out_dir)
 
-    io.outdir = out_dir
-    io.set_prefix(prefix)
-    io.set_convention(convention)
-
     # boilerplate
     basename = 'report'
     extension = 'txt'
@@ -416,11 +412,19 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
 
     LGR.info('Loading input data: {}'.format([f for f in data]))
     catd, ref_img = io.load_data(data, n_echos=n_echos)
+    io_generator = io.OutputGenerator(
+        ref_img,
+        convention=convention,
+        out_dir=out_dir,
+        prefix=prefix,
+        config="auto",
+    )
+
     n_samp, n_echos, n_vols = catd.shape
     LGR.debug('Resulting data shape: {}'.format(catd.shape))
 
     # check if TR is 0
-    img_t_r = ref_img.header.get_zooms()[-1]
+    img_t_r = io_generator.reference_img.header.get_zooms()[-1]
     if img_t_r == 0:
         raise IOError('Dataset has a TR of 0. This indicates incorrect'
                       ' header information. To correct this, we recommend'
@@ -433,20 +437,20 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
     if mixm is not None and op.isfile(mixm):
         mixm = op.abspath(mixm)
         # Allow users to re-run on same folder
-        mixing_name = io.gen_tsv_name("ICA mixing")
+        mixing_name = io_generator.get_name("ICA mixing tsv")
         if mixm != mixing_name:
             shutil.copyfile(mixm, mixing_name)
-            shutil.copyfile(mixm, op.join(out_dir, op.basename(mixm)))
+            shutil.copyfile(mixm, op.join(io_generator.out_dir, op.basename(mixm)))
     elif mixm is not None:
         raise IOError('Argument "mixm" must be an existing file.')
 
     if ctab is not None and op.isfile(ctab):
         ctab = op.abspath(ctab)
         # Allow users to re-run on same folder
-        metrics_name = io.gen_tsv_name("ICA metrics")
+        metrics_name = io_generator.get_name("ICA metrics tsv")
         if ctab != metrics_name:
             shutil.copyfile(ctab, metrics_name)
-            shutil.copyfile(ctab, op.join(out_dir, op.basename(ctab)))
+            shutil.copyfile(ctab, op.join(io_generator.out_dir, op.basename(ctab)))
     elif ctab is not None:
         raise IOError('Argument "ctab" must be an existing file.')
 
@@ -461,10 +465,10 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
         manacc = [int(m) for m in manacc]
 
     if t2smap is not None and op.isfile(t2smap):
-        t2smap_file = io.gen_img_name('t2star map')
+        t2smap_file = io_generator.get_name('t2star img')
         t2smap = op.abspath(t2smap)
         # Allow users to re-run on same folder
-        if t2smap != io.gen_img_name('t2star map'):
+        if t2smap != t2smap_file:
             shutil.copyfile(t2smap, t2smap_file)
     elif t2smap is not None:
         raise IOError('Argument "t2smap" must be an existing file.')
@@ -489,7 +493,7 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
         mask[t2s_limited == 0] = 0  # reduce mask based on T2* map
     else:
         LGR.info('Computing EPI mask from first echo')
-        first_echo_img = io.new_nii_like(ref_img, catd[:, 0, :])
[…]
