### 2021-05-05 ad664ae
[ENH] Generate BIDS Derivatives-compatible outputs (#691)
* BIDS Derivatives-compatible outputs.

* Switches to class-based output generator for file naming

Co-authored-by: Taylor Salo <tsalo006@fiu.edu>
Co-authored-by: Joshua Teves <joshua.teves@nih.gov>
