+++ tedana/io.py
@@ -1,66 +1,331 @@
+"""The io module handles most file input and output in the `tedana` workflow.
+
+Other functions in the module help write outputs which require multiple
+data sources, assist in writing per-echo verbose outputs, or act as helper
+functions for any of the above.
 """
-Functions to handle file input/output
-"""
-import json
 import logging
+import os
 import os.path as op
+import json
+from string import Formatter
 
 import numpy as np
-import pandas as pd
 import nibabel as nib
-from nibabel.filename_parser import splitext_addext
+import pandas as pd
 from nilearn._utils import check_niimg
 from nilearn.image import new_img_like
 
 from tedana import utils
 from tedana.stats import computefeats2, get_coeffs
 
+
 LGR = logging.getLogger(__name__)
 RepLGR = logging.getLogger('REPORT')
 RefLGR = logging.getLogger('REFERENCES')
 
 
-def split_ts(data, mmix, mask, comptable):
-    """
-    Splits `data` time series into accepted component time series and remainder
+class OutputGenerator():
+    """A class for managing tedana outputs.
 
     Parameters
     ----------
-    data : (S x T) array_like
-        Input data, where `S` is samples and `T` is time
-    mmix : (T x C) array_like
-        Mixing matrix for converting input data to component space, where `C`
-        is components and `T` is the same as in `data`
-    mask : (S,) array_like
-        Boolean mask array
-    comptable : (C x X) :obj:`pandas.DataFrame`
-        Component metric table. One row for each component, with a column for
-        each metric. Requires at least two columns: "component" and
-        "classification".
+    reference_img : img_like
+        The reference image which defines affine, shape, etc. of output images.
+    convention : {"bidsv1.5.0", "orig", or other str}, optional
+        Default is "bidsv1.5.0". Must correspond to a key in ``config``.
+    out_dir : str, optional
+        Output directory. Default is current working directory (".").
+    prefix : None or str, optional
+        Prefix to prepend to output filenames. Default is None, which means no prefix will be used.
+    config : str, optional
+        Path to configuration json file, which determines appropriate filenames based on file
+        descriptions. Default is "auto", which uses tedana's default configuration file.
+    make_figures : bool, optional
+        Whether or not to actually make a figures directory
+
+    Attributes
+    ----------
+    config : dict
+        File naming configuration information.
+    reference_img : img_like
+        The reference image which defines affine, shape, etc. of output images.
+    convention : str
+        The naming convention for output files.
+    out_dir : str
+        Directory in which outputs will be saved.
+    figures_dir : str
+        Directory in which figures will be saved.
+        This will correspond to a "figures" subfolder of ``out_dir``.
+    prefix : str
+        Prefix to prepend to output filenames.
+    """
+
+    def __init__(
+        self,
+        reference_img,
+        convention="bidsv1.5.0",
+        out_dir=".",
+        prefix="",
+        config="auto",
+        make_figures=True
+    ):
+
+        if config == "auto":
+            config = op.join(utils.get_resource_path(), "config", "outputs.json")
+
+        if convention == "bids":
+            # modify to update default bids convention number
+            convention = "bidsv1.5.0"
+
+        config = load_json(config)
+
+        cfg = {}
+        for k, v in config.items():
+            if convention not in v.keys():
+                raise ValueError(
+                    f"Convention {convention} is not one of the supported conventions "
+                    f"({', '.join(v.keys())})"
+                )
+            cfg[k] = v[convention]
+        self.config = cfg
+        self.reference_img = check_niimg(reference_img)
+        self.convention = convention
+        self.out_dir = op.abspath(out_dir)
[…]
