+++ tedana/selection/tedica.py
@@ -14,7 +14,7 @@ RepLGR = logging.getLogger('REPORT')
 RefLGR = logging.getLogger('REFERENCES')
 
 
-def manual_selection(comptable, metric_metadata, acc=None, rej=None):
+def manual_selection(comptable, acc=None, rej=None):
     """
     Perform manual selection of components.
 
@@ -22,9 +22,6 @@ def manual_selection(comptable, metric_metadata, acc=None, rej=None):
     ----------
     comptable : (C x M) :obj:`pandas.DataFrame`
         Component metric table, where `C` is components and `M` is metrics
-    metric_metadata : :obj:`dict`
-        Dictionary with metadata about calculated metrics.
-        Each entry corresponds to a column in ``comptable``.
     acc : :obj:`list`, optional
         List of accepted components. Default is None.
     rej : :obj:`list`, optional
@@ -46,26 +43,6 @@ def manual_selection(comptable, metric_metadata, acc=None, rej=None):
             'original_classification' not in comptable.columns):
         comptable['original_classification'] = comptable['classification']
         comptable['original_rationale'] = comptable['rationale']
-        metric_metadata["original_classification"] = {
-            "LongName": "Original classification",
-            "Description": (
-                "Classification from the original decision tree."
-            ),
-            "Levels": {
-                "accepted": "A BOLD-like component included in denoised and high-Kappa data.",
-                "rejected": "A non-BOLD component excluded from denoised and high-Kappa data.",
-                "ignored": (
-                    "A low-variance component included in denoised, "
-                    "but excluded from high-Kappa data."),
-            },
-        }
-        metric_metadata["original_rationale"] = {
-            "LongName": "Original rationale",
-            "Description": (
-                "The reason for the original classification. "
-                "Please see tedana's documentation for information about possible rationales."
-            ),
-        }
 
     comptable['classification'] = 'accepted'
     comptable['rationale'] = ''
@@ -95,34 +72,13 @@ def manual_selection(comptable, metric_metadata, acc=None, rej=None):
     comptable.loc[ign, 'classification'] = 'ignored'
     comptable.loc[ign, 'rationale'] += 'I001;'
 
-    metric_metadata["classification"] = {
-        "LongName": "Component classification",
-        "Description": (
-            "Classification from the manual classification procedure."
-        ),
-        "Levels": {
-            "accepted": "A BOLD-like component included in denoised and high-Kappa data.",
-            "rejected": "A non-BOLD component excluded from denoised and high-Kappa data.",
-            "ignored": (
-                "A low-variance component included in denoised, "
-                "but excluded from high-Kappa data."),
-        },
-    }
-    metric_metadata["rationale"] = {
-        "LongName": "Rationale for component classification",
-        "Description": (
-            "The reason for the original classification. "
-            "Please see tedana's documentation for information about possible rationales."
-        ),
-    }
-
     # Move decision columns to end
     comptable = clean_dataframe(comptable)
     metric_metadata = collect.get_metadata(comptable)
     return comptable, metric_metadata
 
 
-def kundu_selection_v2(comptable, metric_metadata, n_echos, n_vols):
+def kundu_selection_v2(comptable, n_echos, n_vols):
     """
     Classify components as "accepted," "rejected," or "ignored" based on
     relevant metrics.
@@ -138,9 +94,6 @@ def kundu_selection_v2(comptable, metric_metadata, n_echos, n_vols):
     comptable : (C x M) :obj:`pandas.DataFrame`
         Component metric table. One row for each component, with a column for
         each metric. The index should be the component number.
-    metric_metadata : :obj:`dict`
[…]
