### 2021-05-14 #3103 @oliviaguest
I can but not in any immediate way — is it OK to take it on but "pause" it for a few weeks? 😊
