+++ tedana/reporting/html_report.py
@@ -1,9 +1,10 @@
-import pandas as pd
-from bokeh import (embed, layouts, models)
-from bokeh import __version__ as bokehversion
-from pathlib import Path
 from os.path import join as opj
+from pathlib import Path
 from string import Template
+
+import pandas as pd
+from bokeh import __version__ as bokehversion
+from bokeh import embed, layouts, models
 from tedana.info import __version__
 from tedana.reporting import dynamic_figures as df
 
@@ -24,15 +25,13 @@ def _update_template_bokeh(bokeh_id, about, bokeh_js):
     -------
     HTMLReport : an instance of a populated HTML report
     """
-    resource_path = Path(__file__).resolve().parent.joinpath('data', 'html')
+    resource_path = Path(__file__).resolve().parent.joinpath("data", "html")
 
-    body_template_name = 'report_body_template.html'
+    body_template_name = "report_body_template.html"
     body_template_path = resource_path.joinpath(body_template_name)
-    with open(str(body_template_path), 'r') as body_file:
+    with open(str(body_template_path), "r") as body_file:
         body_tpl = Template(body_file.read())
-    body = body_tpl.substitute(content=bokeh_id,
-                               about=about,
-                               javascript=bokeh_js)
+    body = body_tpl.substitute(content=bokeh_id, about=about, javascript=bokeh_js)
     return body
 
 
@@ -45,13 +44,15 @@ def _save_as_html(body):
     body : str
         Body for HTML report with embedded figures
     """
-    resource_path = Path(__file__).resolve().parent.joinpath('data', 'html')
-    head_template_name = 'report_head_template.html'
+    resource_path = Path(__file__).resolve().parent.joinpath("data", "html")
+    head_template_name = "report_head_template.html"
     head_template_path = resource_path.joinpath(head_template_name)
-    with open(str(head_template_path), 'r') as head_file:
+    with open(str(head_template_path), "r") as head_file:
         head_tpl = Template(head_file.read())
 
-    html = head_tpl.substitute(version=__version__, bokehversion=bokehversion, body=body)
+    html = head_tpl.substitute(
+        version=__version__, bokehversion=bokehversion, body=body
+    )
     return html
 
 
@@ -71,50 +72,65 @@ def generate_report(out_dir, tr):
         A generated HTML report
     """
     # Load the component time series
-    comp_ts_path = opj(out_dir, 'ica_mixing.tsv')
-    comp_ts_df = pd.read_csv(comp_ts_path, sep='\t', encoding='utf=8')
+    comp_ts_path = opj(out_dir, "ica_mixing.tsv")
+    comp_ts_df = pd.read_csv(comp_ts_path, sep="\t", encoding="utf=8")
     n_vols, n_comps = comp_ts_df.shape
 
     # Load the component table
-    comptable_path = opj(out_dir, 'ica_decomposition.json')
+    comptable_path = opj(out_dir, "ica_decomposition.json")
     comptable_cds = df._create_data_struct(comptable_path)
 
     # Create kappa rho plot
     kappa_rho_plot = df._create_kr_plt(comptable_cds)
 
     # Create sorted plots
-    kappa_sorted_plot = df._create_sorted_plt(comptable_cds, n_comps,
-                                              'kappa_rank', 'kappa',
-                                              'Kappa Rank', 'Kappa')
-    rho_sorted_plot = df._create_sorted_plt(comptable_cds, n_comps,
-                                            'rho_rank', 'rho',
-                                            'Rho Rank', 'Rho')
+    kappa_sorted_plot = df._create_sorted_plt(
[…]
