### 2021-05-15 f6a90ea
Rename functions to fix building error
