### 2021-05-19 37096ed
Revert "Rename functions to fix building error"
This reverts commit f6a90eac353d39f3f4be390e3ed2eede18cab4ef.
