+++ docs/conf.py
@@ -19,11 +19,11 @@
 #
 import os
 import sys
-
-sys.path.insert(0, os.path.abspath("sphinxext"))
+sys.path.insert(0, os.path.abspath('sphinxext'))
 sys.path.insert(0, os.path.abspath(os.path.pardir))
 
 from github_link import make_linkcode_resolve
+
 
 # -- General configuration ------------------------------------------------
 
@@ -39,42 +39,39 @@ add_module_names = False
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-    "matplotlib.sphinxext.plot_directive",
-    "sphinx.ext.autodoc",
-    "sphinx.ext.autosummary",
-    "sphinx.ext.doctest",
-    "sphinx.ext.ifconfig",
-    "sphinx.ext.intersphinx",
-    "sphinx.ext.linkcode",
-    "sphinx.ext.napoleon",
-    "sphinx.ext.todo",
-    "sphinxarg.ext",
+    'matplotlib.sphinxext.plot_directive',
+    'sphinx.ext.autodoc',
+    'sphinx.ext.autosummary',
+    'sphinx.ext.doctest',
+    'sphinx.ext.ifconfig',
+    'sphinx.ext.intersphinx',
+    'sphinx.ext.linkcode',
+    'sphinx.ext.napoleon',
+    'sphinx.ext.todo',
+    'sphinxarg.ext',
 ]
 
-from distutils.version import LooseVersion
-
 import sphinx
-
-if LooseVersion(sphinx.__version__) < LooseVersion("1.4"):
-    extensions.append("sphinx.ext.pngmath")
+from distutils.version import LooseVersion
+if LooseVersion(sphinx.__version__) < LooseVersion('1.4'):
+    extensions.append('sphinx.ext.pngmath')
 else:
-    extensions.append("sphinx.ext.imgmath")
+    extensions.append('sphinx.ext.imgmath')
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ["_templates"]
+templates_path = ['_templates']
 
 # source_suffix = ['.rst', '.md']
-source_suffix = ".rst"
+source_suffix = '.rst'
 
 # The master toctree document.
-master_doc = "index"
+master_doc = 'index'
 
 # General information about the project.
 from datetime import datetime  # access current time and date
-
-project = "tedana"
-copyright = "2017-" + datetime.today().strftime("%Y") + ", tedana developers"
-author = "tedana developers"
+project = 'tedana'
+copyright = '2017-' + datetime.today().strftime("%Y") + ', tedana developers'
+author = 'tedana developers'
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
@@ -82,7 +79,6 @@ author = "tedana developers"
 #
 # The short X.Y version.
 import tedana
-
 version = tedana.__version__
 # The full version, including alpha/beta/rc tags.
 release = tedana.__version__
@@ -97,10 +93,10 @@ language = None
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This patterns also effect to html_static_path and html_extra_path
-exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
+exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
 
 # The name of the Pygments (syntax highlighting) style to use.
-pygments_style = "sphinx"
+pygments_style = 'sphinx'
 
 # If true, `todo` and `todoList` produce output, else they produce nothing.
 todo_include_todos = False
@@ -128,44 +124,41 @@ napoleon_use_rtype = False
 #
 # installing theme package
 import sphinx_rtd_theme
-
[…]
