### 2021-05-19 fee8b48
Reverted style changes and updated function names.
