### 2021-05-24 c1f80cc
Reorder component table labels correctly
