### 2021-06-11 c1aa4b3
Don't unmask outputs of denoise_ts.
