### 2021-06-11 bb0361e
Use io_generator for carpet plots.
