### 2021-06-11 b723151
Split carpet plots into separate figures.
