### 2021-06-16 1ed70cb
[ENH] Allows iterative clustering (#732)
