### 2021-06-17 3bded74
Use two different adaptive masks.
masksum_denoise: Threshold of 1, used for denoising and optimal combination.
masksum_clf: Threshold of 3, used for decomposition and component selection.
