### 2021-06-17 ec2ccfe
[FIX] Add references back into the HTML report (#737)
