+++ tedana/workflows/tedana.py
@@ -712,38 +712,6 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
 
     LGR.info('Workflow completed')
 
-    if not no_reports:
-        LGR.info('Making figures folder with static component maps and timecourse plots.')
-
-        dn_ts, hikts, lowkts = io.denoise_ts(data_oc, mmix, mask, comptable)
-
-        reporting.static_figures.carpet_plot(
-            optcom_ts=data_oc,
-            denoised_ts=dn_ts,
-            hikts=hikts,
-            lowkts=lowkts,
-            mask=mask,
-            io_generator=io_generator,
-        )
-        reporting.static_figures.comp_figures(
-            data_oc,
-            mask=mask,
-            comptable=comptable,
-            mmix=mmix_orig,
-            io_generator=io_generator,
-            png_cmap=png_cmap,
-        )
-
-        if sys.version_info.major == 3 and sys.version_info.minor < 6:
-            warn_msg = ("Reports requested but Python version is less than "
-                        "3.6.0. Dynamic reports will not be generated.")
-            LGR.warn(warn_msg)
-        else:
-            LGR.info('Generating dynamic report')
-            reporting.generate_report(io_generator, tr=img_t_r)
-
-    LGR.info('Workflow completed')
-
     RepLGR.info("This workflow used numpy (Van Der Walt, Colbert, & "
                 "Varoquaux, 2011), scipy (Jones et al., 2001), pandas "
                 "(McKinney, 2010), scikit-learn (Pedregosa et al., 2011), "
@@ -784,13 +752,26 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
         fo.write(report)
 
     if not no_reports:
-        LGR.info('Making figures folder with static component maps and '
-                 'timecourse plots.')
-        reporting.static_figures.comp_figures(data_oc, mask=mask,
-                                              comptable=comptable,
-                                              mmix=mmix_orig,
-                                              io_generator=io_generator,
-                                              png_cmap=png_cmap)
+        LGR.info('Making figures folder with static component maps and timecourse plots.')
+
+        dn_ts, hikts, lowkts = io.denoise_ts(data_oc, mmix, mask, comptable)
+
+        reporting.static_figures.carpet_plot(
+            optcom_ts=data_oc,
+            denoised_ts=dn_ts,
+            hikts=hikts,
+            lowkts=lowkts,
+            mask=mask,
+            io_generator=io_generator,
+        )
+        reporting.static_figures.comp_figures(
+            data_oc,
+            mask=mask,
+            comptable=comptable,
+            mmix=mmix_orig,
+            io_generator=io_generator,
+            png_cmap=png_cmap,
+        )
 
         if sys.version_info.major == 3 and sys.version_info.minor < 6:
             warn_msg = ("Reports requested but Python version is less than "
