### 2021-06-20 1a54bb7
Derive classifier masks directly from denoising ones.
