+++ tedana/workflows/tedana.py
@@ -504,13 +504,15 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
         catd,
         mask=mask,
         getsum=True,
-        threshold=1
+        threshold=1,
     )
     LGR.debug('Retaining {}/{} samples for denoising'.format(mask_denoise.sum(), n_samp))
     io_generator.save_file(masksum_denoise, "adaptive mask img")
 
     # Create an adaptive mask with at least 3 good echoes, for classification
-    mask_clf, masksum_clf = utils.make_adaptive_mask(catd, mask=mask, getsum=True, threshold=3)
+    masksum_clf = masksum_denoise.copy()
+    masksum_clf[masksum_clf < 3] = 0
+    mask_clf = masksum_clf.astype(bool)
     LGR.debug('Retaining {}/{} samples for denoising'.format(mask_clf.sum(), n_samp))
 
     if t2smap is None:
