### 2021-06-21 1523934
Replace hardcoded dictionaries with json file.
