+++ docs/multi-echo.rst
@@ -1,3 +1,7 @@
+*********************
+About multi-echo fMRI
+*********************
+
 What is multi-echo fMRI?
 ========================
 Most echo-planar image (EPI) sequences collect a single brain image following
@@ -176,3 +180,352 @@ default, which can be inspected in order to determine the quality of denoising.
 See :ref:`outputs` for more information on these outputs.
 
 .. _t2smap: https://tedana.readthedocs.io/en/latest/usage.html#run-t2smap
+
+Acquiring multi-echo data
+=========================
+
+Available multi-echo fMRI sequences
+-----------------------------------
+We have attempted to compile some basic multi-echo fMRI protocols in an `OSF project`_.
+The parameter choices in these protocols run and seem reasonable, but they have
+not been optimized for a specific situation.
+They are a good starting point for someone designing a study, but should not be
+considered canonical.
+If you would like to use one of them, please customize it for your own purposes
+and make sure to run pilot scans to test your choices.
+
+.. _OSF project: https://osf.io/ebkrp/
+
+Siemens
+```````
+**For Siemens** users, there are two options for Works In Progress (WIPs) Sequences.
+
+* | The Center for Magnetic Resonance Research at the University of Minnesota
+  | provides a custom MR sequence that allows users to collect multiple echoes
+  | (termed **Contrasts**). The sequence and documentation can be `found here`_.
+  | For details on obtaining a license follow `this link`_.
+  | By default the number of contrasts is 1, yielding a single-echo sequence.
+  | In order to collect multiple echoes, increase number of Contrasts on the
+  | **Sequence Tab, Part 1** on the MR console.
+* | The Martinos Center at Harvard also has a MR sequence available, with the
+  | details `available here`_. The number of echoes can be specified on the
+  | **Sequence, Special** tab in this sequence.
+
+.. _found here: https://www.cmrr.umn.edu/multiband/
+.. _this link: http://license.umn.edu/technologies/cmrr_center-for-magnetic-resonance-research-software-for-siemens-mri-scanners
+.. _available here: https://www.nmr.mgh.harvard.edu/software/c2p/sms
+
+
+GE
+```
+**For GE users**, there are currently two sharable pulse sequences:
+
+* Multi-echo EPI (MEPI) – Software releases: DV24, MP24 and DV25 (with offline recon)
+* | Hyperband Multi-echo EPI (HyperMEPI) - Software releases: DV26, MP26, DV27, RX27
+  | (here hyperband can be deactivated to do simple Multi-echo EPI – online recon)
+
+Please reach out to the GE Research Operation team or each pulse sequence’s
+author to begin the process of obtaining this software.
+More information can be found on the `GE Collaboration Portal`_
+
+Once logged in, go to Groups > GE Works-in-Progress you can find the description
+of the current ATSM (i.e. prototypes).
+
+.. _GE Collaboration Portal: https://collaborate.mr.gehealthcare.com
+
+Philips
+```````
+**For Philips** users, sequences can be defined using product software.
+
+Multi-echo EPI (ME-EPI) can be acquired using the product software and can be combined with
+SENSE parallel imaging and MultiBand.
+The combination with MultiBand requires a SW release >R5.1 and MultiBand functionality to be present.
+No default ME-EPI are provided, but existing single-echo EPI sequences from the BOLD fMRI folder can be
+modified into multi-echo sequences by increasing the number of echoes.
+As a starting point to develop a 3 echo EPI protocol start by opening the default fMRI protocol and
+modify the following: increase number of echoes to 3 on the Contrast tab, set SENSE = 3, MB-SENSE = 3,
[…]
