+++ docs/outputs.rst
@@ -1,10 +1,11 @@
 .. _outputs:
 
+#################
 Outputs of tedana
-=================
+#################
 
-tedana derivatives
-------------------
+Outputs of the tedana workflow
+==============================
 
 ================================================    =====================================================
 Filename                                            Content
@@ -126,7 +127,7 @@ desc-TEDICAT1cDenoised_mixing.tsv                   T1-GS corrected mixing matri
 ================================================    =====================================================
 
 Component tables
-----------------
+================
 TEDPCA and TEDICA use component tables to track relevant metrics, component
 classifications, and rationales behind classifications.
 The component tables are stored as json files for BIDS-compatibility.
@@ -148,7 +149,7 @@ ignored            Low-variance components included in denoised, but excluded
 ===============    =============================================================
 
 TEDPCA codes
-````````````
+------------
 =====  ===============  ========================================================
 Code   Classification   Description
 =====  ===============  ========================================================
@@ -163,7 +164,7 @@ P007   rejected         Rho below fmin (only in stabilized PCA decision tree)
 =====  ===============  ========================================================
 
 TEDICA codes
-````````````
+------------
 =====  =================  ========================================================
 Code   Classification     Description
 =====  =================  ========================================================
@@ -184,7 +185,7 @@ I012   ignored            ign_add1
 =====  =================  ========================================================
 
 Citable workflow summaries
---------------------------
+==========================
 
 ``tedana`` generates a report for the workflow, customized based on the parameters used and including relevant citations.
 The report is saved in a plain-text file, report.txt, in the output directory.
@@ -216,3 +217,178 @@ An example report
   Sørensen, T. J. (1948). A method of establishing groups of equal amplitude in plant sociology based on similarity of species content and its application to analyses of the vegetation on Danish commons. I kommission hos E. Munksgaard.
 
   Van Der Walt, S., Colbert, S. C., & Varoquaux, G. (2011). The NumPy array: a structure for efficient numerical computation. Computing in Science & Engineering, 13(2), 22.
+
+.. _interactive reports:
+
+ICA Components Report
+=====================
+
+The reporting page for the tedana decomposition presents a series
+of interactive plots designed to help you evaluate the quality of your
+analyses. This page describes the plots forming the reports and well as
+information on how to take advantage of the interactive functionalities.
+You can also play around with `our demo`_.
+
+.. _our demo: https://me-ica.github.io/tedana-ohbm-2020/
+
+Report Structure
+----------------
+
+The image below shows a representative report, which has two sections: a) the summary view,
+and b) the individual component view.
+
+.. image:: /_static/rep01_overallview.png
+  :align: center
+
+.. note::
+  When a report is initially loaded, as no component is selected on the
+  summary view, the individual component view appears empty.
+
+Summary View
+````````````
+This view provides an overview of the decomposition and component
+selection results. It includes four different plots.
+
+* **Kappa/Rho Scatter:** This is a scatter plot of `Kappa` vs. `Rho` features for all components.
+  In the plot, each dot represents a different component. The x-axis represents the kappa feature, and the
+  y-axis represents the rho feature. These are two of the most
+  informative features describing the likelihood of the component
+  being BOLD or non-BOLD. Additional information is provided via color
+  and size. In particular, color informs about its classification
+  status (e.g., accepted, rejected); while size relates to
+  the amount of variance explained by the component (larger dot,
+  larger variance).
+
+.. image:: /_static/rep01_kapparhoScatter.png
+  :align: center
[…]
