### 2021-06-22 e15bdcd
Fix docstring.
