### 2021-07-06 30174b4
Remove sort_df and apply_sort functions.
