### 2021-07-06 ec97921
Remove sort_by and ascending parameters.
