+++ tedana/workflows/tedana.py
@@ -570,7 +570,7 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
             comptable, mmix = metrics.collect.generate_metrics(
                 catd, data_oc, mmix_orig, masksum, tes,
                 io_generator, 'ICA',
-                metrics=required_metrics, sort_by='kappa', ascending=False,
+                metrics=required_metrics,
             )
             comptable, metric_metadata = selection.kundu_selection_v2(
                 comptable, n_echos, n_vols
@@ -599,7 +599,7 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
             comptable, mmix = metrics.collect.generate_metrics(
                 catd, data_oc, mmix_orig, masksum, tes,
                 io_generator, 'ICA',
-                metrics=required_metrics, sort_by='kappa', ascending=False
+                metrics=required_metrics,
             )
             comptable, metric_metadata = selection.kundu_selection_v2(
                     comptable, n_echos, n_vols
