### 2021-07-06 0e4c52c
Clarify communication pathways.
