+++ docs/support.rst
@@ -1,7 +1,9 @@
 .. _support_ref:
 
+*************************
 Support and communication
--------------------------
+*************************
+
 All bugs, concerns and enhancement requests for this software can be submitted here: https://github.com/ME-ICA/tedana/issues.
 
 If you would like to ask a question about tedana-specific usage or tedana's outputs, please submit a question to `NeuroStars`_ with the `tedana tag`_.
@@ -15,3 +17,14 @@ We will also attempt to archive certain common questions and associate answers i
 .. _multi-echo tag: https://neurostars.org/tags/multi-echo
 .. _NeuroStars: https://neurostars.org
 .. _tedana tag: https://neurostars.org/tags/tedana
+
+Internal Communication Pathways
+===============================
+
+There are multiple mechanisms for development-related communication, which can lead to confusion.
+Here we attempt to clarify when and how each communication pathway should be used:
+
+- GitHub Issues and Pull Requests: Issues and PRs should be the **main avenue of communication** for questions and proposals related to tedana's development.
+- Gitter: While tedana does have a Gitter channel, this channel should not be used discussions. Instead, its primary use is to redirect to **existing** Issues and PRs, whether open or closed.
+- Monthly Newsletter: The monthly newsletter is used to make announcements from the tedana developers to the tedana users. Discussions that start in Issues or PRs may move to the newsletter if they are news-worthy.
+- NeuroStars: The NeuroStars forum is to be used by tedana users to ask questions about how to use tedana. tedana's developers may answer usage questions on NeuroStars, but development-related discussions should happen in GitHub Issues and PRs.
