### 2021-07-07 dd42364
Apply suggestions from code review
Co-authored-by: Eneko Uruñuela <13706448+eurunuela@users.noreply.github.com>
