### 2021-07-07 8e62b49
Finish reorganization and heading standardization.
Also add note about heading level formatting to CONTRIBUTING.md.
