+++ docs/multi-echo.rst
@@ -1,19 +1,24 @@
-*********************
+#####################
 About multi-echo fMRI
-*********************
+#####################
 
+
+************************
 What is multi-echo fMRI?
-========================
+************************
+
 Most echo-planar image (EPI) sequences collect a single brain image following
 a radio frequency (RF) pulse, at a rate known as the repetition time (TR).
 This typical approach is known as single-echo fMRI.
 In contrast, multi-echo (ME) fMRI refers to collecting data at multiple echo times,
 resulting in multiple volumes with varying levels of contrast acquired per RF pulse.
 
+
 .. _multi-echo physics:
 
 The physics of multi-echo fMRI
-------------------------------
+==============================
+
 Multi-echo fMRI data is obtained by acquiring multiple echo times (commonly called
 `TEs`_) for each MRI volume during data collection.
 While fMRI signal contains important neural information (termed the blood
@@ -72,7 +77,8 @@ For a more comprehensive review of these topics and others, see `Kundu et al. (2
 
 
 Why use multi-echo?
--------------------
+===================
+
 There are many potential reasons an investigator would be interested in using multi-echo EPI (ME-EPI).
 Among these are the different levels of analysis ME-EPI enables.
 Specifically, by collecting multi-echo data, researchers are able to:
@@ -102,16 +108,21 @@ We can use this information to denoise the optimally combined time series.
 
 
 Considerations for ME-fMRI
---------------------------
+==========================
+
 Multi-echo fMRI acquisition sequences and analysis methods are rapidly maturing.
 Someone who has access to a multi-echo fMRI sequence should seriously consider using it.
 
+
 Costs and benefits of multi-echo fMRI
--------------------------------------
+=====================================
+
 The following are a few points to consider when deciding whether or not to collect multi-echo data.
 
+
 Possible increase in TR
-```````````````````````
+-----------------------
+
 The one difference with multi-echo is a slight time cost.
 For multi-echo fMRI, the shortest echo time (TE) is essentially free since it is collected in the
 gap between the RF pulse and the single-echo acquisition.
@@ -128,8 +139,10 @@ Instead of compromising on slice coverage or TR, one can increase acceleration.
 If one increases acceleration, it is worth doing an empirical comparison to make sure there
 isn't a non-trivial loss in SNR or an increase of artifacts.
 
+
 Weighted averaging may lead to an increase in SNR
-`````````````````````````````````````````````````
+-------------------------------------------------
+
 Multiple studies have shown that a weighted average of the echoes to optimize
 T2* weighting, sometimes called "optimally combined," gives a reliable, modest
 boost in data quality.
@@ -140,8 +153,10 @@ In tedana, the weighted average can be calculated with
 If no other acquisition compromises are necessary to acquire multi-echo data,
 this boost is worthwhile.
 
+
 Consider the life of the dataset
-````````````````````````````````
+--------------------------------
+
 If other compromises are necessary, consider the life of the data set.
 If data is being acquired for a discrete study that will be acquired, analyzed,
 and published in a year or two, it might not be worth making compromises to
@@ -159,8 +174,10 @@ still being actively developed.
 Users need to have the time and knowledge to look at the denoising output from
 every run to make sure denoising worked as intended.
 
+
 You may recover signal in areas affected by dropout
-```````````````````````````````````````````````````
+---------------------------------------------------
+
 Typical single echo fMRI uses an echo time that is appropriate for signal
 across most of the brain.
[…]
