+++ docs/usage.rst
@@ -1,6 +1,6 @@
-**********************************
+##################################
 Using tedana from the command line
-**********************************
+##################################
 
 ``tedana`` minimally requires:
 
@@ -22,10 +22,13 @@ Users can also construct their own preprocessing pipelines from which to call
 .. _fMRIPrep: https://fmriprep.readthedocs.io
 .. _afni_proc.py: https://afni.nimh.nih.gov/pub/dist/doc/program_help/afni_proc.py.html
 
+
 .. _tedana cli:
 
+***************************
 Running the tedana workflow
-===========================
+***************************
+
 This is the full tedana workflow, which runs multi-echo ICA and outputs
 multi-echo denoised data along with many other derivatives.
 To see which files are generated by this workflow, check out the outputs page:
@@ -45,10 +48,13 @@ https://tedana.readthedocs.io/en/latest/outputs.html
     To examine regions-of-interest with multi-echo data, apply masks after TE
     Dependent ANAlysis.
 
+
 .. _t2smap cli:
 
+***************************
 Running the t2smap workflow
-===========================
+***************************
+
 This workflow uses multi-echo data to optimally combine data across echoes and
 to estimate T2* and S0 maps or time series.
 To see which files are generated by this workflow, check out the workflow
