+++ docs/roadmap.rst
@@ -1,8 +1,11 @@
+##################
 The tedana roadmap
-==================
+##################
 
+
+**************
 Project vision
---------------
+**************
 
 ``tedana`` was originally developed as a place for the multi-echo fMRI
 denoising method that was originally defined in ME-ICA
@@ -14,7 +17,114 @@ We have expanded to welcome additional multi-echo fMRI processing
 approaches, and to support communal resources for multi-echo fMRI, whether
 or not they directly involve the tedana software.
 
-A more detailed `project scope is here <https://tedana.readthedocs.io/en/latest/contributing.html#Scope-of-tedana>`_
+
+Scope of tedana
+===============
+
+tedana is a collection of tools, software and a community related to echo time
+(TE) dependent analyses. The umbrella of tedana covers a number of overlapping,
+but somewhat distinct, ideas related to multi-echo analysis. This scope includes
+collecting multi-echo data (Acquisition), combining those echoes together
+(Combination), with optional noise removal (Denoising), inspecting the outputs
+(Visualization) and answering multi-echo related questions (Community). In
+general, tedana accepts previously preprocessed data to produce outputs that
+are ready for further analyses.
+
+
+Acquisition
+-----------
+
+While the development of multi-echo sequences is beyond the current scope
+of tedana, the tedana community is committed to providing guidelines on current
+multi-echo implementations. This will include both specific instructions for
+how to collect multi-echo data for multiple vendors as well as details about
+what types of data have been collected thus far. These details are subject to
+change, and are intended to provide users with an idea of what is possible,
+rather than definitive recommendations.
+
+Our focus is on functional MRI, including both magnitude and phase data,
+however we understand that quantitative mapping has the potential to aid in
+data processing. Thus, we believe that some details on non-functional MRI
+acquisitions, such as detailed T2* mapping, may fall within the scope of
+tedana.
+Acquisition related details can be found in the `tedana Documentation.`_
+
+.. _tedana Documentation.: https://tedana.readthedocs.io/en/latest/acquisition.html
+
+
+Combining echoes
+----------------
+
+An early step in processing data collected with multiple echoes is the
+combination of the data into a single time series. We currently implement
+multiple options to combine multi-echo data and will add more as they continue
+to be developed. This is an area of active development and interest.
+
+
+Denoising
+---------
+
+tedana was developed out of a package known as `multi-echo ICA, ME-ICA, or MEICA`_
+developed by Dr. Prantik Kundu. Though the usage of ICA for classification of
+signal vs noise components has continued in tedana, this is not a rule. The
+tedana community is open and encouraging of new denoising methods, whether or not they
+have a basis in ICA.
+
+Specifically, we are interested in any method that seeks to use the information from multiple
+echoes to identify signal (defined here as BOLD signals arising from neural
+processing) and noise (defined here as changes unrelated to neural
+processing, such as motion, cardiac, respiration).
+
+tedana is primarily intended to work on volume data, that is, data that is
+still in structured voxel space. This is because several of the currently used denoising metrics rely on spatial continuity, and they have not yet been updated to consider continuity over cortical vertices.
+Therefore, surface-based denoising is not currently
+within the scope of tedana, but code could be written so that it is a
+possible option in the future.
+
+Currently tedana works on a single subject, run by run basis; however, methods
+that use information across multiple runs are welcome.
+
[…]

[…]
