### 2021-07-07 7bd6aa7
Temporarily disable, then reenable, report logger.
