### 2021-07-08 2e72e79
add resource info for and link to dash app
