### 2021-07-09 36b15b7
Revert "Added buttons to select carpet plot and added some structure to the reports"
This reverts commit fbe23b44235def50eeb28898da58deddbf8752c4.
