### 2021-07-09 b6e88b0
Move new changes from resources to multi-echo.
