### 2021-07-09 fe88c70
Delete resources.rst.
