+++ docs/resources.rst
@@ -1,110 +1,0 @@
-Resources
-=========
-
-Journal articles describing multi-echo methods
-----------------------------------------------
-* | :ref:`spreadsheet of publications` catalogues papers using multi-echo fMRI,
-  | with information about acquisition parameters.
-* | `Multi-echo acquisition`_
-  | Posse, NeuroImage 2012
-  | Includes an historical overview of multi-echo acquisition and research
-* | `Multi-Echo fMRI A Review of Applications in fMRI Denoising and Analysis of BOLD Signals`_
-  | Kundu et al, NeuroImage 2017
-  | A review of multi-echo denoising with a focus on the MEICA algorithm
-* | `Enhanced identification of BOLD-like components with MESMS and MEICA`_
-  | Olafsson et al, NeuroImage 2015
-  | The appendix includes a good explanation of the math underlying MEICA denoising
-* | `Comparing resting state fMRI de-noising approaches using multi- and single-echo acquisitions`_
-  | Dipasquale et al, PLoS One 2017
-  | The appendix includes some recommendations for multi-echo acquisition
-
-.. _Multi-echo acquisition: https://www.ncbi.nlm.nih.gov/pubmed/22056458
-.. _Multi-Echo fMRI A Review of Applications in fMRI Denoising and Analysis of BOLD Signals: https://www.ncbi.nlm.nih.gov/pubmed/28363836
-.. _Enhanced identification of BOLD-like components with MESMS and MEICA: https://www.ncbi.nlm.nih.gov/pubmed/25743045
-.. _Comparing resting state fMRI de-noising approaches using multi- and single-echo acquisitions: https://www.ncbi.nlm.nih.gov/pubmed/28323821
-
-Videos
-------
-* An `educational session from OHBM 2017`_ by Dr. Prantik Kundu about multi-echo denoising
-* A `series of lectures from the OHBM 2017 multi-echo session`_ on multiple facets of multi-echo data analysis
-* | Multi-echo fMRI lecture from the `2018 NIH FMRI Summer Course`_ by Javier Gonzalez-Castillo
-  | `Slides from 2018 NIH FMRI Summer Course`_
-
-.. _educational session from OHBM 2017: https://www.pathlms.com/ohbm/courses/5158/sections/7788/video_presentations/75977
-.. _series of lectures from the OHBM 2017 multi-echo session: https://www.pathlms.com/ohbm/courses/5158/sections/7822
-.. _2018 NIH FMRI Summer Course: https://fmrif.nimh.nih.gov/course/fmrif_course/2018/14_Javier_20180713
-.. _Slides from 2018 NIH FMRI Summer Course: https://fmrif.nimh.nih.gov/COURSE/fmrif_course/2018/content/14_Javier_20180713.pdf
-
-Multi-echo preprocessing software
----------------------------------
-tedana requires data that has already been preprocessed for head motion, alignment, etc.
-
-AFNI can process multi-echo data natively as well as apply tedana denoising through the use of
-**afni_proc.py**. To see various implementations, start with Example 12 in the `afni_proc.py help`_
-
-.. _afni_proc.py help: https://afni.nimh.nih.gov/pub/dist/doc/program_help/afni_proc.py.html
-
-`fmriprep` can also process multi-echo data, but is currently limited to using the optimally combined
-timeseries.
-For more details, see the `fmriprep workflows page`_.
-
-.. _fmriprep workflows page: https://fmriprep.readthedocs.io/en/stable/workflows.html
-
-Currently SPM and FSL do not natively support multi-echo fmri data processing.
-
-Other software that uses multi-echo fMRI
-----------------------------------------
-``tedana`` represents only one approach to processing multi-echo data.
-Currently there are a number of methods that can take advantage of or use the
-information contained in multi-echo data.
[…]
