### 2021-07-10 f968eb4
Added function to get system info and save into json
