+++ tedana/utils.py
@@ -3,6 +3,7 @@ Utilities for tedana package
 """
 import logging
 import os.path as op
+import platform
 
 import numpy as np
 import nibabel as nib
