### 2021-07-10 8044a99
Info table almost ready, need to figure out how to save json
