+++ tedana/reporting/html_report.py
@@ -5,10 +5,11 @@ from pathlib import Path
 from os.path import join as opj
 from string import Template
 from tedana.info import __version__
+from tedana.io import load_json
 from tedana.reporting import dynamic_figures as df
 
 
-def _update_template_bokeh(bokeh_id, about, bokeh_js):
+def _update_template_bokeh(bokeh_id, info_table, about, bokeh_js):
     """
     Populate a report with content.
 
@@ -16,6 +17,8 @@ def _update_template_bokeh(bokeh_id, about, bokeh_js):
     ----------
     bokeh_id : str
         HTML div created by bokeh.embed.components
+    info_table : str
+        HTML div created by _generate_info_table()
     about : str
         Reporting information for a given run
     bokeh_js : str
@@ -31,6 +34,7 @@ def _update_template_bokeh(bokeh_id, about, bokeh_js):
     with open(str(body_template_path), 'r') as body_file:
         body_tpl = Template(body_file.read())
     body = body_tpl.substitute(content=bokeh_id,
+                               info=info_table,
                                about=about,
                                javascript=bokeh_js)
     return body
@@ -53,6 +57,28 @@ def _save_as_html(body):
 
     html = head_tpl.substitute(version=__version__, bokehversion=bokehversion, body=body)
     return html
+
+
+def _generate_info_table(info_dict):
+    """Generate a table with relevant information about the
+    system and tedana.
+    """
+    resource_path = Path(__file__).resolve().parent.joinpath('data', 'html')
+
+    info_template_name = 'report_body_template.html'
+    info_template_path = resource_path.joinpath(info_template_name)
+    with open(str(info_template_path), 'r') as info_file:
+        info_tpl = Template(info_file.read())
+    info_html = info_tpl.substitute(command=info_dict["Command"],
+                                    system=info_dict["System"],
+                                    node=info_dict["Node"],
+                                    release=info_dict["Release"],
+                                    sysversion=info_dict["Version"],
+                                    machine=info_dict["Machine"],
+                                    processor=info_dict["Processor"],
+                                    python=info_dict["Python"],
+                                    tedana=__version__)
+    return info_html
 
 
 def generate_report(io_generator, tr):
@@ -114,7 +140,14 @@ def generate_report(io_generator, tr):
     with open(opj(io_generator.out_dir, 'report.txt'), 'r+') as f:
         about = f.read()
 
-    body = _update_template_bokeh(kr_div, about, kr_script)
+    # Read info table
+    infojson_path = io_generator.get_name("System info json")
+    info_dict = load_json(infojson_path)
+
+    # Create info table
+    info_table = _generate_info_table(info_dict)
+
+    body = _update_template_bokeh(kr_div, info_table, about, kr_script)
     html = _save_as_html(body)
     with open(opj(io_generator.out_dir, 'tedana_report.html'), 'wb') as f:
         f.write(html.encode('utf-8'))
