### 2021-07-12 952c4d3
Add MIR/GSR figures to carpet plots.
This doesn't toggle the new buttons based on workflow settings, so it will probably fail as is.
