### 2021-07-12 d449925
[REF] Eliminate component sorting from metric calculation (#741)
* Remove sort_df and apply_sort functions.

* Remove sort_by and ascending parameters.

* Update descriptions.
