+++ tedana/decomposition/pca.py
@@ -236,7 +236,7 @@ def tedpca(data_cat, data_oc, combmode, mask, adaptive_mask, t2sG,
     comptable, _ = metrics.collect.generate_metrics(
         data_cat, data_oc, comp_ts, adaptive_mask,
         tes, io_generator, 'PCA',
-        metrics=required_metrics, sort_by=None
+        metrics=required_metrics,
     )
 
     # varex_norm from PCA overrides varex_norm from dependence_metrics,
