### 2021-07-12 a2606e7
Add "optimal sign" metric and do not alter mixing matrix.
