### 2021-07-12 c446156
Coerce optimal signs to integers.
