### 2021-07-12 c1bf640
Drop sign-flipping from method descriptions.
