### 2021-07-12 c393ff8
Write out PCA metrics TSV with index.
