### 2021-07-12 7e7e92e
Note issue with column reordering in metric calculation.
It's not a big deal, but devs might wonder why the columns in the component table don't match the order hardcoded into generate_metrics.
