### 2021-07-12 6a98fe8
Fix stream formatter.
