+++ tedana/decay.py
@@ -8,7 +8,7 @@ from scipy import stats
 
 from tedana import utils
 
-LGR = logging.getLogger(__name__)
+LGR = logging.getLogger("GENERAL")
 RepLGR = logging.getLogger('REPORT')
 RefLGR = logging.getLogger('REFERENCES')
 
