+++ tedana/stats.py
@@ -9,7 +9,7 @@ from scipy import stats
 from tedana import utils
 from tedana.due import due, BibTeX, Doi
 
-LGR = logging.getLogger(__name__)
+LGR = logging.getLogger("GENERAL")
 RepLGR = logging.getLogger('REPORT')
 RefLGR = logging.getLogger('REFERENCES')
 
