### 2021-07-12 e161ffd
Apply to t2smap too.
