### 2021-07-12 5266b3f
Remove ContextFilter.
It's unnecessary now that we're using a single LGR for general logging.
