### 2021-07-12 bebeb13
Remove check.
The issue is that pytest's handler is different from our own.
