### 2021-07-12 8317ff1
Use denoising mask for component figures.
