### 2021-07-13 7925a61
Swap full and limited T2*/S0 maps.
Now that the full maps are used throughout the pipeline, they are the primary T2*/S0 outputs. The limited maps are not used for anything, and are only written out if the workflow is run in verbose mode.
