### 2021-07-13 def354c
Apply the same logic to t2smap.
