### 2021-07-13 6b5f314
Add template and function to only show GSR and MIR if they were calculated.
