### 2021-07-13 953abeb
Fix documentation about full vs. limited maps.
