+++ docs/outputs.rst
@@ -14,15 +14,15 @@ T2starmap.nii.gz                                    Full estimated T2* 3D map.
                                                     Values are in seconds.
                                                     The difference between the limited and full maps
                                                     is that, for voxels affected by dropout where
-                                                    only one echo contains good data, the full map
-                                                    uses the single echo's value while the limited
-                                                    map has a NaN.
+                                                    only one echo contains good data, the full map uses
+                                                    the T2* estimate from the first two echoes, while the
+                                                    limited map has a NaN.
 S0map.nii.gz                                        Full S0 3D map.
                                                     The difference between the limited and full maps
                                                     is that, for voxels affected by dropout where
-                                                    only one echo contains good data, the full map
-                                                    uses the single echo's value while the limited
-                                                    map has a NaN.
+                                                    only one echo contains good data, the full map uses
+                                                    the S0 estimate from the first two echoes, while the
+                                                    limited map has a NaN.
 desc-optcom_bold.nii.gz                             Optimally combined time series.
 desc-optcomDenoised_bold.nii.gz                     Denoised optimally combined time series. Recommended
                                                     dataset for analysis.
@@ -78,11 +78,17 @@ Filename                                                        Content
 ==============================================================  =====================================================
 desc-limited_T2starmap.nii.gz                                   Limited T2* map/time series.
                                                                 Values are in seconds.
-                                                                The difference between the limited and full maps is
-                                                                that, for voxels affected by dropout where only one
-                                                                echo contains good data, the full map uses the
-                                                                single echo's value while the limited map has a NaN.
+                                                                The difference between the limited and full maps
+                                                                is that, for voxels affected by dropout where
+                                                                only one echo contains good data, the full map uses
+                                                                the S0 estimate from the first two echoes, while the
+                                                                limited map has a NaN.
 desc-limited_S0map.nii.gz                                       Limited S0 map/time series.
+                                                                The difference between the limited and full maps
+                                                                is that, for voxels affected by dropout where
+                                                                only one echo contains good data, the full map uses
+                                                                the S0 estimate from the first two echoes, while the
+                                                                limited map has a NaN.
 echo-[echo]_desc-[PCA|ICA]_components.nii.gz                    Echo-wise PCA/ICA component weight maps.
 echo-[echo]_desc-[PCA|ICA]R2ModelPredictions_components.nii.gz  Component- and voxel-wise R2-model predictions,
                                                                 separated by echo.
