### 2021-07-14 319bbba
[REF] Do not modify mixing matrix with sign-flipping (#749)
* Add "optimal sign" metric and do not alter mixing matrix.

* Coerce optimal signs to integers.

* Drop sign-flipping from method descriptions.

* Adjust figure signs.

* Write out PCA metrics TSV with index.

* Note issue with column reordering in metric calculation.

It's not a big deal, but devs might wonder why the columns in the component table don't match the order hardcoded into generate_metrics.

* Use save_file instead of to_csv.
