+++ tedana/tests/test_metrics.py
@@ -48,7 +48,7 @@ def test_smoke_generate_metrics(testdata1):
         "normalized variance explained",
         "d_table_score",
     ]
-    comptable, mixing = collect.generate_metrics(
+    comptable = collect.generate_metrics(
         testdata1["data_cat"],
         testdata1["data_optcom"],
         testdata1["mixing"],
