### 2021-07-15 aded008
Changed where system info is saved and how it is displayed on report
