+++ tedana/workflows/tedana.py
@@ -1,6 +1,7 @@
 """
 Run the "canonical" TE-Dependent ANAlysis workflow.
 """
+import inspect
 import json
 import os
 import sys
@@ -340,6 +341,9 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
     generated by this workflow, please visit
     https://tedana.readthedocs.io/en/latest/outputs.html
     """
+    frame = inspect.currentframe()
+    _, _, _, arg_values = inspect.getargvalues(frame)
+
     out_dir = op.abspath(out_dir)
     if not op.isdir(out_dir):
         os.mkdir(out_dir)
@@ -424,9 +428,7 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
     # Save system info to json
     info_dict = utils.get_system_info()
     info_dict["Python"] = sys.version
-    info_dict["Tedana"] = __version__
     info_dict["Command"] = " ".join(sys.argv)
-    io_generator.save_file(info_dict, "System info json")
 
     n_samp, n_echos, n_vols = catd.shape
     LGR.debug('Resulting data shape: {}'.format(catd.shape))
@@ -700,6 +702,9 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
     if verbose:
         io.writeresults_echoes(catd, mmix, mask, comptable, io_generator)
 
+    # Generate tedana command with all API values
+    tedana_command = utils.write_tedana_command(arg_values)
+
     # Write out BIDS-compatible description file
     derivative_metadata = {
         "Name": "tedana Outputs",
@@ -713,7 +718,17 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
                     "A denoising pipeline for the identification and removal "
                     "of non-BOLD noise from multi-echo fMRI data."
                 ),
-                "CodeURL": "https://github.com/ME-ICA/tedana"
+                "CodeURL": "https://github.com/ME-ICA/tedana",
+                "Node": {
+                    "Name": info_dict["Node"],
+                    "System": info_dict["System"],
+                    "Machine": info_dict["Machine"],
+                    "Processor": info_dict["Processor"],
+                    "Release": info_dict["Release"],
+                    "Version": info_dict["Version"]
+                },
+                "Python": info_dict["Python"],
+                "Command": tedana_command
             }
         ]
     }
