### 2021-07-15 465c81e
Only change button names when files are generated
