### 2021-07-15 9f8467c
Update tedana/utils.py
Co-authored-by: Taylor Salo <tsalo006@fiu.edu>
