### 2021-07-15 2d6d91a
[DOC] Reorganize documentation (#740)
* Start reorganizing docs.

* Fix role in tedpca docstring.

* Fix docstring.

* Finish reorganization and heading standardization.

Also add note about heading level formatting to CONTRIBUTING.md.

* Fix levels in roadmap.

* Fix more heading levels.

* Move new changes from resources to multi-echo.

* Delete resources.rst.
