### 2021-07-15 6472ad0
Add carpet plots figure and section in docs.
