### 2021-07-16 d2c9e84
[FIX] Correct manacc documentation and check for associated inputs (#754)
