### 2021-07-16 95db93d
Log updated masking procedure.
