+++ pyproject.toml
@@ -1,6 +1,3 @@
-[build-system]
-requires = ["setuptools==40.8", "wheel"]
-
 [tool.black]
 line-length = 99
 target-version = ['py37']
