### 2021-07-17 12f16eb
[ENH] Add carpet plot to outputs (#696)
* Create new denoising function.

* Add carpet plots.

* Add docstring.

* Update test outputs.

* Drop output check from test.

Now that write_split_ts doesn't return anything.... we shouldn't check that it does.

* Add carpet plots to report

* Re-drop varexpl.

* Fix style issue.

* Don't unmask outputs of denoise_ts.

* Use io_generator for carpet plots.

* Split carpet plots into separate figures.

* Try to organize the report?

* Fix report generation.

* Added buttons to select carpet plot and added some structure to the reports

* Revert "Added b
[…]
