+++ tedana/selection/tedica.py
@@ -2,16 +2,17 @@
 Functions to identify TE-dependent and TE-independent components.
 """
 import logging
+
 import numpy as np
 from scipy import stats
 
-from tedana.stats import getfbounds
-from tedana.selection._utils import getelbow, clean_dataframe
 from tedana.metrics import collect
+from tedana.selection._utils import clean_dataframe, getelbow
+from tedana.stats import getfbounds
 
 LGR = logging.getLogger(__name__)
-RepLGR = logging.getLogger('REPORT')
-RefLGR = logging.getLogger('REFERENCES')
+RepLGR = logging.getLogger("REPORT")
+RefLGR = logging.getLogger("REFERENCES")
 
 
 def manual_selection(comptable, acc=None, rej=None):
@@ -35,17 +36,21 @@ def manual_selection(comptable, acc=None, rej=None):
         Dictionary with metadata about calculated metrics.
         Each entry corresponds to a column in ``comptable``.
     """
-    LGR.info('Performing manual ICA component selection')
-    RepLGR.info("Next, components were manually classified as "
-                "BOLD (TE-dependent), non-BOLD (TE-independent), or "
-                "uncertain (low-variance).")
-    if ('classification' in comptable.columns and
-            'original_classification' not in comptable.columns):
-        comptable['original_classification'] = comptable['classification']
-        comptable['original_rationale'] = comptable['rationale']
+    LGR.info("Performing manual ICA component selection")
+    RepLGR.info(
+        "Next, components were manually classified as "
+        "BOLD (TE-dependent), non-BOLD (TE-independent), or "
+        "uncertain (low-variance)."
+    )
+    if (
+        "classification" in comptable.columns
+        and "original_classification" not in comptable.columns
+    ):
+        comptable["original_classification"] = comptable["classification"]
+        comptable["original_rationale"] = comptable["rationale"]
 
-    comptable['classification'] = 'accepted'
-    comptable['rationale'] = ''
+    comptable["classification"] = "accepted"
+    comptable["rationale"] = ""
 
     all_comps = comptable.index.values
     if acc is not None:
@@ -59,18 +64,19 @@ def manual_selection(comptable, acc=None, rej=None):
     elif acc is None and rej is not None:
         acc = sorted(np.setdiff1d(all_comps, rej))
     elif acc is None and rej is None:
-        LGR.info('No manually accepted or rejected components supplied. '
-                 'Accepting all components.')
+        LGR.info(
+            "No manually accepted or rejected components supplied. " "Accepting all components."
+        )
         # Accept all components if no manual selection provided
         acc = all_comps[:]
         rej = []
 
     ign = np.setdiff1d(all_comps, np.union1d(acc, rej))
-    comptable.loc[acc, 'classification'] = 'accepted'
-    comptable.loc[rej, 'classification'] = 'rejected'
-    comptable.loc[rej, 'rationale'] += 'I001;'
-    comptable.loc[ign, 'classification'] = 'ignored'
-    comptable.loc[ign, 'rationale'] += 'I001;'
+    comptable.loc[acc, "classification"] = "accepted"
+    comptable.loc[rej, "classification"] = "rejected"
+    comptable.loc[rej, "rationale"] += "I001;"
+    comptable.loc[ign, "classification"] = "ignored"
+    comptable.loc[ign, "rationale"] += "I001;"
 
     # Move decision columns to end
     comptable = clean_dataframe(comptable)
[…]
