+++ tedana/reporting/static_figures.py
@@ -4,19 +4,20 @@ Functions to creating figures to inspect tedana output
 import logging
 import os
 
-import numpy as np
 import matplotlib
-matplotlib.use('AGG')
+import numpy as np
+
+matplotlib.use("AGG")
 import matplotlib.pyplot as plt
 from nilearn import plotting
 
 from tedana import io, stats, utils
 
 LGR = logging.getLogger(__name__)
-MPL_LGR = logging.getLogger('matplotlib')
+MPL_LGR = logging.getLogger("matplotlib")
 MPL_LGR.setLevel(logging.WARNING)
-RepLGR = logging.getLogger('REPORT')
-RefLGR = logging.getLogger('REFERENCES')
+RepLGR = logging.getLogger("REPORT")
+RefLGR = logging.getLogger("REFERENCES")
 
 
 def _trim_edge_zeros(arr):
@@ -37,9 +38,7 @@ def _trim_edge_zeros(arr):
     """
 
     mask = arr != 0
-    bounding_box = tuple(
-                         slice(np.min(indexes), np.max(indexes) + 1)
-                         for indexes in np.where(mask))
+    bounding_box = tuple(slice(np.min(indexes), np.max(indexes) + 1) for indexes in np.where(mask))
     return arr[bounding_box]
 
 
@@ -191,32 +190,30 @@ def comp_figures(ts, mask, comptable, mmix, io_generator, png_cmap):
 
     # Create indices for 6 cuts, based on dimensions
     cuts = [ts_B.shape[dim] // 6 for dim in range(3)]
-    expl_text = ''
+    expl_text = ""
 
     # Remove trailing ';' from rationale column
-    comptable['rationale'] = comptable['rationale'].str.rstrip(';')
+    comptable["rationale"] = comptable["rationale"].str.rstrip(";")
     for compnum in comptable.index.values:
-        if comptable.loc[compnum, "classification"] == 'accepted':
-            line_color = 'g'
-            expl_text = 'accepted'
-        elif comptable.loc[compnum, "classification"] == 'rejected':
-            line_color = 'r'
-            expl_text = 'rejection reason(s): ' + comptable.loc[compnum, "rationale"]
-        elif comptable.loc[compnum, "classification"] == 'ignored':
-            line_color = 'k'
-            expl_text = 'ignored reason(s): ' + comptable.loc[compnum, "rationale"]
+        if comptable.loc[compnum, "classification"] == "accepted":
+            line_color = "g"
+            expl_text = "accepted"
+        elif comptable.loc[compnum, "classification"] == "rejected":
+            line_color = "r"
+            expl_text = "rejection reason(s): " + comptable.loc[compnum, "rationale"]
+        elif comptable.loc[compnum, "classification"] == "ignored":
+            line_color = "k"
+            expl_text = "ignored reason(s): " + comptable.loc[compnum, "rationale"]
         else:
             # Classification not added
             # If new, this will keep code running
-            line_color = '0.75'
-            expl_text = 'other classification'
+            line_color = "0.75"
+            expl_text = "other classification"
 
         allplot = plt.figure(figsize=(10, 9))
-        ax_ts = plt.subplot2grid((5, 6), (0, 0),
-                                 rowspan=1, colspan=6,
-                                 fig=allplot)
+        ax_ts = plt.subplot2grid((5, 6), (0, 0), rowspan=1, colspan=6, fig=allplot)
 
-        ax_ts.set_xlabel('TRs')
+        ax_ts.set_xlabel("TRs")
[…]
