+++ tedana/decomposition/__init__.py
@@ -1,7 +1,7 @@
 # emacs: -*- mode: python-mode; py-indent-offset: 4; tab-width: 4; indent-tabs-mode: nil -*-
 # ex: set sts=4 ts=4 sw=4 et:
 
-from .pca import tedpca
 from .ica import tedica
+from .pca import tedpca
 
-__all__ = ['tedpca', 'tedica']
+__all__ = ["tedpca", "tedica"]
