+++ tedana/due.py
@@ -16,8 +16,8 @@ Copyright:  2015-2016  DueCredit developers
 License:    BSD-2
 """
 
-from builtins import str
-from builtins import object
+from builtins import object, str
+
 __version__ = '0.0.5'
 
 
@@ -45,7 +45,7 @@ def _donothing_func(*args, **kwargs):
 
 
 try:
-    from duecredit import due, BibTeX, Doi, Url
+    from duecredit import BibTeX, Doi, Url, due
     if 'due' in locals() and not hasattr(due, 'cite'):
         raise RuntimeError(
             "Imported due lacks .cite. DueCredit is now disabled")
