+++ tedana/reporting/html_report.py
@@ -6,6 +6,7 @@ from string import Template
 import pandas as pd
 from bokeh import __version__ as bokehversion
 from bokeh import embed, layouts, models
+
 from tedana.info import __version__
 from tedana.reporting import dynamic_figures as df
 
@@ -37,12 +38,14 @@ def _generate_buttons(out_dir):
     with open(str(buttons_template_path), "r") as buttons_file:
         buttons_tpl = Template(buttons_file.read())
 
-    buttons_html = buttons_tpl.substitute(optcomdisp=optcom_nogsr_disp,
-                                          denoiseddisp=denoised_mir_disp,
-                                          accepteddisp=accepted_mir_disp,
-                                          optcomname=optcom_name,
-                                          denoisedname=denoised_name,
-                                          acceptedname=accepted_name)
+    buttons_html = buttons_tpl.substitute(
+        optcomdisp=optcom_nogsr_disp,
+        denoiseddisp=denoised_mir_disp,
+        accepteddisp=accepted_mir_disp,
+        optcomname=optcom_name,
+        denoisedname=denoised_name,
+        acceptedname=accepted_name,
+    )
 
     return buttons_html
 
@@ -88,9 +91,7 @@ def _save_as_html(body):
     with open(str(head_template_path), "r") as head_file:
         head_tpl = Template(head_file.read())
 
-    html = head_tpl.substitute(
-        version=__version__, bokehversion=bokehversion, body=body
-    )
+    html = head_tpl.substitute(version=__version__, bokehversion=bokehversion, body=body)
     return html
 
 

[…]
