### 2021-07-17 eb363b6
Add an informative error to getelbow functions.
