### 2021-07-17 84d868e
Expand elbow calculations.
Makes for more informative tracebacks.
