### 2021-07-17 cbaa57d
Fix combined multiline strings.
It's the one thing black always messes up.
