+++ tedana/reporting/static_figures.py
@@ -240,7 +240,7 @@ def comp_figures(ts, mask, comptable, mmix, io_generator, png_cmap):
         comp_var = "{0:.2f}".format(comptable.loc[compnum, "variance explained"])
         comp_kappa = "{0:.2f}".format(comptable.loc[compnum, "kappa"])
         comp_rho = "{0:.2f}".format(comptable.loc[compnum, "rho"])
-        plt_title = "Comp. {}: variance: {}%, kappa: {}, rho: {}, " "{}".format(
+        plt_title = "Comp. {}: variance: {}%, kappa: {}, rho: {}, {}".format(
             compnum, comp_var, comp_kappa, comp_rho, expl_text
         )
         title = ax_ts.set_title(plt_title)
