+++ tedana/io.py
@@ -571,7 +571,7 @@ def load_data(data, n_echos=None):
         if len(data) == 1:  # a z-concatenated file was provided
             data = data[0]
         elif len(data) == 2:  # inviable -- need more than 2 echos
-            raise ValueError("Cannot run `tedana` with only two echos: " "{}".format(data))
+            raise ValueError("Cannot run `tedana` with only two echos: {}".format(data))
         else:  # individual echo files were provided (surface or volumetric)
             fdata = np.stack([utils.load_image(f) for f in data], axis=1)
             ref_img = check_niimg(data[0])

[…]
