+++ tedana/due.py
@@ -18,25 +18,28 @@ License:    BSD-2
 
 from builtins import object, str
 
-__version__ = '0.0.5'
+__version__ = "0.0.5"
 
 
 class InactiveDueCreditCollector(object):
     """Just a stub at the Collector which would not do anything"""
+
     def _donothing(self, *args, **kwargs):
         """Perform no good and no bad"""
         pass
 
     def dcite(self, *args, **kwargs):
         """If I could cite I would"""
+
         def nondecorating_decorator(func):
             return func
+
         return nondecorating_decorator
 
     cite = load = add = _donothing
 
     def __repr__(self):
-        return self.__class__.__name__ + '()'
+        return self.__class__.__name__ + "()"
 
 
 def _donothing_func(*args, **kwargs):
@@ -46,15 +49,18 @@ def _donothing_func(*args, **kwargs):
 
 try:
     from duecredit import BibTeX, Doi, Url, due
-    if 'due' in locals() and not hasattr(due, 'cite'):
-        raise RuntimeError(
-            "Imported due lacks .cite. DueCredit is now disabled")
+
+    if "due" in locals() and not hasattr(due, "cite"):
+        raise RuntimeError("Imported due lacks .cite. DueCredit is now disabled")
 except Exception as e:
-    if type(e).__name__ != 'ImportError':
+    if type(e).__name__ != "ImportError":
         import logging
+
         logging.getLogger("duecredit").error(
             'Module `duecredit` not successfully imported due to "%s". '
-            'Package functionality unaffected.', str(e))
+            "Package functionality unaffected.",
+            str(e),
+        )
 
     # Initiate due stub
     due = InactiveDueCreditCollector()
