+++ tedana/info.py
@@ -7,62 +7,63 @@ Base module variables
 
 from ._version import get_versions
 
-__version__ = get_versions()['version']
+__version__ = get_versions()["version"]
 del get_versions
 
-__author__ = 'tedana developers'
-__copyright__ = 'Copyright 2018, tedana developers'
-__license__ = 'LGPL 2.1'
-__maintainer__ = 'Elizabeth DuPre'
-__email__ = 'emd222@cornell.edu'
-__status__ = 'Prototype'
-__url__ = 'https://github.com/me-ica/tedana'
-__packagename__ = 'tedana'
-__description__ = ('TE-Dependent Analysis (tedana) of multi-echo functional '
-                   'magnetic resonance imaging (fMRI) data.')
+__author__ = "tedana developers"
+__copyright__ = "Copyright 2018, tedana developers"
+__license__ = "LGPL 2.1"
+__maintainer__ = "Elizabeth DuPre"
+__email__ = "emd222@cornell.edu"
+__status__ = "Prototype"
+__url__ = "https://github.com/me-ica/tedana"
+__packagename__ = "tedana"
+__description__ = (
+    "TE-Dependent Analysis (tedana) of multi-echo functional "
+    "magnetic resonance imaging (fMRI) data."
+)
 
-DOWNLOAD_URL = (
-    'https://github.com/ME-ICA/{name}/archive/{ver}.tar.gz'.format(
-        name=__packagename__, ver=__version__))
+DOWNLOAD_URL = "https://github.com/ME-ICA/{name}/archive/{ver}.tar.gz".format(
+    name=__packagename__, ver=__version__
+)
 
 REQUIRES = [
-    'bokeh',
-    'mapca~=0.0.1',
-    'matplotlib',
-    'nibabel>=2.5.1',
-    'nilearn>=0.5.2',
-    'numpy>=1.15',
-    'pandas',
-    'scikit-learn>=0.19.1',
-    'scipy>=1.1.0',
-    'threadpoolctl'
+    "bokeh",
+    "mapca~=0.0.1",
+    "matplotlib",
+    "nibabel>=2.5.1",
+    "nilearn>=0.5.2",
+    "numpy>=1.15",
+    "pandas",
+    "scikit-learn>=0.19.1",
+    "scipy>=1.1.0",
+    "threadpoolctl",
 ]
 
 TESTS_REQUIRES = [
-    'codecov',
-    'coverage<5.0',
-    'flake8>=3.7',
+    "codecov",
+    "coverage<5.0",
+    "flake8>=3.7",
     "flake8-black",
     "flake8-isort",
-    'pytest',
-    'pytest-cov',
-    'requests',
+    "pytest",
+    "pytest-cov",
+    "requests",
 ]
 
 EXTRA_REQUIRES = {
-    'dev': ['versioneer'],
-    'doc': [
-        'sphinx>=1.5.3',
-        'sphinx_rtd_theme',
-        'sphinx-argparse',
+    "dev": ["versioneer"],
+    "doc": [
+        "sphinx>=1.5.3",
+        "sphinx_rtd_theme",
+        "sphinx-argparse",
     ],
-    'tests': TESTS_REQUIRES,
-    'duecredit': ['duecredit'],
+    "tests": TESTS_REQUIRES,
+    "duecredit": ["duecredit"],
 }
 
 # Enable a handle to install all extra dependencies at once
-EXTRA_REQUIRES['all'] = list(set([
-    v for deps in EXTRA_REQUIRES.values() for v in deps]))
+EXTRA_REQUIRES["all"] = list(set([v for deps in EXTRA_REQUIRES.values() for v in deps]))
 
[…]
