+++ tedana/selection/tedica.py
@@ -245,11 +245,20 @@ def kundu_selection_v2(comptable, n_echos, n_vols):
     # Compute elbows from other elbows
     f05, _, f01 = getfbounds(n_echos)
     kappas_nonsig = comptable.loc[comptable['kappa'] < f01, 'kappa']
+    if not kappas_nonsig.size:
+        LGR.warning(
+            "No nonsignificant kappa values detected. "
+            "Only calculating elbow from all values instead."
+        )
+        kappas_nonsig_elbow = np.nan
+    else:
+        kappas_nonsig_elbow = getelbow(kappas_nonsig, return_val=True)
+
+    kappas_all_elbow = getelbow(comptable['kappa'], return_val=True)
+
     # NOTE: Would an elbow from all Kappa values *ever* be lower than one from
     # a subset of lower (i.e., nonsignificant) values?
-    kappas_nonsig_elbow = getelbow(kappas_nonsig, return_val=True)
-    kappas_all_elbow = getelbow(comptable['kappa'], return_val=True)
-    kappa_elbow = np.min((kappas_all_elbow, kappas_nonsig_elbow))
+    kappa_elbow = np.nanmin((kappas_all_elbow, kappas_nonsig_elbow))
     rhos_ncls_elbow = getelbow(comptable.loc[ncls, 'rho'], return_val=True)
     rhos_all_elbow = getelbow(comptable['rho'], return_val=True)
     rho_elbow = np.mean((rhos_ncls_elbow, rhos_all_elbow, f05))
