### 2021-07-17 ac9e35c
Replace single-quote strings.
