### 2021-07-19 983a48c
[ENH] Coerces images to 32-bit (#759)
