### 2021-07-20 cc7fc8f
[FIX] Ignore non-significant kappa elbow when no non-significant kappa values exist (#760)
* Add an informative error to getelbow functions.

* Expand elbow calculations.

Makes for more informative tracebacks.

* Compromise with warning.

* Improve warning text.

* Replace random seed recommendation.

* Add tests.

* Update error messages.
