### 2021-07-22 d5cbf86
[REF] Move logger management to new functions (#750)
* First attempt.

* Hm.

* Fix stream formatter.

* Replace file-specific loggers with a general one.

* Try fixing it.

* Apply to t2smap too.

* Remove ContextFilter.

It's unnecessary now that we're using a single LGR for general logging.

* Add check in CI.

* Remove check.

The issue is that pytest's handler is different from our own.
