+++ tedana/selection/tedica.py
@@ -9,7 +9,7 @@ from tedana.stats import getfbounds
 from tedana.selection._utils import getelbow, clean_dataframe
 from tedana.metrics import collect
 
-LGR = logging.getLogger(__name__)
+LGR = logging.getLogger("GENERAL")
 RepLGR = logging.getLogger('REPORT')
 RefLGR = logging.getLogger('REFERENCES')
 
