### 2021-07-22 7e1cf2c
Updated dependency version numbering (#763)
