### 2021-07-25 c4a7918
Overhaul the walkthrough.
