+++ docs/multi-echo_walkthrough.rst
@@ -24,51 +24,183 @@ The physics of fMRI
 In a typical fMRI sequence,
 the protons in a brain slice are aligned perpendicularly to the main magnetic
 field of the scanner (:math:`B_0`) with an "excitation pulse".
-Those protons start releasing energy from the excitation pulse as they fall
-back in line with :math:`B_0`,
+Those protons start releasing energy from the excitation pulse as they fall back in line with :math:`B_0`,
 and that energy is actively measured at an "echo time" (also known as `TE`_).
 
 The time it takes for 37% of the protons fall back in line with :math:`B_0`
 is known as the transverse relaxation time, or "T2".
 The exact value of T2 varies across voxels in the brain, based on tissue type,
-which is why protocols which are T2-weighted or which quantitatively measure
-T2 are useful for structural analyses.
+which is why protocols which are T2-weighted or which quantitatively measure T2 are useful for structural analyses.
 
-However,
+fMRI works based on the fact that differences in blood oxygenation
+(indicative of gross differences in neural metabolism ostensibly driven by neural activity)
+also impact observed T2, also known as T2*.
+Thus, changes in the measured signal from an fMRI sequence reflect (at least in part) changes in blood oxygenation.
 
-To start, let's look at the most common type of fMRI, single-echo EPI.
+
+*****************************
+BOLD signal and BOLD contrast
+*****************************
+
+As mentioned above, energy is released after the excitation pulse, and this is our observed fMRI signal.
+The echo time is the point at which that signal is recorded.
+
+Let's take a look at how fMRI signal varies as a function of echo time.
+
+.. image:: /_static/physics_signal_decay.png
+
+As you can see, signal decays as echo time increases.
+However, one very important feature of this signal decay is that the signal decays differently
+depending on the level of blood oxygenation in the voxel.
+
+When a voxel contains more deoxygenated blood,
+its signal decays more slowly than when the blood within it is more oxygenated.
+
+.. image:: /_static/physics_signal_decay_activity.png
+
+This is the "BOLD contrast" functional neuroimagers care about.
+Namely, you can compare the signal from a voxel during a cognitive task
+(when you expect that voxel's brain region to be more "active")
+against that voxel's signal from a different cognitive task in order to determine if there is a
+meaningful different in activation between the two tasks.
+
+There are two relevant features to the BOLD contrast.
+First, as repeatedly noted above, overall signal decays as echo time increases.
+Eventually, with a long enough echo time, there is not enough signal to detect over noise
+(i.e., signal-to-noise ratio is too low)
+and meaningful signal cannot be observed.
+Second, BOLD contrast increases as echo time increases.
+That is, the relative difference in signal between active and inactive states increases with echo time,
+even as the overall signal decreases.
+
+While the signal decay curve can be described using many models,
+one of the most useful approximations (and the one we used to simulate the signals in this walkthrough)
+is a monoexponential decay curve.
+In this model, the signal is driven by two factors:
+the "intercept" (signal at echo time = 0s), also known as S0; and the "slope" (decay rate), also known as T2*.
+For more information about different models of fMRI signal decay, please see XXXXX.
+
+.. note:: T2*
+    We said above that observed T2 is T2*, and now we're saying that the slope of the decay is T2*.
+    That's because they're the same thing!
+
+In the above figure, the difference between the "inactive" and "active" signals was driven by a change in T2* from 20ms to 40ms.
+Thus, the point at which the difference in signal is maximized between the two states in this voxel is ~30ms.
+
+And, in fact, this is the standard approach for the most common version of fMRI: single-echo fMRI.
 
 
 *************************
 What is single-echo fMRI?
[…]
