### 2021-07-25 13b5cd1
Fix misnamed figure.
