### 2021-07-29 3ede972
Revert T2* changes.
