+++ tedana/workflows/tedana.py
@@ -503,13 +503,13 @@ def tedana_workflow(data, tes, out_dir='.', mask=None,
 
         # set a hard cap for the T2* map
         # anything that is 10x higher than the 99.5 %ile will be reset to 99.5 %ile
-        cap_t2s = stats.scoreatpercentile(t2s_full.flatten(), 99.5,
+        cap_t2s = stats.scoreatpercentile(t2s_limited.flatten(), 99.5,
                                           interpolation_method='lower')
         LGR.debug('Setting cap on T2* map at {:.5f}s'.format(
             utils.millisec2sec(cap_t2s)))
-        t2s_full[t2s_full > cap_t2s * 10] = cap_t2s
-        io_generator.save_file(utils.millisec2sec(t2s_full), 't2star img')
-        io_generator.save_file(s0_full, 's0 img')
+        t2s_limited[t2s_limited > cap_t2s * 10] = cap_t2s
+        io_generator.save_file(utils.millisec2sec(t2s_limited), 't2star img')
+        io_generator.save_file(s0_limited, 's0 img')
 
         if verbose:
             io_generator.save_file(utils.millisec2sec(t2s_limited), 'limited t2star img')
