+++ docs/multi-echo_walkthrough.rst
@@ -48,7 +48,7 @@ The echo time is the point at which that signal is recorded.
 Let's take a look at how fMRI signal varies as a function of echo time.
 
 .. image:: https://osf.io/m7aw3/
-   alt: physics_signal_decay.png
+   :alt: physics_signal_decay.png
 
 As you can see, signal decays as echo time increases.
 However, one very important feature of this signal decay is that the signal decays differently
@@ -57,7 +57,8 @@ depending on the level of blood oxygenation in the voxel.
 When a voxel contains more deoxygenated blood,
 its signal decays more slowly than when the blood within it is more oxygenated.
 
-.. image:: /_static/physics_signal_decay_activity.png
+.. image:: https://osf.io/ve7cf/
+   :alt: physics_signal_decay_activity.png
 
 This is the "BOLD contrast" functional neuroimagers care about.
 Namely, you can compare the signal from a voxel during a cognitive task
@@ -144,7 +145,8 @@ There are, in fact, many factors that impact observed fMRI signal, but we will f
 
 Let's take a look at what single-echo data looks like over time.
 
-.. image:: /_static/fluctuations_single-echo.gif
+.. image:: https://osf.io/g9dqc/
+   :alt: fluctuations_single-echo.gif
 
 As you can see, the single data point fluctuates over time.
 Let's assume that those fluctuations reflect meaningful BOLD signal.
@@ -152,14 +154,16 @@ Nothing to be concerned about, right?
 
 Okay, let's check out the underlying signal decay curve we're sampling from.
 
-.. image:: /_static/fluctuations_single-echo_with_curve.gif
+.. image:: https://osf.io/5yjwx/
+   :alt: fluctuations_single-echo_with_curve.gif
 
 Everything still looks fine, right?
 We know there's an underlying signal decay curve, and we're sampling that curve at a single point, at our TE.
 
 What if we describe the curve in terms of S0 and T2*?
 
-.. image:: /_static/fluctuations_single-echo_with_curve_and_t2s_s0.gif
+.. image:: https://osf.io/6a7nv/
+   :alt: fluctuations_single-echo_with_curve_and_t2s_s0.gif
 
 Now we see that the changes in the signal are driven by changes in _both_ S0 and T2*.
 Why should we care about that?
@@ -179,7 +183,8 @@ To make sure we can _really_ see the curves, we'll also make the S0 and T2* chan
 They have different scales, so we'll use the same time series of fluctuations,
 scaled to have matching percent signal changes between the two values.
 
-.. image:: /_static/fluctuations_t2s_s0.gif
+.. image:: https://osf.io/g29ez/
+   :alt: fluctuations_t2s_s0.gif
 
 Hey, look at that!
 The curves change differently!
@@ -187,7 +192,8 @@ If you look at the whole curve, you can differentiate S0 changes from T2* change
 
 Now that we know that, what about single-echo fMRI?
 
-.. image:: /_static/fluctuations_t2s_s0_single-echo.gif
+.. image:: https://osf.io/mx4ku/
+   :alt: fluctuations_t2s_s0_single-echo.gif
 
 Hm... with only one data point per time point, we really can't tell whether the changes are due to S0 or T2*.
 
@@ -205,7 +211,8 @@ though more are possible if you make certain compromises with your parameters.
 
 Here we have some simulated data with six echoes.
 
-.. image:: /_static/fluctuations_t2s_s0_multi-echo.gif
+.. image:: https://osf.io/mf3ae/
+   :alt: fluctuations_t2s_s0_multi-echo.gif
 
 Now we can tell the two curves apart again!
 
@@ -237,9 +244,11 @@ By using the information across multiple echos these images can be combined in
[…]
