### 2021-07-31 49f9f46
Started working on React-based reports
