+++ tedana/reporting/html_report.py
@@ -37,12 +37,14 @@ def _generate_buttons(out_dir):
     with open(str(buttons_template_path), "r") as buttons_file:
         buttons_tpl = Template(buttons_file.read())
 
-    buttons_html = buttons_tpl.substitute(optcomdisp=optcom_nogsr_disp,
-                                          denoiseddisp=denoised_mir_disp,
-                                          accepteddisp=accepted_mir_disp,
-                                          optcomname=optcom_name,
-                                          denoisedname=denoised_name,
-                                          acceptedname=accepted_name)
+    buttons_html = buttons_tpl.substitute(
+        optcomdisp=optcom_nogsr_disp,
+        denoiseddisp=denoised_mir_disp,
+        accepteddisp=accepted_mir_disp,
+        optcomname=optcom_name,
+        denoisedname=denoised_name,
+        acceptedname=accepted_name,
+    )
 
     return buttons_html
 
@@ -69,11 +71,13 @@ def _update_template_bokeh(bokeh_id, about, bokeh_js, buttons):
     body_template_path = resource_path.joinpath(body_template_name)
     with open(str(body_template_path), "r") as body_file:
         body_tpl = Template(body_file.read())
-    body = body_tpl.substitute(content=bokeh_id, about=about, javascript=bokeh_js, buttons=buttons)
+    body = body_tpl.substitute(
+        content=bokeh_id, about=about, javascript=bokeh_js, buttons=buttons
+    )
     return body
 
 
-def _save_as_html(body):
+def _save_as_html(path, body, mode):
     """
     Save an HTML report out to a file.
 
@@ -83,15 +87,35 @@ def _save_as_html(body):
         Body for HTML report with embedded figures
     """
     resource_path = Path(__file__).resolve().parent.joinpath("data", "html")
-    head_template_name = "report_head_template.html"
-    head_template_path = resource_path.joinpath(head_template_name)
-    with open(str(head_template_path), "r") as head_file:
+    # head_template_name = "report_head_template.html"
+    # head_template_path = resource_path.joinpath(head_template_name)
+    with open(str(path), "r") as head_file:
         head_tpl = Template(head_file.read())
 
-    html = head_tpl.substitute(
-        version=__version__, bokehversion=bokehversion, body=body
-    )
-    return html
+    if mode == "data":
+        html = head_tpl.substitute(componentsData=body)
+    elif mode == "about":
+        html = head_tpl.substitute(about=body)
+
+    with open(path, "wb") as f:
+        f.write(html.encode("utf-8"))
+
+
+def react_structure(df):
+
+    # {
+    #     x: "IC 1",
+    #     var: 30,
+    #     kappa: 40,
+    #     rho: 10,
+    #     status: "accepted",
+    #     color: acceptedColor,
+    # },
+    data = ""
+    for index, row in df.iterrows():
+        row_data = f'x: "Component {row["component"]}", var: {row["var_exp"]}, kappa: {row["kappa"]}, kappa_rank: {row["kappa_rank"]}, rho: {row["rho"]}, rho_rank: {row["rho_rank"]}, classification: "{row["classification"]}", color: "{row["color"]}",'
+        data += f"{{ {row_data} }},"
+    return data
 
 
 def generate_report(io_generator, tr):
@@ -118,53 +142,60 @@ def generate_report(io_generator, tr):
     comptable_path = io_generator.get_name("ICA metrics tsv")
     comptable_cds = df._create_data_struct(comptable_path)
 
-    # Create kappa rho plot
-    kappa_rho_plot = df._create_kr_plt(comptable_cds)
[…]
