+++ tedana/reporting/dynamic_figures.py
@@ -2,11 +2,9 @@ import numpy as np
 from math import pi
 import pandas as pd
 from sklearn.preprocessing import MinMaxScaler
-from bokeh import (events, models, plotting, transform)
+from bokeh import events, models, plotting, transform
 
-color_mapping = {'accepted': '#2ecc71',
-                 'rejected': '#e74c3c',
-                 'ignored': '#3498db'}
+color_mapping = {"accepted": "#2ecc71", "rejected": "#e74c3c", "ignored": "#3498db"}
 
 tap_callback_jscode = """
     // Accessing the selected component ID
@@ -68,58 +66,70 @@ def _create_data_struct(comptable_path, color_mapping=color_mapping):
     cds: bokeh.models.ColumnDataSource
         Data structure with all the fields to plot or hover over
     """
-    unused_cols = ['normalized variance explained',
-                   'countsigFT2', 'countsigFS0',
-                   'dice_FS0', 'countnoise', 'dice_FT2',
-                   'signal-noise_t', 'signal-noise_p',
-                   'd_table_score', 'kappa ratio',
-                   'rationale', 'd_table_score_scrub']
+    unused_cols = [
+        "normalized variance explained",
+        "countsigFT2",
+        "countsigFS0",
+        "dice_FS0",
+        "countnoise",
+        "dice_FT2",
+        "signal-noise_t",
+        "signal-noise_p",
+        "d_table_score",
+        "kappa ratio",
+        "rationale",
+        "d_table_score_scrub",
+    ]
 
     df = pd.read_table(comptable_path)
     n_comps = df.shape[0]
 
     # remove space from column name
-    df.rename(columns={'variance explained': 'var_exp'}, inplace=True)
+    df.rename(columns={"variance explained": "var_exp"}, inplace=True)
 
     # For providing sizes based on Var Explained that are visible
     mm_scaler = MinMaxScaler(feature_range=(4, 20))
-    df['var_exp_size'] = mm_scaler.fit_transform(
-        df[['var_exp', 'normalized variance explained']])[:, 0]
+    df["var_exp_size"] = mm_scaler.fit_transform(
+        df[["var_exp", "normalized variance explained"]]
+    )[:, 0]
 
     # Calculate Kappa and Rho ranks
-    df['rho_rank'] = df['rho'].rank(ascending=False).values
-    df['kappa_rank'] = df['kappa'].rank(ascending=False).values
-    df['var_exp_rank'] = df['var_exp'].rank(ascending=False).values
+    df["rho_rank"] = df["rho"].rank(ascending=False).values
+    df["kappa_rank"] = df["kappa"].rank(ascending=False).values
+    df["var_exp_rank"] = df["var_exp"].rank(ascending=False).values
 
     # Remove unused columns to decrease size of final HTML
     # set errors to 'ignore' in case some columns do not exist in
     # a given data frame
-    df.drop(unused_cols, axis=1, inplace=True, errors='ignore')
+    df.drop(unused_cols, axis=1, inplace=True, errors="ignore")
 
     # Create additional Column with colors based on final classification
-    df['color'] = [color_mapping[i] for i in df['classification']]
+    df["color"] = [color_mapping[i] for i in df["classification"]]
 
     # Create additional column with component ID
-    df['component'] = np.arange(n_comps)
+    df["component"] = np.arange(n_comps)
 
     # Compute angle and re-sort data for Pie plots
-    df['angle'] = df['var_exp'] / df['var_exp'].sum() * 2 * pi
-    df.sort_values(by=['classification', 'var_exp'], inplace=True)
+    df["angle"] = df["var_exp"] / df["var_exp"].sum() * 2 * pi
[…]
