### 2021-08-01 c9c0fde
First working version of React reports, replicating current Bokeh reports
