+++ tedana/reporting/data/src/Plotsgood.js
@@ -1,0 +1,296 @@
+import React, { useState } from "react";
+import { Line, Pie } from "react-chartjs-2";
+
+const acceptedColor = "#62bc6c";
+const rejedtecColor = "#f2563c";
+const ignoredColor = "#4e85f5";
+
+const data = [
+  {
+    x: "IC 1",
+    var: 30,
+    kappa: 40,
+    rho: 10,
+    status: "accepted",
+    color: acceptedColor,
+  },
+  {
+    x: "IC 2",
+    var: 25,
+    kappa: 30,
+    rho: 5,
+    status: "rejected",
+    color: rejedtecColor,
+  },
+  {
+    x: "IC 3",
+    var: 20,
+    kappa: 10,
+    rho: 2,
+    status: "accepted",
+    color: acceptedColor,
+  },
+  {
+    x: "IC 4",
+    var: 10,
+    kappa: 35,
+    rho: 7,
+    status: "ignored",
+    color: ignoredColor,
+  },
+  {
+    x: "IC 5",
+    var: 10,
+    kappa: 25,
+    rho: 8,
+    status: "accepted",
+    color: acceptedColor,
+  },
+  {
+    x: "IC 6",
+    var: 5,
+    kappa: 12,
+    rho: 2,
+    status: "rejected",
+    color: rejedtecColor,
+  },
+];
+
+let kappa_rho = {
+  labels: data.map((e) => e.x),
+  datasets: [
+    {
+      type: "scatter",
+      borderColor: "black",
+      pointBackgroundColor: data.map((e) => e.color),
+      pointBorderColor: data.map((e) => e.color),
+      pointRadius: 10,
+      borderWidth: 1,
+      fill: false,
+      data: data.map((e) => ({ x: e.rho, y: e.kappa })),
+    },
+  ],
+};
+
+let rho = {
+  labels: data.map((e) => e.x),
+  datasets: [
+    {
+      type: "line",
+      borderColor: "black",
+      pointBackgroundColor: data.map((e) => e.color),
+      pointBorderColor: data.map((e) => e.color),
+      pointRadius: 10,
+      borderWidth: 1,
+      fill: false,
+      data: data.map((e) => e.rho),
+    },
+  ],
+};
+
+let kappa = {
+  labels: data.map((e) => e.x),
+  datasets: [
+    {
+      type: "line",
+      borderColor: "black",
+      pointBackgroundColor: data.map((e) => e.color),
+      pointBorderColor: data.map((e) => e.color),
+      pointRadius: 10,
+      borderWidth: 1,
+      fill: false,
+      data: data.map((e) => e.kappa),
+    },
+  ],
+};
+
+let variance = {
+  labels: data.map((e) => e.x),
+  datasets: [
+    {
+      label: data.map((e) => e.status),
+      borderColor: "black",
+      backgroundColor: data.map((e) => e.color),
+      borderWidth: 2,
+      data: data.map((e) => e.var),
+    },
+  ],
+};
+
+const options_kappa_rho = {
+  plugins: {
+    legend: {
+      display: false,
+    },
+    title: {
+      display: true,
+      text: "Kappa / Rho Plot",
+      font: {
+        size: 20,
+        weight: "bold",
+      },
+    },
+  },
+};
+
+const options_rho = {
+  plugins: {
+    legend: {
+      display: false,
+    },
+    title: {
+      display: true,
+      text: "Rho Rank",
+      font: {
+        size: 20,
+        weight: "bold",
+      },
+    },
+  },
+};
+
+const options_kappa = {
+  plugins: {
+    legend: {
+      display: false,
+    },
+    title: {
+      display: true,
+      text: "Kappa Rank",
+      font: {
+        size: 20,
[…]
