+++ tedana/reporting/html_report.py
@@ -2,12 +2,82 @@ import os
 from os.path import join as opj
 from pathlib import Path
 from string import Template
+import shutil
+import subprocess
 
 import pandas as pd
 from bokeh import __version__ as bokehversion
 from bokeh import embed, layouts, models
 from tedana.info import __version__
-from tedana.reporting import dynamic_figures as df
+import numpy as np
+from sklearn.preprocessing import MinMaxScaler
+from math import pi
+
+color_mapping = {"accepted": "#2ecc71", "rejected": "#e74c3c", "ignored": "#3498db"}
+
+
+def _create_data_struct(comptable_path, color_mapping=color_mapping):
+    """
+    Create Bokeh ColumnDataSource with all info dynamic plots need
+
+    Parameters
+    ----------
+    comptable: str
+        file path to component table, JSON format
+
+    Returns
+    -------
+    cds: bokeh.models.ColumnDataSource
+        Data structure with all the fields to plot or hover over
+    """
+    unused_cols = [
+        "normalized variance explained",
+        "countsigFT2",
+        "countsigFS0",
+        "dice_FS0",
+        "countnoise",
+        "dice_FT2",
+        "signal-noise_t",
+        "signal-noise_p",
+        "d_table_score",
+        "kappa ratio",
+        "rationale",
+        "d_table_score_scrub",
+    ]
+
+    df = pd.read_table(comptable_path)
+    n_comps = df.shape[0]
+
+    # remove space from column name
+    df.rename(columns={"variance explained": "var_exp"}, inplace=True)
+
+    # For providing sizes based on Var Explained that are visible
+    mm_scaler = MinMaxScaler(feature_range=(4, 20))
+    df["var_exp_size"] = mm_scaler.fit_transform(
+        df[["var_exp", "normalized variance explained"]]
+    )[:, 0]
+
+    # Calculate Kappa and Rho ranks
+    df["rho_rank"] = df["rho"].rank(ascending=False).values
+    df["kappa_rank"] = df["kappa"].rank(ascending=False).values
+    df["var_exp_rank"] = df["var_exp"].rank(ascending=False).values
+
+    # Remove unused columns to decrease size of final HTML
+    # set errors to 'ignore' in case some columns do not exist in
+    # a given data frame
+    df.drop(unused_cols, axis=1, inplace=True, errors="ignore")
+
+    # Create additional Column with colors based on final classification
+    df["color"] = [color_mapping[i] for i in df["classification"]]
+
+    # Create additional column with component ID
+    df["component"] = np.arange(n_comps)
+
+    # Compute angle and re-sort data for Pie plots
+    df["angle"] = df["var_exp"] / df["var_exp"].sum() * 2 * pi
+    df.sort_values(by=["classification", "var_exp"], inplace=True)
+
+    return df
 
 
 def _generate_buttons(out_dir):
@@ -49,35 +119,7 @@ def _generate_buttons(out_dir):
     return buttons_html
 
 
-def _update_template_bokeh(bokeh_id, about, bokeh_js, buttons):
-    """
-    Populate a report with content.
-
-    Parameters
-    ----------
-    bokeh_id : str
-        HTML div created by bokeh.embed.components
-    about : str
-        Reporting information for a given run
-    bokeh_js : str
-        Javascript created by bokeh.embed.components
-    Returns
-    -------
-    HTMLReport : an instance of a populated HTML report
-    """
-    resource_path = Path(__file__).resolve().parent.joinpath("data", "html")
-
-    body_template_name = "report_body_template.html"
-    body_template_path = resource_path.joinpath(body_template_name)
[…]
