+++ tedana/reporting/data/src/Panel.js
@@ -1,0 +1,6 @@
+import React from "react";
+// import { TiChartLine, TiInfoLarge, TiWaves } from "react-icons/ti";
+
+export default function Panel(props) {
+  return <div className="panel">{props.children}</div>;
+}
