### 2021-08-01 92fe778
Update checks for tests
