+++ tedana/tests/test_integration.py
@@ -32,13 +32,15 @@ def check_integration_outputs(fname, outpath):
     """
 
     # Gets filepaths generated by integration test
-    existing = [os.path.relpath(f, outpath) for f in
-                glob.glob(os.path.join(outpath, '**'), recursive=True)[1:]]
+    existing = [
+        os.path.relpath(f, outpath)
+        for f in glob.glob(os.path.join(outpath, "**"), recursive=True)[1:]
+    ]
 
     # Checks for log file
-    log_regex = ('^tedana_'
-                 '[12][0-9]{3}-[0-9]{2}-[0-9]{2}T[0-9]{2}'
-                 '[0-9]{2}[0-9]{2}.tsv$')
+    log_regex = (
+        "^tedana_" "[12][0-9]{3}-[0-9]{2}-[0-9]{2}T[0-9]{2}" "[0-9]{2}[0-9]{2}.tsv$"
+    )
     logfiles = [out for out in existing if re.match(log_regex, out)]
     assert len(logfiles) == 1
 
@@ -46,7 +48,7 @@ def check_integration_outputs(fname, outpath):
     existing.remove(logfiles[0])
 
     # Compares remaining files with those expected
-    with open(fname, 'r') as f:
+    with open(fname, "r") as f:
         tocheck = f.read().splitlines()
     tocheck = [os.path.normpath(path) for path in tocheck]
     assert sorted(tocheck) == sorted(existing)
@@ -72,20 +74,18 @@ def download_test_data(osf, outpath):
 
 
 def test_integration_five_echo(skip_integration):
-    """ Integration test of the full tedana workflow using five-echo test data
-    """
+    """Integration test of the full tedana workflow using five-echo test data"""
 
     if skip_integration:
-        pytest.skip('Skipping five-echo integration test')
-    out_dir = '/tmp/data/five-echo/TED.five-echo'
+        pytest.skip("Skipping five-echo integration test")
+    out_dir = "/tmp/data/five-echo/TED.five-echo"
     if os.path.exists(out_dir):
         shutil.rmtree(out_dir)
 
     # download data and run the test
-    download_test_data('https://osf.io/9c42e/download',
-                       os.path.dirname(out_dir))
-    prepend = '/tmp/data/five-echo/p06.SBJ01_S09_Task11_e'
-    suffix = '.sm.nii.gz'
+    download_test_data("https://osf.io/9c42e/download", os.path.dirname(out_dir))
+    prepend = "/tmp/data/five-echo/p06.SBJ01_S09_Task11_e"
+    suffix = ".sm.nii.gz"
     datalist = [prepend + str(i + 1) + suffix for i in range(5)]
     echo_times = [15.4, 29.7, 44.0, 58.3, 72.6]
     tedana_cli.tedana_workflow(
@@ -93,129 +93,157 @@ def test_integration_five_echo(skip_integration):
         tes=echo_times,
         out_dir=out_dir,
         tedpca=0.95,
-        fittype='curvefit',
+        fittype="curvefit",
         fixed_seed=49,
         tedort=True,
-        verbose=True)
+        verbose=True,
+    )
 
     # Just a check on the component table pending a unit test of load_comptable
-    comptable = os.path.join(out_dir, 'desc-tedana_metrics.tsv')
+    comptable = os.path.join(out_dir, "desc-tedana_metrics.tsv")
     df = pd.read_table(comptable)
     assert isinstance(df, pd.DataFrame)
 
     # Test re-running, but use the CLI
-    out_dir2 = '/tmp/data/five-echo/TED.five-echo-manual'
-    acc_comps = df.loc[df['classification'] == 'accepted'].index.values
+    out_dir2 = "/tmp/data/five-echo/TED.five-echo-manual"
[…]
