### 2021-08-01 a4d8755
Add npm installation for integration tests
