+++ .circleci/config.yml
@@ -6,7 +6,6 @@ version: 2.1
 orbs:
   codecov: codecov/codecov@1.0.5
 jobs:
-
   makeenv_37:
     docker:
       - image: continuumio/miniconda3
@@ -26,7 +25,7 @@ jobs:
       - save_cache:
           key: conda-py37-v1-{{ checksum "tedana/info.py" }}
           paths:
-              - /opt/conda/envs/tedana_py37
+            - /opt/conda/envs/tedana_py37
 
   unittest_36:
     docker:
@@ -56,11 +55,11 @@ jobs:
       - save_cache:
           key: conda-py36-v1-{{ checksum "tedana/info.py" }}
           paths:
-              - /opt/conda/envs/tedana_py36
+            - /opt/conda/envs/tedana_py36
       - persist_to_workspace:
           root: /tmp
           paths:
-              - src/coverage/.coverage.py36
+            - src/coverage/.coverage.py36
 
   unittest_37:
     docker:
@@ -82,7 +81,7 @@ jobs:
       - persist_to_workspace:
           root: /tmp
           paths:
-              - src/coverage/.coverage.py37
+            - src/coverage/.coverage.py37
 
   unittest_38:
     docker:
@@ -112,11 +111,11 @@ jobs:
       - save_cache:
           key: conda-py38-v1-{{ checksum "tedana/info.py" }}
           paths:
-              - /opt/conda/envs/tedana_py38
+            - /opt/conda/envs/tedana_py38
       - persist_to_workspace:
           root: /tmp
           paths:
-              - src/coverage/.coverage.py38
+            - src/coverage/.coverage.py38
 
   unittest_39:
     docker:
@@ -146,11 +145,11 @@ jobs:
       - save_cache:
           key: conda-py39-v1-{{ checksum "tedana/info.py" }}
           paths:
-              - /opt/conda/envs/tedana_py39
+            - /opt/conda/envs/tedana_py39
       - persist_to_workspace:
           root: /tmp
           paths:
-              - src/coverage/.coverage.py39
+            - src/coverage/.coverage.py39
 
   style_check:
     docker:
@@ -191,7 +190,7 @@ jobs:
       - persist_to_workspace:
           root: /tmp
           paths:
-              - src/coverage/.coverage.three-echo
+            - src/coverage/.coverage.three-echo
 
   four-echo:
     docker:
@@ -207,6 +206,7 @@ jobs:
           command: |
             apt-get update
             apt-get install -yqq make
+            apt-get install npm
             source activate tedana_py37  # depends on makeenv_37
             make four-echo
             mkdir /tmp/src/coverage
@@ -216,7 +216,7 @@ jobs:
       - persist_to_workspace:
           root: /tmp
           paths:
-              - src/coverage/.coverage.four-echo
+            - src/coverage/.coverage.four-echo
 
   five-echo:
     docker:
@@ -241,7 +241,7 @@ jobs:
       - persist_to_workspace:
           root: /tmp
           paths:
-              - src/coverage/.coverage.five-echo
+            - src/coverage/.coverage.five-echo
 
   t2smap:
     docker:
@@ -266,7 +266,7 @@ jobs:
       - persist_to_workspace:
           root: /tmp
           paths:
-              - src/coverage/.coverage.t2smap
+            - src/coverage/.coverage.t2smap
 
   merge_coverage:
     working_directory: /tmp/src/tedana
