### 2021-08-01 0e79c96
Add missing installs
