+++ .circleci/config.yml
@@ -181,7 +181,7 @@ jobs:
           command: |
             apt-get update
             apt-get install -yqq make
-            apt-get install npm
+            apt-get install -yqq npm
             source activate tedana_py37  # depends on makeenv_37
             make three-echo
             mkdir /tmp/src/coverage
@@ -207,7 +207,7 @@ jobs:
           command: |
             apt-get update
             apt-get install -yqq make
-            apt-get install npm
+            apt-get install -yqq npm
             source activate tedana_py37  # depends on makeenv_37
             make four-echo
             mkdir /tmp/src/coverage
@@ -233,7 +233,7 @@ jobs:
           command: |
             apt-get update
             apt-get install -yqq make
-            apt-get install npm
+            apt-get install -yqq npm
             source activate tedana_py37  # depends on makeenv_37
             make five-echo
             mkdir /tmp/src/coverage
