### 2021-08-13 3147e01
[ENH] Use different masking thresholds for denoising and classification (#736)
* Use two different adaptive masks.

masksum_denoise: Threshold of 1, used for denoising and optimal combination.
masksum_clf: Threshold of 3, used for decomposition and component selection.

* Fix long line.

* Derive classifier masks directly from denoising ones.

* Use denoising mask for component figures.

* Swap full and limited T2*/S0 maps.

Now that the full maps are used throughout the pipeline, they are the primary T2*/S0 outputs. The limited maps are not used for anything, and are only written out if the w
[…]
