+++ tedana/workflows/t2smap.py
@@ -171,21 +171,27 @@ def t2smap_workflow(data, tes, out_dir='.', mask=None,
     -----
     This workflow writes out several files, which are described below:
 
-    ==========================    =================================================
+    ============================= =================================================
     Filename                      Content
-    ==========================    =================================================
-    T2starmap.nii.gz              Limited estimated T2* 3D map or 4D timeseries.
+    ============================= =================================================
+    T2starmap.nii.gz              Estimated T2* 3D map or 4D timeseries.
                                   Will be a 3D map if ``fitmode`` is 'all' and a
                                   4D timeseries if it is 'ts'.
-    S0map.nii.gz                  Limited S0 3D map or 4D timeseries.
-    desc-full_T2starmap.nii.gz    Full T2* map/timeseries. The difference between
+    S0map.nii.gz                  S0 3D map or 4D timeseries.
+    desc-limited_T2starmap.nii.gz Limited T2* map/timeseries. The difference between
                                   the limited and full maps is that, for voxels
                                   affected by dropout where only one echo contains
-                                  good data, the full map uses the single echo's
-                                  value while the limited map has a NaN.
-    desc-full_S0map.nii.gz        Full S0 map/timeseries.
+                                  good data, the full map uses the T2* estimate
+                                  from the first two echos, while the limited map
+                                  will have a NaN.
+    desc-limited_S0map.nii.gz     Limited S0 map/timeseries. The difference between
+                                  the limited and full maps is that, for voxels
+                                  affected by dropout where only one echo contains
+                                  good data, the full map uses the S0 estimate
+                                  from the first two echos, while the limited map
+                                  will have a NaN.
     desc-optcom_bold.nii.gz       Optimally combined timeseries.
-    ==========================    =================================================
+    ============================= =================================================
     """
     out_dir = op.abspath(out_dir)
     if not op.isdir(out_dir):
@@ -234,11 +240,11 @@ def t2smap_workflow(data, tes, out_dir='.', mask=None,
 
     # set a hard cap for the T2* map/timeseries
     # anything that is 10x higher than the 99.5 %ile will be reset to 99.5 %ile
-    cap_t2s = stats.scoreatpercentile(t2s_limited.flatten(), 99.5,
+    cap_t2s = stats.scoreatpercentile(t2s_full.flatten(), 99.5,
                                       interpolation_method='lower')
     cap_t2s_sec = utils.millisec2sec(cap_t2s * 10.)
     LGR.debug('Setting cap on T2* map at {:.5f}s'.format(cap_t2s_sec))
-    t2s_limited[t2s_limited > cap_t2s * 10] = cap_t2s
+    t2s_full[t2s_full > cap_t2s * 10] = cap_t2s
 
     LGR.info('Computing optimal combination')
     # optimally combine data
@@ -246,24 +252,24 @@ def t2smap_workflow(data, tes, out_dir='.', mask=None,
                                  combmode=combmode)
 
     # clean up numerical errors
-    for arr in (OCcatd, s0_limited, t2s_limited):
+    for arr in (OCcatd, s0_full, t2s_full):
         np.nan_to_num(arr, copy=False)
 
-    s0_limited[s0_limited < 0] = 0
-    t2s_limited[t2s_limited < 0] = 0
+    s0_full[s0_full < 0] = 0
+    t2s_full[t2s_full < 0] = 0
 
     io_generator.save_file(
-        utils.millisec2sec(t2s_limited),
+        utils.millisec2sec(t2s_full),
         't2star img',
     )
-    io_generator.save_file(s0_limited, 's0 img')
+    io_generator.save_file(s0_full, 's0 img')
[…]
