+++ tedana/tests/test_t2smap.py
@@ -31,9 +31,9 @@ class TestT2smap():
         assert len(img.shape) == 3
         img = nib.load(op.join(out_dir, 'S0map.nii.gz'))
         assert len(img.shape) == 3
-        img = nib.load(op.join(out_dir, 'desc-full_T2starmap.nii.gz'))
+        img = nib.load(op.join(out_dir, 'desc-limited_T2starmap.nii.gz'))
         assert len(img.shape) == 3
-        img = nib.load(op.join(out_dir, 'desc-full_S0map.nii.gz'))
+        img = nib.load(op.join(out_dir, 'desc-limited_S0map.nii.gz'))
         assert len(img.shape) == 3
         img = nib.load(op.join(out_dir, 'desc-optcom_bold.nii.gz'))
         assert len(img.shape) == 4
@@ -57,9 +57,9 @@ class TestT2smap():
         assert len(img.shape) == 4
         img = nib.load(op.join(out_dir, 'S0map.nii.gz'))
         assert len(img.shape) == 4
-        img = nib.load(op.join(out_dir, 'desc-full_T2starmap.nii.gz'))
+        img = nib.load(op.join(out_dir, 'desc-limited_T2starmap.nii.gz'))
         assert len(img.shape) == 4
-        img = nib.load(op.join(out_dir, 'desc-full_S0map.nii.gz'))
+        img = nib.load(op.join(out_dir, 'desc-limited_S0map.nii.gz'))
         assert len(img.shape) == 4
         img = nib.load(op.join(out_dir, 'desc-optcom_bold.nii.gz'))
         assert len(img.shape) == 4
@@ -83,9 +83,9 @@ class TestT2smap():
         assert len(img.shape) == 3
         img = nib.load(op.join(out_dir, 'S0map.nii.gz'))
         assert len(img.shape) == 3
-        img = nib.load(op.join(out_dir, 'desc-full_T2starmap.nii.gz'))
+        img = nib.load(op.join(out_dir, 'desc-limited_T2starmap.nii.gz'))
         assert len(img.shape) == 3
-        img = nib.load(op.join(out_dir, 'desc-full_S0map.nii.gz'))
+        img = nib.load(op.join(out_dir, 'desc-limited_S0map.nii.gz'))
         assert len(img.shape) == 3
         img = nib.load(op.join(out_dir, 'desc-optcom_bold.nii.gz'))
         assert len(img.shape) == 4
@@ -109,9 +109,9 @@ class TestT2smap():
         assert len(img.shape) == 4
         img = nib.load(op.join(out_dir, 'S0map.nii.gz'))
         assert len(img.shape) == 4
-        img = nib.load(op.join(out_dir, 'desc-full_T2starmap.nii.gz'))
+        img = nib.load(op.join(out_dir, 'desc-limited_T2starmap.nii.gz'))
         assert len(img.shape) == 4
-        img = nib.load(op.join(out_dir, 'desc-full_S0map.nii.gz'))
+        img = nib.load(op.join(out_dir, 'desc-limited_S0map.nii.gz'))
         assert len(img.shape) == 4
         img = nib.load(op.join(out_dir, 'desc-optcom_bold.nii.gz'))
         assert len(img.shape) == 4
@@ -135,9 +135,9 @@ class TestT2smap():
         assert len(img.shape) == 3
         img = nib.load(op.join(out_dir, 'S0map.nii.gz'))
         assert len(img.shape) == 3
-        img = nib.load(op.join(out_dir, 'desc-full_T2starmap.nii.gz'))
+        img = nib.load(op.join(out_dir, 'desc-limited_T2starmap.nii.gz'))
         assert len(img.shape) == 3
[…]
