### 2021-08-16 fdcdb22
Add resources folder to package data.
