+++ setup.py
@@ -32,6 +32,7 @@ def main():
             'tests/data/*',
             'reporting/data/*',
             'reporting/data/html/*',
+            'resources/*',
         ]
     }
 
