### 2021-08-19 d4649ee
[FIX] Add resources folder to package data (#772)
* Add resources folder to package data.

* It's not recursive.
