+++ docs/multi-echo.rst
@@ -559,6 +559,8 @@ This list is not necessarily up to date, so please check out OpenNeuro to potent
 * `Valence processing differs across stimulus modalities`_
 * `Cambridge Centre for Ageing Neuroscience (Cam-CAN)`_
 * `rt-me-fMRI - A task and resting state dataset for real-time, multi-echo fMRI methods development and validation`_
+* `EuskalIBUR`_: 7 subjects, 10 sessions per subject, five-echo fMRI with one task (breath-hold).
+  Also includes physio.
 
 .. _Multi-echo fMRI replication sample of autobiographical memory, prospection and theory of mind reasoning tasks: https://openneuro.org/datasets/ds000210/
 .. _Multi-echo Cambridge: https://openneuro.org/datasets/ds000258
@@ -566,6 +568,7 @@ This list is not necessarily up to date, so please check out OpenNeuro to potent
 .. _Valence processing differs across stimulus modalities: https://openneuro.org/datasets/ds001491
 .. _Cambridge Centre for Ageing Neuroscience (Cam-CAN): https://camcan-archive.mrc-cbu.cam.ac.uk/dataaccess/
 .. _rt-me-fMRI - A task and resting state dataset for real-time, multi-echo fMRI methods development and validation: https://doi.org/10.34894/R1TNL8
+.. _EuskalIBUR: https://openneuro.org/datasets/ds003192
 
 
 .. _spreadsheet of publications:
