### 2021-08-25 0dfb91a
Suppresses divide by 0 warning
