### 2021-08-30 #3669 @whedon
[REVIEW]: TE-dependent analysis of multi-echo fMRI with *tedana*
**Submitting author:** @tsalo (<a href="http://orcid.org/0000-0001-9813-3167">Taylor Salo</a>)
**Repository:** <a href="https://github.com/ME-ICA/tedana" target ="_blank">https://github.com/ME-ICA/tedana</a>
**Version:** 0.0.9
