### 2021-09-02 ab7afd8
[DOC] Add fMRIPrep collection information to FAQ (#773)
* Add gist to FAQ.

* Got it working!! (cue manic laughter)

* Add missing links.

* Nothing to see here...

* Add link to original issue.

* Address @eurunuela's review.
