### 2021-09-08 6644a17
Add console log and accept components with index 0 for updating figures on reports
