### 2021-09-09 c98dae4
normalization text with more nuance
