### 2021-09-10 46081f1
Add kappa, rho and variance ranks to metrics table for React reports.
