+++ tedana/metrics/collect.py
@@ -4,14 +4,11 @@ import os.path as op
 
 import numpy as np
 import pandas as pd
-
-from . import dependence
-from ._utils import determine_signs, flip_components, dependency_resolver
-
-from tedana import io
-from tedana import utils
+from tedana import io, utils
 from tedana.stats import getfbounds
 
+from . import dependence
+from ._utils import dependency_resolver, determine_signs, flip_components
 
 LGR = logging.getLogger("GENERAL")
 RepLGR = logging.getLogger("REPORT")
@@ -116,9 +113,7 @@ def generate_metrics(
     n_components = mixing.shape[1]
     comptable = pd.DataFrame(index=np.arange(n_components, dtype=int))
     comptable["Component"] = [
-        io.add_decomp_prefix(
-            comp, prefix=label, max_value=comptable.shape[0]
-        )
+        io.add_decomp_prefix(comp, prefix=label, max_value=comptable.shape[0])
         for comp in comptable.index.values
     ]
 
@@ -140,9 +135,7 @@ def generate_metrics(
             data_optcom, mixing
         )
         if io_generator.verbose:
-            metric_maps["map echo betas"] = dependence.calculate_betas(
-                data_cat, mixing
-            )
+            metric_maps["map echo betas"] = dependence.calculate_betas(data_cat, mixing)
 
     if "map percent signal change" in required_metrics:
         LGR.info("Calculating percent signal change maps")
@@ -158,7 +151,7 @@ def generate_metrics(
         if io_generator.verbose:
             io_generator.save_file(
                 utils.unmask(metric_maps["map Z"] ** 2, mask),
-                label + ' component weights img',
+                label + " component weights img",
             )
 
     if ("map FT2" in required_metrics) or ("map FS0" in required_metrics):
@@ -231,11 +224,22 @@ def generate_metrics(
             Z_maps=metric_maps["map Z"],
         )
 
+    if ("kappa rank" in required_metrics) or ("rho rank" in required_metrics):
+        LGR.info("Ranking kappa and rho")
+        comptable["rho_rank"] = comptable["rho"].rank(ascending=False).values
+        comptable["kappa_rank"] = comptable["kappa"].rank(ascending=False).values
+
     # Generic metrics
     if "variance explained" in required_metrics:
         LGR.info("Calculating variance explained")
         comptable["variance explained"] = dependence.calculate_varex(
             metric_maps["map optcom betas"]
+        )
+
+    if "variance explained rank" in required_metrics:
+        LGR.info("Ranking variance explained")
+        comptable["variance explained rank"] = (
+            comptable["variance explained"].rank(ascending=False).values
         )
 
     if "normalized variance explained" in required_metrics:
@@ -341,37 +345,47 @@ def generate_metrics(
                 echo_betas = betas[:, i_echo, :]
                 io_generator.save_file(
                     utils.unmask(echo_betas, mask),
-                    'echo weight ' + label + ' map split img',
-                    echo=(i_echo + 1)
+                    "echo weight " + label + " map split img",
+                    echo=(i_echo + 1),
                 )
 
             if write_T2S0:
                 echo_pred_T2_maps = pred_T2_maps[:, i_echo, :]
                 io_generator.save_file(
                     utils.unmask(echo_pred_T2_maps, mask),
-                    'echo T2 ' + label + ' split img',
-                    echo=(i_echo + 1)
+                    "echo T2 " + label + " split img",
+                    echo=(i_echo + 1),
                 )
 
                 echo_pred_S0_maps = pred_S0_maps[:, i_echo, :]
                 io_generator.save_file(
                     utils.unmask(echo_pred_S0_maps, mask),
-                    'echo S0 ' + label + ' split img',
[…]
