+++ tedana/workflows/tedana.py
@@ -1,31 +1,40 @@
 """
 Run the "canonical" TE-Dependent ANAlysis workflow.
 """
+import argparse
+import datetime
 import json
+import logging
 import os
-import sys
 import os.path as op
 import shutil
-import logging
-import datetime
+import sys
 from glob import glob
 
-import argparse
 import numpy as np
 import pandas as pd
+from nilearn.masking import compute_epi_mask
 from scipy import stats
 from threadpoolctl import threadpool_limits
-from nilearn.masking import compute_epi_mask
 
-from tedana import (decay, combine, decomposition, io, metrics,
-                    reporting, selection, utils, __version__)
 import tedana.gscontrol as gsc
+from tedana import (
+    __version__,
+    combine,
+    decay,
+    decomposition,
+    io,
+    metrics,
+    reporting,
+    selection,
+    utils,
+)
 from tedana.stats import computefeats2
-from tedana.workflows.parser_utils import is_valid_file, check_tedpca_value
+from tedana.workflows.parser_utils import check_tedpca_value, is_valid_file
 
 LGR = logging.getLogger("GENERAL")
-RepLGR = logging.getLogger('REPORT')
-RefLGR = logging.getLogger('REFERENCES')
+RepLGR = logging.getLogger("REPORT")
+RefLGR = logging.getLogger("REFERENCES")
 
 
 def _get_parser():
@@ -37,230 +46,322 @@ def _get_parser():
     parser.parse_args() : argparse dict
     """
     from ..info import __version__
-    verstr = 'tedana v{}'.format(__version__)
+
+    verstr = "tedana v{}".format(__version__)
     parser = argparse.ArgumentParser()
     # Argument parser follow templtate provided by RalphyZ
     # https://stackoverflow.com/a/43456577
     optional = parser._action_groups.pop()
-    required = parser.add_argument_group('Required Arguments')
-    required.add_argument('-d',
-                          dest='data',
-                          nargs='+',
-                          metavar='FILE',
-                          type=lambda x: is_valid_file(parser, x),
-                          help=('Multi-echo dataset for analysis. May be a '
-                                'single file with spatially concatenated data '
-                                'or a set of echo-specific files, in the same '
-                                'order as the TEs are listed in the -e '
-                                'argument.'),
-                          required=True)
-    required.add_argument('-e',
-                          dest='tes',
-                          nargs='+',
-                          metavar='TE',
-                          type=float,
-                          help='Echo times (in ms). E.g., 15.0 39.0 63.0',
-                          required=True)
-    optional.add_argument('--out-dir',
-                          dest='out_dir',
-                          type=str,
-                          metavar='PATH',
-                          help='Output directory.',
-                          default='.')
-    optional.add_argument('--mask',
-                          dest='mask',
-                          metavar='FILE',
-                          type=lambda x: is_valid_file(parser, x),
-                          help=("Binary mask of voxels to include in TE "
-                                "Dependent ANAlysis. Must be in the same "
-                                "space as `data`. If an explicit mask is not "
-                                "provided, then Nilearn's compute_epi_mask "
-                                "function will be used to derive a mask "
-                                "from the first echo's data."),
-                          default=None)
-    optional.add_argument('--prefix',
-                          dest='prefix',
-                          type=str,
-                          help="Prefix for filenames generated.",
-                          default='')
-    optional.add_argument('--convention',
-                          dest='convention',
-                          action='store',
-                          choices=['orig', 'bids'],
-                          help=("Filenaming convention. bids will use "
-                                "the latest BIDS derivatives version."),
-                          default='bids')
-    optional.add_argument('--fittype',
-                          dest='fittype',
-                          action='store',
-                          choices=['loglin', 'curvefit'],
[…]
