### 2021-09-10 bb02ac3
[DOC] Generalize preprocessing recommendations (#769)
* Generalize preproc recommendations.

* normalization text with more nuance

* Just two minor typos.

Co-authored-by: handwerkerd <7406227+handwerkerd@users.noreply.github.com>
