### 2021-09-13 e4f1c36
Force Bokeh version to be lower than 3.0
