### 2021-09-13 bb57384
[STY] Use `black` and `isort` to manage library code style (#758)
* Add black and isort to config and README.

* Remove build system requirements.

* Run black and isort on package.

* Fix combined multiline strings.

It's the one thing black always messes up.

* Run black again.

* Replace single-quote strings.

* Fix import.

* isort

* Document our preferred code style.

* Run black again.
