+++ tedana/workflows/parser_utils.py
@@ -1,9 +1,8 @@
 """
 Functions for parsers.
 """
-import os.path as op
-
 import argparse
+import os.path as op
 
 
 def check_tedpca_value(string, is_parser=True):
@@ -16,9 +15,7 @@ def check_tedpca_value(string, is_parser=True):
     try:
         floatarg = float(string)
     except ValueError:
-        msg = "Argument to tedpca must be a float or one of: {}".format(
-            ", ".join(valid_options)
-        )
+        msg = "Argument to tedpca must be a float or one of: {}".format(", ".join(valid_options))
         raise error(msg)
 
     if not (0 <= floatarg <= 1):
@@ -31,6 +28,6 @@ def is_valid_file(parser, arg):
     Check if argument is existing file.
     """
     if not op.isfile(arg) and arg is not None:
-        parser.error('The file {0} does not exist!'.format(arg))
+        parser.error("The file {0} does not exist!".format(arg))
 
     return arg
