+++ tedana/io.py
@@ -4,14 +4,14 @@ Other functions in the module help write outputs which require multiple
 data sources, assist in writing per-echo verbose outputs, or act as helper
 functions for any of the above.
 """
+import json
 import logging
 import os
 import os.path as op
-import json
 from string import Formatter
 
-import numpy as np
 import nibabel as nib
+import numpy as np
 import pandas as pd
 from nilearn._utils import check_niimg
 from nilearn.image import new_img_like
@@ -19,13 +19,12 @@ from nilearn.image import new_img_like
 from tedana import utils
 from tedana.stats import computefeats2, get_coeffs
 
-
 LGR = logging.getLogger("GENERAL")
-RepLGR = logging.getLogger('REPORT')
-RefLGR = logging.getLogger('REFERENCES')
+RepLGR = logging.getLogger("REPORT")
+RefLGR = logging.getLogger("REFERENCES")
 
 
-class OutputGenerator():
+class OutputGenerator:
     """A class for managing tedana outputs.
 
     Parameters
@@ -170,9 +169,9 @@ class OutputGenerator():
         for key, value in kwargs.items():
             if key not in name_variables:
                 raise ValueError(
-                    f'Argument {key} passed but has no match in format '
-                    f'string. Available format variables: '
-                    f'{name_variables} from {kwargs} and {name}.'
+                    f"Argument {key} passed but has no match in format "
+                    f"string. Available format variables: "
+                    f"{name_variables} from {kwargs} and {name}."
                 )
 
         name = name.format(**kwargs)
@@ -221,14 +220,10 @@ class OutputGenerator():
         """
         data_type = type(data)
         if not isinstance(data, np.ndarray):
-            raise TypeError(
-                f"Data supplied must of type np.ndarray, not {data_type}."
-            )
+            raise TypeError(f"Data supplied must of type np.ndarray, not {data_type}.")
         if data.ndim not in (1, 2):
-            raise TypeError(
-                "Data must have number of dimensions in (1, 2), not "
-                f"{data.ndim}"
-            )
+            raise TypeError(f"Data must have number of dimensions in (1, 2), not {data.ndim}")
+
         # Coerce data to be 32-bit max in the cases of float64, int64
         # Note that int64 niftis cannot be read by mricroGL, AFNI
         vox_type = data.dtype
@@ -236,6 +231,7 @@ class OutputGenerator():
             data = np.int32(data)
         elif vox_type == np.float64:
             data = np.float32(data)
+
         # Make new img and save
         img = new_nii_like(self.reference_img, data)
         img.to_filename(name)
@@ -305,7 +301,7 @@ def load_json(path: str) -> dict:
     FileNotFoundError if the file does not exist
     IsADirectoryError if the path is a directory instead of a file
     """
-    with open(path, 'r') as f:
+    with open(path, "r") as f:
         try:
             data = json.load(f)
         except json.decoder.JSONDecodeError:
@@ -335,8 +331,8 @@ def add_decomp_prefix(comp_num, prefix, max_value):
         Component name in the form <prefix>_<zeros><comp_num>
     """
     n_digits = int(np.log10(max_value)) + 1
-    comp_name = '{0:08d}'.format(int(comp_num))
-    comp_name = '{0}_{1}'.format(prefix, comp_name[8 - n_digits:])
+    comp_name = "{0:08d}".format(int(comp_num))
+    comp_name = "{0}_{1}".format(prefix, comp_name[8 - n_digits :])
     return comp_name
 
 
[…]
