+++ tedana/tests/test_reporting.py
@@ -2,6 +2,7 @@
 Tests for tedana.reporting
 """
 import numpy as np
+
 from tedana import reporting
 
 
@@ -11,5 +12,6 @@ def test_smoke_trim_edge_zeros():
     """
     arr = np.random.random((100, 100))
     assert reporting.static_figures._trim_edge_zeros(arr) is not None
+
 
 # TODO: Test other functions in reporting?

[…]
