### 2021-09-14 ME-ICA/tedana#798 @martinagvilas
[JOSS review] Documentation suggestions
I found the documentation super complete and easy to follow. I really don't have more than a few minor suggestions following the JOSS reviewing guidelines:

- [ ] In the [installation instructions](https://github.com/ME-ICA/tedana/blob/bb573840da9176374ee8f4e9cb50e15ff7f097e1/CONTRIBUTING.md#3-run-the-developer-setup) for the developer version, I think it should say `pip install -e .'[all]'` (with quotes surrounding `[all]`), at least for my operating system (MacOS).

- [ ] Maybe some interactive demos might be useful? I see some example notebooks [here](https://github.com/ME-ICA/tedana/tree/8477aabc2081c1b97ad646b203516ccea9ad75d8/examples) that are not linked in the online documentation (but maybe I missed them). I see they are basically showing the same steps as outlined [here](https://tedana.readthedocs.io/en/latest/approach.html) but I still think it would be useful for users to be able to interact with that walkthrough and for you to also provide (or point to) a downloadable dataset that can be used for the demo, if possible.

- [ ] Feel free to disregard if you don't agree, but I think it would be clearer for new users (at least in terms of functionality) if the about section in the readme and online documentation first highlights the purpose of the toolbox, and then the context and prospective user case. Something like this rearrangement seems clearer to me:

    > `TE`-`de`pendent `ana`lysis (`tedana`) is a Python module for denoising multi-echo functional magnetic resonance imaging (fMRI) data. The `tedana` package is part of the ME-ICA pipeline, performing TE-dependent analysis of multi-echo functional magnetic resonance imaging (fMRI) data. The ME-ICA pipeline originally performed both pre-processing and TE-dependent analysis of multi-echo fMRI data; however, `tedana` currently assumes that you’re working with data which has been previously preprocessed.

- [ ] Maybe it would be nice to explicitly tell users how to check if their installation is correct. I find the example recommendation of [nilearn](https://nilearn.github.io/introduction.html) very user-friendly, but of course you are probably aware of it :)

https://github.com/openjournals/joss-reviews/issues/3669
