### 2021-09-14 ME-ICA/tedana#799 @martinagvilas
[JOSS review] Manuscript suggestions
Following JOSS [paper guidelines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain), I have a few suggestions for the manuscript:

- [ ] The text in the figure looks small and is difficult to read, and it also downscales the rest of the images in the figure. I'm guessing you intend to show how exactly the report looks as output, but I think in this case it would be more friendly if you include the images separately from the text, and change the format of the text to be more readable and in line with the manuscript style. The real report can then be accessed by following the link you provided.
- [ ] I think in the _statement of need_ section you need to clearly state again which two use-case driven workflows is your toolbox targeting. Maybe to avoid repeating too much, the last paragraph of the _summary_ section can be integrated in the _statement of need_ section when your refer to "its two use-case (...)" in line 63.
- [ ] This is a stylistic comment, so maybe you won't agree with it. In my opinion, the first paragraphs of the manuscript might be too technical for a non-specialized audience. But I really found the _motivation_ section of your [posters](https://tedana.readthedocs.io/en/latest/index.html#posters) really approachable. Perhaps some of its structure could be copied for the _summary_ part of the manuscript? Maybe the first paragraph could be an overview paragraph mentioning the points of the poster (what fmri is, that is noisy, that multi-echo is poised to ameliorate some of the artifacts), and then in the next sections introduce the more specific paragraphs that you have already written which expand on these topics.

https://github.com/openjournals/joss-reviews/issues/3669
