### 2021-09-14 423491f
[FIX] Figures of components with index 0 get rendered now (#793)
* Add console log and accept components with index 0 for updating figures on reports

* Remove breakpoint

* Force Bokeh version to be lower than 3.0

* Fix typo

* Changed auto-save settings to line length of 99
