### 2021-09-14 9670fad
[DOC] Fix link to developer guidelines in README (#797)
