### 2021-09-17 3ef3ccf
Address martinagvilas JOSS review comments
