+++ paper/paper.md
@@ -102,13 +102,13 @@ Rather than collect data at a single TE, an alternative approach is to collect m
 
 Continuing these efforts, we present *tedana* (TE-Dependent ANAlysis) as an open-source Python package for processing and denoising multi-echo fMRI data. *tedana* implements two approaches to multi-echo preprocessing: (1) estimating a T~2~^\*^ map and using these values to generate a weighted sum of individual echos, and  (2) using echo-time dependent information in analysis and denoising [@Kundu2012-bq].
 
-We have tightly scoped *tedana* processing to focus on those portions of the fMRI analysis workflow which are multi-echo specific. The primary interfaces for users are (1) a ``t2smap`` workflow, which estimates voxel-wise T~2~^\*^ and S~0~ and combines data across echos to increase temporal SNR, and (2) a full ``tedana`` workflow, which performs the same steps as the ``t2smap`` workflow and additionally performs ICA-based denoising to remove components exhibiting noise-like signal decay patterns across echos [@Kundu2012-bq]. The ``tedana`` workflow additionally generates interactive HTML reports through which users may visually inspect their denoising results and evaluate each component’s classification. An example report is presented in \autoref{fig:report}.
-
 # Statement of Need
 
 To date, multi-echo fMRI has not been widely adopted within the neuroimaging community. This is likely due to two constraints: (1) until recently, the lack of available multi-echo fMRI acquisition protocols, and (2) the lack of software for processing multi-echo fMRI data in a way that integrates with existing platforms, such as AFNI [@Cox1996-up], SPM [@Penny2011-vi], FSL [@Jenkinson2012-eh], and fMRIPrep [@Esteban2020-ul].
 
-*tedana* helps to address these gaps both as a software tool and as a community of practice. As a software tool, its two use-case driven workflows each cover distinct, popular analyses in multi-echo fMRI processing. The limited focus and modularity of each workflow allows for easy integration into existing fMRI processing platforms. Individual modules also allow researchers to flexibly perform T~2~^\*^/S~0~ estimation, combination across echos, decomposition with PCA or ICA, and component selection outside of a specific workflow call. As a community of practice, *tedana* serves as a resource for researchers looking to learn more about multi-echo fMRI, from theory to collection to analysis. To specifically increase the availability of multi-echo protocols, *tedana’s* documentation (available at https://tedana.readthedocs.io) consolidates acquisition guidelines for multi-echo sequences across a variety of field strengths and scanner vendors, as well as general recommendations for balancing relevant trade-offs in fMRI acquisition parameter choices. It further serves to consolidate community knowledge, including guides explaining the underlying principles of multi-echo fMRI and information on publicly available multi-echo datasets and general recommendations for balancing relevant trade-offs in sequence development.
+*tedana* helps to address these gaps both as a software tool and as a community of practice. We have tightly scoped *tedana* processing to focus on those portions of the fMRI analysis workflow which are multi-echo specific in order to maximize their compatibility with other community tools. The primary interfaces for users are (1) a ``t2smap`` workflow, which estimates voxel-wise T~2~^\*^ and S~0~ and combines data across echos to increase temporal SNR, and (2) a full ``tedana`` workflow, which performs the same steps as the ``t2smap`` workflow and additionally performs ICA-based denoising to remove components exhibiting noise-like signal decay patterns across echos [@Kundu2012-bq]. The ``tedana`` workflow additionally generates interactive HTML reports through which users may visually inspect their denoising results and evaluate each component’s classification. An example report is presented in \autoref{fig:report}.
+
[…]
