+++ docs/installation.rst
@@ -18,4 +18,16 @@ You can then install ``tedana`` with:
   pip install tedana
 
 In addition to the Python package, installing ``tedana`` will add the ``tedana``
-and ``t2smap`` workflow CLIs to your path.
+and ``t2smap`` workflow CLIs to your path.You can confirm that ``tedana`` has successfully installed by launching a Python instance and running:
+
+.. code-block:: python
+
+  import tedana
+
+You can check that it is available through the command line interface (CLI) with:
+
+.. code-block:: bash
+
+  tedana --help
+
+If no error occurs, ``tedana`` has correctly installed in your environment!
