### 2021-09-17 35ef357
Rename utils.load_image to utils.reshape_niimg
This should make it much clearer what the function does.
