### 2021-09-20 ME-ICA/tedana#804 @stebo85
[JOSS review] Documentation suggestions
The Documentation of the project is very comprehensive and includes all information one could hope for. I only have small suggestions for further potential improvement if the authors think this is feasible and relevant:

1) I was wondering if it would be possible to add examples in the section "Using tedana" including a small dataset that users could use to run tedana or t2smap on before applying it to their own data. Ideally the most important outputs of these examples could be briefly explained and then linked to the full "Outputs of tedana" section.

2) The duecredit warning is described in the FAQ, but for a naive user it looks like something went wrong. Would it make sense to link to this explanation from the Installation section?

https://github.com/openjournals/joss-reviews/issues/3669
