### 2021-09-20 ME-ICA/tedana#798 @emdupre
Thanks for your review, @martinagvilas ! We've pushed changes to the `JOSS` branch in https://github.com/ME-ICA/tedana/commit/3ef3ccfaed49eed879a279c5ca95502b702a5f1b which I believe address your comments  :smile_cat:

The one point I want to explicitly follow-up on is the idea for interactive demos. We definitely agree that these would be useful ! We're actually working on making datasets easily available within `tedana` (see #720), which will help with this goal. However, this work is still ongoing -- and using these datasets in the walk-through will require a larger update of the documentation. Specifically, we'll need to move away from the current 8-echo demonstration dataset (which is not openly available) to another publicly available dataset with a large number of echos.

That said, we have renamed the previous `Examples` folder to `Notebooks` to make it clearer to users that these -- at present -- are not intended to be runnable examples. We've also moved the folder from the top-level directory into the documentation, which should hopefully better situate it in its appropriate context !

Please let me know if this (and the linked commit !) addresses your concerns, here !
