### 2021-09-20 ME-ICA/tedana#799 @emdupre
Thanks again, @martinagvilas ! :sunflower:

We pushed changed to address your first two points in https://github.com/ME-ICA/tedana/commit/3ef3ccfaed49eed879a279c5ca95502b702a5f1b on the `JOSS` branch ! You can find the specific diff for the paper [here](https://github.com/ME-ICA/tedana/commit/3ef3ccfaed49eed879a279c5ca95502b702a5f1b#diff-e504eb580b095a7e65428b098183a581e475f0fb316db95287eacd7d4f344424).

To follow-up on the last point : I absolutely understand your concern about a non-technical audience, but we're somewhat limited in space based on the [JOSS paper guidelines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain). In drafting the paper, we had elected to spend that space on the Statement of Need rather than a longer summary -- but if you have suggestions for what could be cut across the paper to make room for another paragraph about fMRI more generally, please let us know !
