### 2021-09-20 0f04f76
[FIX] Check data type in `io.load_data` (#802)
* Close tuple loophole.

* Rename utils.load_image to utils.reshape_niimg

This should make it much clearer what the function does.

* Add comment.

* Expand tests.
