### 2021-09-20 b6a9cc4
Apply suggestions from code review
Hopefully this will fix the bugs we experienced during testing.
