### 2021-09-20 6d53f13
[enh] initial commit of datasets module (#720)
* initial commit of datasets module

* Light clean-up, add in Cambridge regressor fetching

* Update docstrings with correct references, details

* Functioning cambridge fetcher

* save in tedana directory rather than nilean directory

* Add missing function

* return list of tuples for iterating

* Apply suggestions from code review

Hopefully this will fix the bugs we experienced during testing.

Co-authored-by: Taylor Salo <tsalo006@fiu.edu>
