+++ tedana/datasets/__init__.py
@@ -1,0 +1,1 @@
+from .fetchers import fetch_cambridge
