+++ tedana/datasets/data/cambridge.rst
@@ -1,0 +1,16 @@
+Notes
+-----
+The original data is downloaded from OpenNeuro
+https://openneuro.org/datasets/ds000258/versions/1.0.0
+This fetcher downloads preprocessed data that are available on Open
+Science Framework (OSF): https://osf.io/9wcb8/
+
+References
+----------
+.. [1] Power, J., Plitt, M., Gotts, S., Kundu, P., Voon, V.,
+    Bandettini, P., & Martin, A. (2018).
+    Ridding fMRI data of motion-related influences:
+    removal of signals with distinct spatial and physical bases
+    in multi-echo data.
+    PNAS, 115(9), E2105-2114.
+    www.pnas.org/content/115/9/E2105
