### 2021-09-20 c9164ea
Fix up docstrings. I think newlines were removed when pasted.
