+++ tedana/datasets/__init__.py
@@ -1,1 +1,2 @@
+"""Helper functions for downloading partially preprocessed multi-echo datasets."""
 from .fetchers import fetch_cambridge
