+++ tedana/datasets/__init__.py
@@ -1,2 +1,6 @@
 """Helper functions for downloading partially preprocessed multi-echo datasets."""
 from .fetchers import fetch_cambridge
+
+__all__ = [
+    "fetch_cambridge",
+]
