+++ tedana/datasets/fetchers.py
@@ -1,3 +1,4 @@
+"""Functions for fetching datasets."""
 import os
 import numpy as np
 import pandas as pd
