+++ tedana/datasets/fetchers.py
@@ -23,7 +23,7 @@ def _reduce_confounds(regressors, keep_confounds):
     return reduced_regressors
 
 
-def _fetch_cambridge_functional(n_subjects, data_dir, url, resume, verbose):
+def _fetch_cambridge_functional(n_subjects, low_resolution, data_dir, url, resume, verbose):
     """Helper function to fetch_cambridge.
 
     This function helps in downloading multi-echo functional MRI data
@@ -37,6 +37,8 @@ def _fetch_cambridge_functional(n_subjects, data_dir, url, resume, verbose):
     n_subjects : int
         The number of subjects to load. If None, all the subjects are
         loaded. Total 88 subjects.
+    low_resolution : bool
+        If True, download downsampled versions of the fMRI files, which is useful for testing.
     data_dir : str
         Path of the data directory. Used to force data storage in a specified
         location. If None is given, data are stored in home directory.
@@ -62,20 +64,22 @@ def _fetch_cambridge_functional(n_subjects, data_dir, url, resume, verbose):
         # https://gist.github.com/emdupre/3cb4d564511d495ea6bf89c6a577da74
         url = "https://osf.io/download/{}/"
 
-    func = "{0}_task-rest_{1}_space-scanner_desc-partialPreproc_bold.nii.gz"
+    if low_resolution:
+        func = "{0}_task-rest_{1}_space-scanner_desc-partialPreproc_res-5mm_bold.nii.gz"
+        csv_col = "key_bold_lowres"
+    else:
+        func = "{0}_task-rest_{1}_space-scanner_desc-partialPreproc_bold.nii.gz"
+        csv_col = "key_bold_nativeres"
 
     # The gzip contains unique download keys per Nifti file and confound
     # pre-extracted from OSF. Required for downloading files.
     package_directory = os.path.dirname(os.path.abspath(__file__))
-    dtype = [("participant_id", "U12"), ("echo_id", "U12"), ("key_bold", "U24")]
-    names = ["participant_id", "echo_id", "key_b"]
+    dtype = [("participant_id", "U12"), ("echo_id", "U12"), (csv_col, "U24")]
+    names = ["participant_id", "echo_id", csv_col]
+
     # csv file contains download information related to OpenScience(osf)
-    osf_data = csv_to_array(
-        os.path.join(package_directory, "data", "cambridge_echos.csv"),
-        skip_header=True,
-        dtype=dtype,
-        names=names,
-    )
+    csv_file = os.path.join(package_directory, "data", "cambridge_echos.csv")
+    osf_data = csv_to_array(csv_file, skip_header=True, dtype=dtype, names=names)
     funcs = []
     participant_id, echo_id, uuid = zip(*osf_data)
     participants = np.unique(participant_id)[:n_subjects]
@@ -87,7 +91,7 @@ def _fetch_cambridge_functional(n_subjects, data_dir, url, resume, verbose):
         for entry in this_osf_id:
             echo_id = entry["echo_id"]
             # Download bold images for each echo
-            func_url = url.format(entry["key_b"])
+            func_url = url.format(entry[csv_col])
             func_file = [
                 (
                     func.format(participant_id, echo_id),
@@ -168,7 +172,14 @@ def _fetch_cambridge_regressors(n_subjects, data_dir, url, resume, verbose):
 
 
 @due.dcite(Doi("10.1073/pnas.1720985115"), description="Introduces the Cambridge dataset.")
-def fetch_cambridge(n_subjects=None, reduce_confounds=True, data_dir=None, resume=True, verbose=1):
+def fetch_cambridge(
+    n_subjects=None,
+    low_resolution=False,
+    reduce_confounds=True,
+    data_dir=None,
+    resume=True,
+    verbose=1,
+):
     """Fetch Cambridge multi-echo data.
 
[…]
