### 2021-09-20 6a55640
Get fetcher working with new csv files.
