### 2021-09-21 cd3a41b
Sync update index, installation in docs
