### 2021-09-21 ME-ICA/tedana#804 @emdupre
Thank you for your review, @stebo85 !

We've added a note about `duecredit` as an optional dependency to the installation section in https://github.com/ME-ICA/tedana/commit/cd3a41b77fa46f1c6b960a79707c3f6a1c190a63, [diff visible here](https://github.com/ME-ICA/tedana/commit/cd3a41b77fa46f1c6b960a79707c3f6a1c190a63#diff-d9b149498982c0663c3b7170398773361ed5678f1a627e9c2fd8d2c955c563db).

Otherwise, I definitely agree that having more interactive examples in the documentation would be useful -- this is also something @martinagvilas brought up as part of her review in #798. I'm not sure, though, if this is immediately feasible, for the reasons outlined there.

Please let me know if this (and the linked commit !) addresses your concerns, here !
