### 2021-09-21 ME-ICA/tedana#804 @stebo85
Dear @emdupre,

Thank you for adding this - this is now much clearer.

I still think that better examples would be important, but if you believe that it's not feasible in a realistic timeframe than I will not be in your way.

Kind regards
Steffen
