### 2021-09-22 b1e1f38
added check for t2s 0 voxels
