### 2021-09-22 2457fc8
Adds warning on zero-counts
