### 2021-09-23 63493f5
getelbow for min 2 components
