+++ tedana/selection/_utils.py
@@ -94,10 +94,10 @@ def getelbow(arr, return_val=False):
     if arr.ndim != 1:
         raise ValueError("Parameter arr should be 1d, not {0}d".format(arr.ndim))
 
-    if not arr.size:
+    if arr.size <= 1:
         raise ValueError(
-            "Empty array detected during elbow calculation. "
-            "This error happens when getelbow is incorrectly called on no components. "
+            "Only 0 or 1 components used in elbow calculation. "
+            "This error happens when getelbow is incorrectly called on too few components. "
             "If you see this message, please open an issue at "
             "https://github.com/ME-ICA/tedana/issues with the full traceback and any data "
             "necessary to reproduce this error, so that we create additional data checks to "
