+++ tedana/selection/tedica.py
@@ -273,9 +273,16 @@ def kundu_selection_v2(comptable, n_echos, n_vols):
     # NOTE: Would an elbow from all Kappa values *ever* be lower than one from
     # a subset of lower (i.e., nonsignificant) values?
     kappa_elbow = np.nanmin((kappas_all_elbow, kappas_nonsig_elbow))
+    if kappa_elbow == kappas_all_elbow:
+        LGR.info(f"Kappa elbow is {kappa_elbow} and calculated based on all components")
+    else:
+        LGR.info(
+            f"Kappa elbow is {kappa_elbow} and calculated based on {kappas_nonsig.size} components with non-significant kappa fits"
+        )
     rhos_ncls_elbow = getelbow(comptable.loc[ncls, "rho"], return_val=True)
     rhos_all_elbow = getelbow(comptable["rho"], return_val=True)
     rho_elbow = np.mean((rhos_ncls_elbow, rhos_all_elbow, f05))
+    LGR.info(f"Rho elbow is {rho_elbow}")
 
     # Provisionally accept components based on Kappa and Rho elbows
     acc_prov = ncls[
