### 2021-09-23 e6f714c
Remove unnecessary fstring 'f'
Co-authored-by: Taylor Salo <tsalo006@fiu.edu>
