### 2021-09-23 84a7133
Adjusted info message wording
