### 2021-09-24 ME-ICA/tedana#798 @martinagvilas
Hi @emdupre, thanks for addressing these points.

I see how creating the demos takes a lot of time, and I think the publication of your package in JOSS shouldn't be halted for that reason if it is a work in progress.

So I considered this issue resolved 👍
