### 2021-09-24 ME-ICA/tedana#799 @martinagvilas
Thanks @emdupre for addressing these points.

About the last point: my comment was more style-related, so if following my advice makes you lose valuable information because of the space constrain in JOSS guidelines, I think you should ignore it. In any case, your documentation does a great job at explaining the advantages of fmri and multi-echo approaches, and is very beginner-friendly.

I consider this issue resolved, feel free to close it 👍
