### 2021-09-28 4d5c645
[JOSS] Add accepted JOSS manuscript to `main` (#813)
* Add manuscript.

Still need to check references.

* [skip-ci] Add newline at end.

* Update Asyraff reference.

* Remove unused citations.

* Fix citations.

* [skip ci] Minor change from @emdupre.

* Fix in-text citations.

* Add co-first authorship, corresponding, and multi-affiliations.

* Remove DuPre's middle initial.

Co-authored-by: Elizabeth DuPre <emd222@cornell.edu>

* Apply suggestions from code review

Co-authored-by: Dan Handwerker <7406227+handwerkerd@users.noreply.github.com>

* Apply suggestions from code review

Co-authore
[…]
