+++ paper/paper.bib
@@ -1,0 +1,859 @@
+@ARTICLE{Caballero-Gaudes2019-lv,
+  title    = "A deconvolution algorithm for multi-echo functional {MRI}:
+              Multi-echo Sparse Paradigm Free Mapping",
+  author   = "Caballero-Gaudes, C{\'e}sar and Moia, Stefano and Panwar, Puja
+              and Bandettini, Peter A and Gonzalez-Castillo, Javier",
+  abstract = "This work introduces a novel algorithm for deconvolution of the
+              BOLD signal in multi-echo fMRI data: Multi-echo Sparse Paradigm
+              Free Mapping (ME-SPFM). Assuming a linear dependence of the BOLD
+              percent signal change on the echo time (TE) and using
+              sparsity-promoting regularized least squares estimation, ME-SPFM
+              yields voxelwise time-varying estimates of the changes in the
+              apparent transverse relaxation ($\Delta$R) without prior
+              knowledge of the timings of individual BOLD events. Our results
+              in multi-echo fMRI data collected during a multi-task
+              event-related paradigm at 3 Tesla demonstrate that the maps of R
+              changes obtained with ME-SPFM at the times of the stimulus trials
+              show high spatial and temporal concordance with the activation
+              maps and BOLD signals obtained with standard model-based
+              analysis. This method yields estimates of $\Delta$R having
+              physiologically plausible values. Owing to its ability to blindly
+              detect events, ME-SPFM also enables us to map $\Delta$R
+              associated with spontaneous, transient BOLD responses occurring
+              between trials. This framework is a step towards deciphering the
+              dynamic nature of brain activity in naturalistic paradigms,
+              resting-state or experimental paradigms with unknown timing of
+              the BOLD events.",
+  journal  = "Neuroimage",
+  volume   =  202,
+  pages    = "116081",
+  month    =  nov,
+  year     =  2019,
+  keywords = "BOLD fMRI; Deconvolution; Multi-echo; Single-trial",
+  language = "en",
+  doi = {10.1016/j.neuroimage.2019.116081}
+}
+
+@ARTICLE{Kundu2013-xm,
+  title    = "Integrated strategy for improving functional connectivity mapping
+              using multiecho {fMRI}",
+  author   = "Kundu, Prantik and Brenowitz, Noah D and Voon, Valerie and Worbe,
+              Yulia and V{\'e}rtes, Petra E and Inati, Souheil J and Saad, Ziad
+              S and Bandettini, Peter A and Bullmore, Edward T",
+  abstract = "Functional connectivity analysis of resting state blood oxygen
+              level-dependent (BOLD) functional MRI is widely used for
+              noninvasively studying brain functional networks. Recent findings
+              have indicated, however, that even small ($\leq$1 mm) amounts of
+              head movement during scanning can disproportionately bias
+              connectivity estimates, despite various preprocessing efforts.
+              Further complications for interregional connectivity estimation
+              from time domain signals include the unaccounted reduction in
+              BOLD degrees of freedom related to sensitivity losses from high
+              subject motion. To address these issues, we describe an
+              integrated strategy for data acquisition, denoising, and
+              connectivity estimation. This strategy builds on our previously
+              published technique combining data acquisition with multiecho
+              (ME) echo planar imaging and analysis with spatial independent
+              component analysis (ICA), called ME-ICA, which distinguishes BOLD
+              (neuronal) and non-BOLD (artifactual) components based on linear
+              echo-time dependence of signals-a characteristic property of BOLD
+              T*2 signal changes. Here we show for 32 control subjects that
+              this method provides a physically principled and nearly
+              operator-independent way of removing complex artifacts such as
+              motion from resting state data. We then describe a robust
+              estimator of functional connectivity based on interregional
[…]
