+++ docs/approach.rst
@@ -19,6 +19,8 @@ This is performed in a series of steps, including:
 .. image:: /_static/tedana-workflow.png
   :align: center
 
+We provide more detail on each step below.
+The figures shown in this walkthrough are generated in the `provided notebooks <https://github.com/ME-ICA/tedana/tree/joss/docs/notebooks>`_.
 
 ***************
 Multi-echo data
