### 2021-09-30 #3669 @tsalo
@oliviaguest I'm sorry for the delay. We merged the `joss` branch into `main` and released 0.0.11. The new DOI is `10.5281/zenodo.5541689`.
